URLs for the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts

Generate the URL for a specific request to the NMME API based on parameters. See arguments for specifics and Details for links.

NMME_urls(
  start = Sys.Date(),
  end = as.Date("2050-01-01"),
  model = "ENSMEAN",
  lat = 31.9555,
  lon = -109.0744,
  freq = "daily",
  data = c("tasmin", "tasmean", "tasmax", "pr")
)

Arguments

start, end

Date for the start and end of the cast.

model

character value of the model to use, one of "ENSMEAN", (Multi-Model Mean), "CMC1" (CMC1-CanCM3), "CMC2" (CMC2-CanCM4), "CFCSv2" (NCEP-CFSv2), "GFDL" (GFDL-CM2.1), "GFDL-FLOR" (GFDL-FLOR), or "NCAR" (NCAR-CCSM4). Presently can only take one value.

lat, lon

numeric latitude and longitude values used to downscale the model. Presently can only take one value for each.

freq

character value of the frequency of the data, can be "daily" or "XmonthAverage", where "X" is a number between 1 and 7. Presently can only take one value.

data

character value of the type of data, one of "tasmin" (minimum temperature), "tasmean" (mean temperature), "tasmax" (maximum temperature), "pr" (precipitation), "dps" (dew point), "rsds" (shortwave radiation; sun intensity), "was" (wind speed).

Value

Named character vector of URLs, or NULL if data, freq, or model is NULL.

Details

The Northwest Knowledge Network (NKN) at the University of Idaho provides a simple API to download downscaled climate forecasts using the North American Multi-Model Ensemble (NMME) set.

Examples

  NMME_urls()