Provide URLs for the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts

The Northwest Knowledge Network (NKN) at at the University of Idaho provides a simple API to download downscaled climate forecasts using the North American Multi-Model Ensemble (NMME) set. This function generates the URL for specific request based on all possible user inputs including, time window, location, model, frequency of data, and data type. Given the construction of the URL, only data is vectorized. See arguments for specifics.

NMME_urls(start = Sys.Date(), end = as.Date("2050-01-01"),
  model = "ENSMEAN", lat = 31.9555, lon = -109.0744,
  freq = "daily", data = c("tasmin", "tasmean", "tasmax", "pr"),
  arg_checks = TRUE)

Arguments

start, end
Date for the start and end of the cast.

model
character value of the model, one of "ENSMEAN", (Multi-Model Mean), "CMC1" (CMC1-CanCM3), "CMC2" (CMC2-CanCM4), "CFCSv2" (NCEP-CFSv2), "GFDL" (GFDL-CM2.1), "GFDL-FLOR" (GFDL-FLOR), or "NCAR" (NCAR-CCSM4).

Presently can only take one value.

lat, lon
numeric latitude and longitude values used to downscale the model.

Presently can only take one value for each.

freq
character value of the frequency of the data, can be "daily" or "XmonthAverage", where "X" is a number between 1 and 7.

Presently can only take one value.

data
character value of the type of data, one of "tasmin" (minimum temperature), "tasmean" (mean temperature), "tasmax" (maximum temperature), "pr" (precipitation), "dps" (dew point), "rsds" (shortwave radiation; sun intensity), "was" (wind speed).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

Named character vector of URLs, or NULL if data, freq, or model is NULL.

Examples

  NMME_urls()