Add the Associated Values to a Cast Tab

Add values to a cast's cast tab. If necessary components are missing (such as no observations added yet and the user requests errors), the missing components are added.

add_lead_to_cast_tab adds a column of lead times.

add_obs_to_cast_tab appends a column of observations.

add_err_to_cast_tab adds a column of raw error values.

add_covered_to_cast_tab appends a logical column indicating if the observation was within the prediction interval.

add_lead_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_err_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_covered_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

add_obs_to_cast_tab(
  main = ".",
  settings = directory_settings(),
  cast_tab = NULL
)

Arguments

main

character value of the name of the main component of the directory tree.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

cast_tab

A data.frame of a cast's output. See read_cast_tab.

Value

data.frame of cast_tab with an additional column or columns if needed.

Details

If an interpolated data set is used to fit a model, add_obs_to_cast_tab adds the true (non-interpolated) observations so that model predictions are all compared to the same data.