Prepare options for a portalcasting directory

Suite of functions used to generate options lists:

all_options creates the full hierarchy of lists

dir_options creates a list of control options for the directory set-up

PortalData_options creates a list of options for filling the PortalData subdirectory

data_options creates a list of lists of options for filling the data subdirectory

moons_options creates a list of options for the moons data

rodents_options creates a list of options for the rodents data

covariates_options creates a list of options for the covariates data

metadata_options creates a list of options for the metadata creation

predictions_options creates a list of options for populating the predictions subdirectory

models_options creates a list of options for populating the models subdirectory

cast_options creates a list of control options for running portalcast

all_options(base = ".", main = "", subs = subdirs(), quiet = FALSE,
  fdate = today(), append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, model = models(),
  ensemble = TRUE, version = "latest", from_zenodo = TRUE)

dir_options(base = ".", main = "", subs = subdirs(), quiet = FALSE)

PortalData_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, version = "latest", from_zenodo = TRUE)

data_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, fdate = today(), append_missing_to_raw = TRUE,
  m_save = TRUE, m_filename = "moons.csv", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml")

moons_options(n_future_moons = 12, fdate = today(),
  append_missing_to_raw = TRUE, save = TRUE, filename = "moons.csv",
  tree = dirtree(), quiet = FALSE)

rodents_options(cast_type = "forecasts", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", save = TRUE, filename = "all.csv",
  tree = dirtree(), quiet = FALSE)

covariates_options(cast_type = "forecasts", cov_hist = TRUE,
  cov_fcast = TRUE, fdate = today(), yr = as.numeric(format(today(),
  "%Y")), start = 217, end = NULL, hind_step = 1, lead_time = 12,
  min_lag = 6, fcast_nms = NULL, nfcnm = 0,
  append_fcast_csv = TRUE, hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", save = TRUE,
  filename = "covariates.csv", tree = dirtree(), quiet = FALSE)

metadata_options(fdate = today(), cast_type = "forecasts",
  confidence_level = 0.9, lead_time = 12, save = TRUE,
  filename = "metadata.yaml", quiet = FALSE, tree = dirtree())

predictions_options(base = ".", main = "", subs = subdirs(),
  download_existing_predictions = FALSE, quiet = FALSE)

models_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, model = models())

cast_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, model = models(), cast_type = "forecasts",
  fdate = today(), ensemble = TRUE, start = 217, end = NULL,
  hind_step = 1, min_plots = 24, min_traps = 1)

Arguments

base
name of the base directory where the forecasting directory should exist (will be created if it doesn't)

main
name of the portalcasting directory

subs
names of the portalcasting subdirectories

quiet
logical indicator if progress messages should be quieted

fdate
date from which future is defined, typically today's date.

append_missing_to_raw
logical indicating if the missing moon dates should be appended to the raw data file (should be TRUE to allow the imported portalr functions to work properly)

m_save
logical if the moons data should be saved out

m_filename
the name of the file for saving moons data

tmnt_type
"all" or "controls"

start
newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01

end
newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample. If cast_type is "hindcasts" and end is NULL, default becomes 490:403, corresponding to 2017-01-01 back to 2010-01-01

hind_step
an iteration parameter to work across the input values for end. Default is 1, which is likely all it should be run using

drop_spp
species names to drop from the table

min_plots
minimum number of plots surveyed for a survey to be used

min_traps
minimum number of traps trapped for a plot to be used

level
for get_rodent_data, automatically set by tmnt_type

treatment
for get_rodent_data, automatically set by tmnt_type

plots
for get_rodent_data, automatically set by tmnt_type

output
for get_rodent_data, automatically set by tmnt_type

r_save
logical if the data should be saved out

r_filename
the name of the file to save the rodent data in

cov_hist
logical indicator whether or not historical covariates are to be included

cov_fcast
logical indicator whether or not forecasted covariates are to be included

yr
year of today

lead_time
number of moons into the future the rodents are forecast

min_lag
the minimum lag time used in any model

fcast_nms
newmoon numbers to be forecast for covariates

nfcnm
number of forecast newmoons for covariates

append_fcast_csv
logical if the new forecast should be appended to the historical forecasts for the purposes of hindcasting

hist_fcast_file
name of the file where the historical covariate forecasts are held

source_name
character value for the name to give the covariaate forecast. Currently is "current_archive". Previous to "current_archive", the data were retroactively filled in and are given the source name "retroactive"

c_save
logical if the covariate data should be saved out

c_filename
the name of the covariate file for the saving

cast_type
"forecasts" or "hindcasts"

confidence_level
confidence level used in summarizing model output

meta_save
logical if the metadata should be saved out

meta_filename
the name of the metadata file for the saving

download_existing_predictions
logical indicator is the existing predictions files should be retrieved from the portalPredictions repo

model
names of model scripts to include

ensemble
logical indicator of whether to create an ensemble model

version
version number or "latest" (default) for the Portal Data download

from_zenodo
logical indicator of whether or not the Portal Data should come from Zenodo (come from GitHub if not)

n_future_moons
integer value for the number of future moons to add

save
logical if the specific data should be saved out

filename
the name of the file for the saving

tree
directory tree

Value

all_options: list of control options lists ("options_dir" generated by dir_options, "options_PortalData" generated by PortalData_options, "options_data" generated by data_options, "options_predictions" generated by predictions_options, "options_models" generated by models_options, and "options_cast" generated by cast_options.

dir_options: a list of settings controlling the directory structure creation

PortalData_options: a list of settings controlling the creation and population of the PortalData subdirectory

data_options: list of control options lists ("moons" generated by moons_options, "rodents" generated by rodents_options, "covariates" generated by covariates_options, and "metadata" generated by metadata_options, as well as items "cast_type", "tree", and "quiet" for simple use with all or ancillary data.

moons_options: a list of settings controlling the moon data creation

rodents_options: a list of settings controlling the rodents data creation

covariates_options: a list of settings controlling the covariates data creation

metadata_options: a list of settings controlling the metadata creation

predictions_options: a list of settings controlling the population of the predictions subdirectory

models_options: a list of settings controlling the population of the models subdirectory

cast_options: a list of settings controlling the running of portalcast