Check a function's arguments' values for validity

Provides toggle-able argument checking for nearly all functions in the codebase (save a few utilities), that can be turned off to facilitate sandboxing or kept on to support reliability.

check_args checks that all of the arguments to a given function have valid values within the pipeline to avoid naming collision and improper formatting, by wrapping around check_arg for each argument.

check_arg returns NULL if the argument is valid or character vector of error messages associated with the specific situation (<fun_name>(<arg_name> = arg_value)) if the argument is invalid.

check_arg_list produces a list of argument details for checking within check_arg.

check_args should only be called within a function.

See Details for argument-specific rules.

check_args(arg_checks = TRUE)

check_arg(arg_name, arg_value, fun_name = NULL)

check_arg_list()

Arguments

arg_checks
logical value of if the arguments should be checked.

arg_name
character value of the argument name.

arg_value
Input value for the argument.

fun_name
character value of the function name or NULL.

Value

check_args: throws an error or NULL.

check_arg: character vector of error messages associated with the specific situation (<fun_name>(<arg_name> = arg_value)).

check_arg_list: list of argument checking component lists.

Details

Argument details used for checking are produced simply via check_arg_list. If you are adding to the codebase and use novel arguments, add them to that function's production.
Each argument should have its own element that is a list with the following elements: class (character value of its class), null (logical indicator if the argument can be NULL), na (logical indicator if the argument can be NA), length (numeric value of how many elements the argument can be or NULL if unrestricted), and addl if needed (an unevaluated expression that can be evaluated and allows for simple evaluation of, e.g., element names of a list; see the source code).

Usage rules for arguments are as follows:

Must be length-1 logical values, cannot be NULL or NA: bline, force_model_updates, only_if_missing.

Must be length-1 logical values, can be NULL, but cannot be NA: add_error, add_in_window, add_lead, add_obs, append_cast_csv, arg_checks, cast_obs, clean, cleanup, covariatesTF, effort, ensemble, fillweight, interpolate, include_interp, na_drop, nadot, NULLname, overwrite, quiet, retain_target_moons, return_version, save, silent_jags, tail, total, unknowns, update_prefab_models, verbose, with_census, zero_drop.

Must be logical values, can be any length or NULL, but cannot be NA: movedTF,

Must be length-1 character values, can be NULL, but cannot be NA: colname, data_name, data_set, directory, extension (if not NULL, must have a "." character), filename, filename_cov, filename_cov_casts, filename_config, filename_meta, filename_moons, freq, jags_model, level, list_function, main, method (if not NULL, must be "unwtavg", "unwtavg", "rjags", "simple", "interruptible", "parallel", "rjparallel", "background", "bgparallel", or "snow"), model (inputted values are checked via verify_models), name, output (if not NULL, must be "abundance"), path, plots (if not NULL, must be "all" or "longterm"), raw_data, sep_char, set, shape, source_name, source_url, species_id (if not NULL, must be within the species produced by rodent_spp("wtotal")), sub (must be one of the named sub directories: "tmp", "raw", "data", "models", "fits", or "casts"), time, title, type, url, winner (if not NULL, must be "hist" or "cast"), zip_destin.

Must be length-1 character cannot be NULL or NA: column

Must be character values, can be any length, can be NULL, but cannot be NA: colnames, concept_rec_id, data, data_sets, downloads_versions, eval_args, factories, files, models (inputted values are checked via verify_models), modules, monitor, msg, names, paths, rec_id, rec_version, ref_species, species (if not NULL, must be all within the species produced by all_secies(total = TRUE)), subs (must all be one of the named sub directories: "tmp", "raw", "data", "models", "fits", or "casts"), target_cols, treatment (if not NULL, must be "control")

Must be data.frames, can be any length, can be NULL, but cannot be NA: casts, cast_cov, cast_tab, casts, covariates, df, hist_cov, hist_tab, moons, obs, preds, rodents_tab.

Must be a data.frame or vector can be any length, can be NULL, but cannot be NA: dfv.

Must be a data.frame or list can be any length, can be NULL, but cannot be NA: dfl.

Must be lists, can be any length, can be NULL, but cannot be NA: control_climate_dl, control_files, control_list, control_model, control_runjags, controls_models, controls_rodents, data_set_controls, downloads, filenames, metadata, rodents.

Must be lists, can be any length, can be NULL or NA: run_status.

Must be length-1 Date-conformable values, can be NULL, but cannot be NA: cast_date, date end, from_date, start.

Must be Date-conformable values, can be any length, can be NULL, but cannot be NA: cast_dates. dates.

Must be length-1 numeric values, can be NULL, but cannot be NA: confidence_level (must be between 0 and 1), lat, lon.

Must be length-1 integer-conformable values can be NULL, but cannot be NA: adapt (must be positive), burnin (must be positive), cast_id (must be non-negative), end_moon (must be positive), lead (must be positive), lead_time (must be non-negative), lev, max_E (must be positive, 1 to 24), min_observed (must be positive), min_plots (must be positive), min_traps (must be positive), moon (must be positive), nchains (must be positive), ndates (must be positive), nmoons (must be non-negative), sample (must be positive), start_moon (must be positive), thin (must be positive), topx (must be positive).

Must be length-1 integer-conformable values can be NULL, and can be NA: lag (must be positive), min_lag (must be non-negative).

Must be integer-conformable values, can be any length, can be NULL, but cannot be NA: cast_groups (must be non-negative), cast_ids (must be non-negative), end_moons (must be positive), target_moons (must be positive).

Must be a list or function; can be NULL or NA: mutate.

Must be a character vector (any length), list, or function; can be NULL but not NA: inits.

Examples

 check_arg_list()
 check_arg("end_moon", 400)
 check_arg("end_moon", 399.5)