Cast the covariates

Cast the covariates for a model run.

cast_covariates is the primary function, which produces the covariates (min, mean, and max temperature; precipitation; and NDVI) required for a model to be cast.

prep_cast_covariates provides a wrapper on cast_covariates that includes saving the cast data out via save_cast_cov_csv and tidying the data for use within the model (by removing the vestigial columns that are only needed for saving).

cast_covariates(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_files = files_control(),
  control_climate_dl = climate_dl_control(), quiet = TRUE,
  verbose = FALSE, arg_checks = TRUE)

prep_cast_covariates(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_climate_dl = climate_dl_control(),
  control_files = files_control(), quiet = TRUE, verbose = FALSE,
  arg_checks = TRUE)

cast_ndvi(main = ".", hist_cov = NULL, lead_time = 12, min_lag = 6,
  arg_checks = TRUE)

cast_weather(main = ".", moons = NULL, hist_cov = NULL,
  end_moon = NULL, lead_time = 12, min_lag = 6,
  cast_date = Sys.Date(), control_climate_dl = climate_dl_control(),
  control_files = files_control(), quiet = TRUE, verbose = TRUE,
  arg_checks = TRUE)

save_cast_cov_csv(main = ".", moons = NULL, end_moon = NULL,
  cast_date = Sys.Date(), cast_cov = NULL,
  control_files = files_control(), quiet = TRUE, verbose = FALSE,
  arg_checks = TRUE)

Arguments

main
character name of the main directory tree component.

moons
Lunar data data.frame. See prep_moons.

hist_cov
data.frame of historic covariates. See prep_hist_covariates.

end_moon
Forecast origin. integer-conformable newmoon number or NULL, which equates to the most recently included census' newmoon number.

lead_time
integer-conformable number of timesteps forward a cast will proceed.

min_lag
integer-conformable minimum non-0 covariate lag time used in any model.

cast_date
Date on which the cast happens. Generally should be Sys.Date.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

control_climate_dl
list of specifications for the climate download. Sent to NMME_urls to create the specific URLs. See climate_dl_control.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

cast_cov
data.frame of cast covariates. See cast_covariates.

Value

cast_covariates: data.frame of the covariates cast for the moons as defined as needed for saving.

prep_cast_covariates: data.frame of the cast covariates after saving, with excess columns removed.

cast_ndvi: data.frame of -casted NDVI values.

cast_weather: data.frame of -casted weather values.

save_cast_cov_csv: cast_cov data.frame exactly as input.

Details

If forecasting from the current newmoon, weather is forecast using cast_weather and NDVI is forecast using cast_ndvi. If casting from a previous newmoon, the historical forecasted weather and NDVI data are retrieved from the file pointed to by filename_cov_casts.

Examples

 # \donttest{
  setup_dir()
  hist_cov <- prep_hist_covariates()
  prep_cast_covariates(hist_cov = hist_cov)
  cast_cov <- cast_covariates(hist_cov = hist_cov)
  save_cast_cov_csv(cast_cov = cast_cov)
  cast_ndvi_data <- cast_ndvi(hist_cov = hist_cov)
  cast_weather_data <- cast_weather(hist_cov = hist_cov)
 # }