Determine the start and end calendar dates for a cast window

Based on the cast origin (cast_date), lead time (lead_time), and minimum non-0 lag (min_lag), determines the dates bracketing the requested window.

cast_window(
  main = ".",
  moons = NULL,
  cast_date = Sys.Date(),
  lead_time = 12,
  min_lag = 6,
  control_files = files_control(),
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

moons
Moons data.frame. See prep_moons.

cast_date
Date from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using Sys.Date.

lead_time
integer (or integer numeric) value for the number of timesteps forward a cast will cover.

min_lag
integer (or integer numeric) of the minimum covariate lag time used in any model.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

Named list with elements start and end, which are both Dates.

Examples

 # \donttest{
   create_dir()
   fill_raw()
   cast_window()
 # }