Check a function's arguments' values for validity

check_args checks that all of the arguments to a given function have valid values within the pipeline to avoid naming collision and improper formatting, by wrapping around check_arg for each argument.

check_arg produces (returns) a NULL if the argument is valid or character vector of error messages associated with the specific situation (<fun_name>(<arg_name> = arg_value)) if the argument is invalid.

check_args should only be called within a function.

See Details for argument-specific rules.

check_args()

check_arg(arg_name, arg_value, fun_name = NULL)

Arguments

arg_name
character value of the argument name.

arg_value
Input value for the argument.

fun_name
character value of the function name or NULL.

Value

check_arg: character vector of error messages associated with the specific situation (<fun_name>(<arg_name> = arg_value)).

Details

Usage and rules for arguments are as follows:

add: must be NULL or a character vector in model_names.

add_error: must be a length-1 logical vector in append_observed_to_cast.

add_in_window: must be a length-1 logical vector in append_observed_to_cast.

add_lead: must be a length-1 logical vector in append_observed_to_cast.

add_obs: must be a length-1 logical vector in plot_cast_ts.

all: must be a list with elements named forecast and aic in save_forecast_output

all_forecasts: must be a data.frame in make_ensemble.

append_fcast_csv: must be a length-1 logical vector in all_options, covariates_options, and data_options.

append_missing_to_raw: must be a length-1 logical vector in all_options, data_options, and moons_options.

base must be a length-1 character vector in all_options, cast_options, data_options, dirtree, dir_options, models_options, PortalData_options, and predictions_options.

cast: must be a valid -cast data.frame as checked by cast_is_valid in verify_cast.

casts: must be of class casts in append_observed_to_cast and measure_cast_error.

cast_date: must be NULL or a length-1 Date or Date-conformable vector in add_addl_future_moons, all_options, cast_options, covariates_options, data_options, metadata_options, moons_options, plot_cast_point, plot_cast_ts, and read_cast.

cast_dates: must be NULL or a Date or Date-conformable vector in plot_cov_RMSE_mod_spp, read_casts, and select_most_ab_spp.

cast_to_check: must be a data.frame in cast_is_valid.

cast_type: must be a length-1 character vector in most_recent_cast, plot_cast_point, plot_cast_ts, plot_cov_RMSE_mod_spp, plot_err_lead_spp_mods, read_cast, read_casts, and select_most_ab_spp.

class: must be NULL or a character vector in classy.

colname: must be a length-1 character vector in fcast0, na_conformer, and remove_incompletes.

confidence_level: must be a length-1 numeric value between 0 and 1 in all_options, data_options, make_ensemble, and metadata_options.

controls: must be a list with elements named forecast and aic in save_forecast_output.

covariates: must be a data.frame of class covariates in forecast_covariates, forecast_ndvi, and prep_metadata.

covariate_source: must be NULL or a length-1 character vector of "current_archive" or "retroactive" in all_options, data_options, metadata_options.

covariate_date_made: must be NULL or a length-1 character vector corresponding to the date_made column in the covariate_forecasts data table in all_options, data_options, metadata_options.

cov_fcast: must be a length-1 logical vector in all_options, covariates_options, and data_options.

cov_hist: must be a length-1 logical vector in all_options, covariates_options, and data_options.

c_filename: must be a length-1 character vector in all_options, covariates_options, and data_options.

c_save: must be a length-1 logical vector in all_options, covariates_options, and data_options.

data_name: must be a length-1 character vector of value "all", "controls", "covariates", "covariate_forecasts", "moons", or "metadata" in read_data.

dates: must be NULL or a Date or Date-conformable vector in foy.

df: must be a data.frame in append_csv, dataout, and remove_incompletes.

download: must be a length-1 logical vector in download_predictions.

download_existing_predictions: must be a length-1 logical vector in all_options and predictions_options.

drop_spp: must be a character vector in all_options, data_options, remove_spp, and rodents_options.

end: must be NULL or a positive integer or integer-conformable vector in all_options, cast_options, covariates_options, data_options, and rodents_options.

ensemble: must be a length-1 logical vector in all_options and cast_options.

extension: must be a length-1 character vector with a single period in model_paths.

fcast_nms: must be NULL or a non-negative integer or integer-conformable vector in all_options, covariates_options, and data_options

filename: must be a length-1 character vector in append_csv and verify_PortalData.

from_date: must be NULL or a length-1 Date or Date-conformable vector in plot_cast_point and select_most_ab_spp.

from_zenodo: must be a length-1 logical vector in all_options and PortalData_options.

future_moons: must be a data.frame of class moons in add_addl_future_moons.

hind_step: must be a length-1 positive integer or integer-conformable vector in all_options, cast_options, covariates_options, data_options, and rodents_options.

hist_cov: must be a data.frame of class covariates in prep_fcast_covariates and update_covfcast_options.

hist_fcast_file: must be a length-1 character vector in all_options, covariates_options, and data_options.

lag: must be NULL or a length-1 non-negative integer or integer-conformable vector in lag_covariates, model_options, and pevGARCH.

lead: must be a length-1 positive integer or integer-conformable vector in plot_cast_point and select_most_ab_spp.

lead_time: must be a length-1 non-negative integer or integer-conformable vector in all_options, covariates_options, data_options, and metadata_options.

level: must be NULL or a length-1 character vector of value "All" or "Controls" in plot_cast_point, plot_cast_ts, plot_cast_ts_ylab, plot_cov_RMSE_mod_spp, plot_err_lead_spp_mods, select_most_ab_spp, and select_casts or a length-1 character vector of value "Site" or "Treatment" in AutoArima, all_options, data_options, ESSS, metadata_options, nbGARCH, nbsGARCH, pevGARCH, and rodents_options.

local_paths: must be a character vector in file_paths.

main must be a length-1 character vector in all_options, cast_options, data_options, dirtree, dir_options, PortalData_options, models_options, and predictions_options.

metadata: must be of class metadata in save_forecast_output.

meta_filename: must be a length-1 character vector in all_options, data_options, and metadata_options.

meta_save: must be a length-1 logical vector in all_options, data_options, and metadata_options.

min_lag: must be a length-1 non-negative integer or integer-conformable vector in all_options, covariates_options, and data_options.

min_observed: must be a length-1 positive integer or integer-conformable vector in measure_cast_error and plot_cov_RMSE_mod_spp.

min_plots: must be a length-1 positive integer or integer-conformable vector in all_options, cast_options, data_options, and rodents_options.

min_traps: must be a length-1 positive integer or integer-conformable vector in all_options, cast_options, data_options, and rodents_options.

model: must be a length-1 character vector in model_options, plot_cast_point, plot_cast_ts, plot_cast_ts_ylab, save_forecast_output, and select_most_ab_spp.

models: must be NULL or a character vector in all_options, cast_options, models_options, model_paths, and select_casts, plot_cov_RMSE_mod_spp, and plot_err_lead_spp_mods.

model_set: must be NULL or a length-1 character vector with value "prefab" or "wEnsemble" in model_names.

mod_covariates: must be a length-1 logical vector in model_options.

mod_type: must be "pevGARCH" in covariate_models .

moons: must be a data.frame of class moons in add_future_moons, append_past_moons_to_raw, forecast_covariates, forecast_ndvi, forecast_weather, format_moons, get_climate_forecasts, prep_covariates, prep_fcast_covariates, prep_metadata, prep_rodents, prep_rodents_list, trim_moons_fcast, update_covariates, and update_covfcast_options.

msg: must be NULL or a character vector in messageq.

m_filename: must be a length-1 character vector in all_options, data_options, and moons_options.

m_save: must be a length-1 logical vector in all_options, data_options, and moons_options.

nadot: must be a length-1 logical vector in rodent_spp.

ndates: must be a length-1 positive integer or integer-conformable vector in plot_err_lead_spp_mods.

newmoonnumbers: must be NULL or a length-1 non-negative integer or integer-conformable vector in select_casts.

new_forecast_covariates: must be a data.frame of class covariates in append_cov_fcast_csv.

nfcnm: must be a length-1 non-negative integer or integer-conformable vector in all_options, covariates_options, and data_options.

n_future_moons: must be a length-1 non-negative integer or integer-conformable vector in moons_options.

options_all: must be of class all_options in casts, cleanup_dir, fill_dir, portalcast, setup_dir, step_casts,and step_hind_forward.

options_cast: must be of class cast_options in add_ensemble, cast, cast_models, check_to_skip, combine_forecasts, models_to_cast, and verify_models.

options_covariates: must be of class covariates_options in append_cov_fcast_csv, forecast_covariates, forecast_ndvi, forecast_weather, get_climate_forecasts, prep_covariates, prep_fcast_covariates, prep_hist_covariates, trim_moons_fcast, update_covariates, and update_covfcast_options.

options_data: must be of class data_options in fill_data, prep_data, transfer_hist_covariate_forecasts, transfer_trapping_table, and update_data.

options_dir: must be of class dir_options in create_dir, create_main_dir, and create_sub_dirs.

options_metadata: must be of class metadata_options in prep_metadata.

options_model: must be of class model_options in model_template and write_model.

options_models: must be of class models_options in fill_models.

options_moons: must be of class moons_options in add_future_moons, append_past_moons_to_raw, and prep_moons.

options_out: must be an options list in dataout.

options_PortalData: must be of class PortalData_options in fill_PortalData.

options_predictions: must be of class predictions_options in fill_predictions.

options_rodents: must be of class rodents_options in enforce_rodents_options, prep_rodents, prep_rodents_list, trim_treatment, and update_rodents_list.

output: must be "abundance" in all_options, data_options, and rodents_options.

plots: must be a length-1 character vector of value "all" or "longerm" in all_options, data_options, and rodents_options.

pred_dir: must be a length-1 character vector in compile_aic_weights and make_ensemble.

quiet: must be a length-1 logical vector in all_options, AutoArima, cast_options, covariates_options, create_sub_dir, data_options, download_predictions, dir_options, ESSS, metadata_options, models_options, model_options, predictions_options, PortalData_options, moons_options, messageq, nbGARCH, nbsGARCH, pevGARCH, rodents_options, and verify_PortalData.

rangex: must be a length-2 positive integer or integer-conformable vector in plot_cast_ts_xaxis.

rodents: must be a data.table of class rodents in interpolate_abundance, is.spcol, remove_spp, and trim_treatment.

rodents_list: must be of class rodents_list in prep_metadata.

r_filename: must be a length-1 character vector in all_options, data_options, and rodents_options.

r_save: must be a length-1 logical vector in all_options, data_options, and rodents_options.

save: must be a length-1 logical vector in covariates_options, metadata_options, moons_options, and rodents_options.

source_name: must be a length-1 character vector in all_options, covariates_options, and data_options.

species: must be NULL or a character vector in plot_cast_point, plot_cast_point_yaxis, plot_cov_RMSE_mod_spp, plot_err_lead_spp_mods, select_casts, and select_most_ab_spp; specifically length-1 in plot_cast_ts and plot_cast_ts_ylab.

species_set: must be NULL or a length-1 character vector of value "base", "wtotal", or "evalplot" in rodent_spp.

specific_subs: must be NULL or a length-1 character vector in sub_paths.

spp_names: must be a character vector in is.spcol.

start: must be a length-1 positive integer or integer-conformable vector in all_options, cast_options, covariates_options, data_options, and rodents_options.

start_newmoon: must be a length-1 positive integer or integer-conformable vector in plot_cast_ts.

subs must be a class-character vector in all_options, cast_options, data_options, dirtree, dir_options, models_options, PortalData_options, and predictions_options.

subs_names: must be NULL or a character vector in subdirs.

subs_type: must be NULL or a length-1 character vector in subdirs.

sub_path: must be NULL or a length-1 character vector in create_sub_dir.

tail: must be a length-1 logical vector in lag_covariates.

temp_dir: must be a length-1 character vector in save_forecast_output.

time: must be a length-1 logical vector in today.

tmnt_type: must be NULL or a length-1 character vector of value "all" or "controls" in all_options, data_options, enforce_rodents_options, and rodents_options.

to_cleanup: must be NULL or a character vector in all_options and dir_options.

topx: must be a length-1 positive integer or integer-conformable vector in select_most_ab_spp.

treatment: must be NULL or "control" in all_options, data_options, and rodents_options.

tree: must be of class dirtree in append_observed_to_cast, AutoArima, base_path, clear_tmp, covariates_options, create_tmp, download_predictions, ESSS, file_paths, main_path, metadata_options, model_options, model_paths, model_scripts, moons_options, most_recent_cast, most_recent_census, nbGARCH, nbsGARCH, pevGARCH, plot_cast_point, plot_cast_point_yaxis, plot_cast_ts, plot_cast_ts_xaxis, plot_cast_ts_ylab, plot_cov_RMSE_mod_spp, plot_err_lead_spp_mods, prep_weather_data, read_all, read_cast, read_casts, read_controls, read_covariates, read_covariate_forecasts, read_data, read_metadata, read_moons, rodents_options, select_most_ab_spp, sub_paths, and verify_PortalData.

verbose: must be a length-1 logical vector in cast_is_valid.

version: must be a length-1 character vector in all_options and PortalData_options.

with_census: must be a length-1 logical vector in most_recent_cast and plot_cast_point.

x: must be a data.frame or vector in na_conformer and must simply exist as an object that can have a class attribute in classy.

yr: must be a length-1 integer or integer-conformable value after 1970 in all_options, covariates_options, and data_options.