Create a Directory Settings List

Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.

directory_settings(
  files = directory_files(),
  subdirectories = directory_subdirectories(),
  resources = directory_resources(),
  save = TRUE,
  overwrite = TRUE,
  unzip_pause = 30,
  download_timeout = getOption("timeout")
)

directory_files(
  directory_configuration = "directory_configuration.yaml",
  moons = "moon_dates.csv",
  covariates = "covariates.csv",
  historical_covariates = "historical_covariates.csv",
  forecast_covariates = "forecast_covariates.csv",
  dataset_controls = "dataset_controls.yaml",
  model_controls = "model_controls.yaml",
  cast_evaluations = "cast_evaluations.csv",
  metadata = "metadata.yaml"
)

directory_subdirectories(
  forecasts = "forecasts",
  fits = "fits",
  models = "models",
  resources = "resources",
  data = "data"
)

directory_resources(
  PortalData = list(source = "github", version = "latest"),
  portalPredictions = list(source = "github", version = NULL),
  climate_forecast = list(source = "NMME", version = as.character(Sys.Date()), data =
    c("tasmin", "tasmean", "tasmax", "pr"))
)

production_settings(download_timeout = max(getOption("timeout"), 600))

sandbox_settings()

Arguments

files

list of character names of standard files, see directory_files.

subdirectories

list of character names of standard subdirectories, see directory_subdirectories.

resources

character name for the resources subdirectory.

save

logical indicator controlling if the output should be saved out.

overwrite

logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

unzip_pause

Positive integer or integer numeric seconds for pausing during steps around unzipping that require time delay.

download_timeout

Positive integer or integer numeric seconds for timeout on downloads. Temporarily overrides the "timeout" option in options.

directory_configuration

character name for the directory configuration YAML.

moons

character name for the lunar data csv.

covariates

character name for the combined historical and forecast covariates csv.

historical_covariates

character name for the historical covariates csv.

forecast_covariates

character name for the forecast covariates csv.

dataset_controls

character name for the YAML of dataset control list(s).

model_controls

character name for the YAML of model controls list(s).

cast_evaluations

character name for the forecast evaluations csv.

metadata

character name for the Forecast metadata YAML.

forecasts

character name for the forecasts subdirectory.

fits

character name for the fits subdirectory.

models

character name for the models subdirectory.

data

character name for the data subdirectory.

PortalData

list of source and version elements of character values for the Portal Data download. Default values retrieve the latest data from github

portalPredictions

list of source and version elements of character values for the archive download. Default values point to github, but version = NULL indicates no download.

climate_forecast

list of source, version, and data elements of character values for the climate forecasts download. Default values retrieve the current day's forecast of min, mean, and max temperature and precipitation from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts.

Value

Named list of settings for the directory (for directory_settings) or list of settings components (for directory_files, directory_subdirectories, and directory_resources).