Create, update, and read the directory configuration file

The directory configuration file is a special file within the portalcasting directory setup and has its own set of functions.

write_directory_config creates the dir_config.yaml file. It is (and should only be) called from within create_dir, it captures information about the compute environment used to create the directory.

update_directory_config adds key components to the dir_config.yaml file; presently only adding the versions of the downloaded data from within fill_raw (the only place it is and should be called from.

read_directory_config reads contents of the dir_config.yaml file into the R session.

write_directory_config(main = ".", filename_config = "dir_config.yaml",
  quiet = FALSE, arg_checks = TRUE)

update_directory_config(main = ".",
  filename_config = "dir_config.yaml",
  downloads_versions = downloads_versions, quiet = FALSE,
  arg_checks = TRUE)

read_directory_config(main = ".", filename_config = "dir_config.yaml",
  quiet = FALSE, arg_checks = TRUE)

Arguments

main
character value of the name of the main component of the directory tree.

filename_config
character value of the path to the directory config YAML.

quiet
logical indicator if progress messages should be quieted.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

downloads_versions
character vector returned from fill_raw of the successfully downloaded versions of downloads.

Value

write_directory_config and update_directory_config both write out the dir_config.yaml file and return NULL.

read_directory_config: list of directory configurations.