Create a Directory Settings List

Most users will not want or need to change the directory folders and file names, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for pipeline functionality.

directory_settings(
  directory_config_file = "dir_config.yaml",
  moons_file = "moon_dates.csv",
  covariates_file = "covariates.csv",
  historical_covariates_file = "historical_covariates.csv",
  forecast_covariates_file = "forecast_covariates.csv",
  dataset_controls_file = "dataset_controls.yaml",
  model_controls_file = "model_controls.yaml",
  cast_evaluations_file = "cast_evaluations.csv",
  metadata_file = "metadata.yaml",
  subdirectories = list(forecasts = "casts", fits = "fits", models = "models",
    resources = "raw", data = "data"),
  PortalData = list(source = "github", version = "latest"),
  portalPredictions = list(source = "github", version = NULL),
  climate_forecast = list(source = "NMME", version = as.character(Sys.Date()), data =
    c("tasmin", "tasmean", "tasmax", "pr")),
  save = TRUE,
  overwrite = TRUE,
  unzip_pause = 30,
  download_timeout = getOption("timeout")
)

Arguments

directory_config_file

character value of the path to the directory config YAML.

moons_file

character name of the file for saving the moons data.

covariates_file, historical_covariates_file, forecast_covariates_file

character name of the files for saving the covariate data (combined historical and forecast) and each component.

dataset_controls_file

character name of the file for saving the rodent dataset controls.

model_controls_file

character name of the file for saving the model controls.

cast_evaluations_file

character name of the file for saving the cast evaluations.

metadata_file

character name of the file for saving the forecast metadata.

subdirectories

character vector of the subdirectory names. Default includes raw, data, models, fits, and casts.

PortalData

list with source and version elements that are character values for the source and version of the Portal Data to download. Default values retrieve the latest data from github.

See download_observations.

portalPredictions

list with source and version elements that are character values for the source and version of the archive to download. Default values point to github, but version = NULL indicates no download.

See download_archive.

climate_forecast

list with source and version elements that are character values for the source and version of the climate forecasts to download. Default values retrieve the current day's forecast from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts.

See download_climate_forecasts.

save

logical indicator controlling if the output should be saved out.

overwrite

logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

unzip_pause

Positive integer or integer numeric seconds for pausing during steps around unzipping that require time delay.

download_timeout

Positive integer or integer numeric seconds for timeout on downloads. Temporarily overrides the "timeout" option in options.

Value

Named list of settings for the directory.