Download the Portal Predictions Repository Archive

Downloads a specific version of the Portal Predictions repository from either GitHub or Zenodo (based on source) into the <main>/raw sub.

download_archive(
  main = ".",
  resources_sub = "raw",
  version = "latest",
  source = "github",
  quiet = FALSE,
  verbose = FALSE,
  pause = 30,
  timeout = getOption("timeout")
)

Arguments

main

character value defining the main component of the portalcasting directory tree.

resources_sub

character value defining the resources subdirectory of the portalcasting directory tree.

version

character version of the data to download. Default "latest" downloads the most recent (by date published). NULL means no download.

source

character indicator of the source for the download. Either "github" (default) or "github".

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator if detailed messages should be printed.

pause

Positive integer or integer numeric seconds for pausing during steps around unzipping that require time delay.

timeout

Positive integer or integer numeric seconds for timeout on downloads. Temporarily overrides the "timeout" option in options.

Value

NULL, invisible-ly.

Note

There are two calls to link[base]{Sys.sleep} for pause seconds each to allow for the file unzipping, copying, and such to catch up.

Examples

 # \donttest{

  create_dir(main = "./portalcasting")
  download_archive(main = "./portalcasting")
 # }