Download and read covariate casts (presently only the NMME)

Given the details specified in control_cdl, download climate forecasts from the Northwest Knowledge Network's (NKN) simple API to the North American Multi-Model Ensemble (NMME) and them ready for analyses.

download_climate_casts: downloads the files from the server.

read_climate_casts: reads the downloaded files into R and does some minimal tidying: makes sure -9999s become NAs, makes sure the date column is named as such and is actually a Date, and converts the temperatures to C and the precipitation to mm.

climate_dl_control: manages the control list for the URL generation via NMME_urls.

download_climate_casts(
  main = ".",
  control_climate_dl = climate_dl_control(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)

climate_dl_control(
  start = Sys.Date(),
  end = as.Date("2050-01-01"),
  model = "ENSMEAN",
  lat = 31.9555,
  lon = -109.0744,
  freq = "daily",
  data = c(mintemp = "tasmin", meantemp = "tasmean", maxtemp = "tasmax", precipitation
    = "pr"),
  arg_checks = TRUE
)

read_climate_casts(
  main = ".",
  control_climate_dl = climate_dl_control(),
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

control_climate_dl
list of specifications for the download, which are sent to NMME_urls to create the specific URLs.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

start, end
Date for the start and end of the cast.

model
character value of the model, one of "ENSMEAN", (Multi-Model Mean), "CMC1" (CMC1-CanCM3), "CMC2" (CMC2-CanCM4), "CFCSv2" (NCEP-CFSv2), "GFDL" (GFDL-CM2.1), "GFDL-FLOR" (GFDL-FLOR), or "NCAR" (NCAR-CCSM4).

Presently can only take one value.

lat, lon
numeric latitude and longitude values used to downscale the model.

Presently can only take one value for each.

freq
character value of the frequency of the data, can be "daily" or "XmonthAverage", where "X" is a number between 1 and 7.

Presently can only take one value.

data
character value of the type of data, one of "tasmin" (minimum temperature), "tasmean" (mean temperature), "tasmax" (maximum temperature), "pr" (precipitation), "dps" (dew point), "rsds" (shortwave radiation; sun intensity), "was" (wind speed).

Value

Named character vector of URLs, or NULL if data, freq, or model is NULL.

download_climate_casts: downloads the files and returns a character vector of URLs where the downloads came from, as generated by NMME_urls.

read_climate_casts: data.frame of the downloaded data formatted for casting.

climate_dl_control: a list of control arguments to pass to NMME_urls to generate the URLs for downloading.

Examples

 # \donttest{
  setup_dir()
  download_climate_casts()
  read_climate_casts()
 # }