Combine (ensemble) casts

Combine multiple casts' output into a single ensemble. Presently, only a general average ensemble is available.

ensemble_casts(
  main = ".",
  method = "unwtavg",
  cast_groups = NULL,
  cast_ids = NULL,
  cast_tab = NULL,
  end_moon = NULL,
  models = NULL,
  data_set = NULL,
  include_interp = TRUE,
  species = NULL,
  arg_checks = TRUE
)

Arguments

main

character value of the name of the main component of the directory tree.

method

character value of the name of the ensemble method to use. Presently, only "unwtavg" (unweighted average) is allowed.

cast_groups

integer (or integer numeric) value of the cast group to combine with an ensemble. If NULL (default), the most recent cast group is ensembled.

cast_ids

integer (or integer numeric) values representing the casts of interest for restricting ensembling, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information.

cast_tab

Optional data.frame of cast table outputs. If not input, will be loaded.

end_moon

integer (or integer numeric) newmoon number of the forecast origin. Default value is NULL, which equates to no selection with respect to end_moon.

models

character value(s) of the name of the model to include. Default value is NULL, which equates to no selection with respect to model. NULL translates to all models in the table.

data_set

character value of the rodent data set to include Default value is NULL, which equates to the first data set encountered.

include_interp

logical indicator of if the basic data set names should also be inclusive of the associated interpolated data sets.

species

character vector of the species code(s) or "total" for the total across species) to be plotted NULL translates to the species defined by evalplot_species.

arg_checks

logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

NULL. Plot is generated.

Details

A pre-loaded table of casts can be input, but if not (default), the table will be efficiently (as defined by the inputs) loaded and trimmed.
The casts can be trimmed specifically using the cast_ids input, otherwise, all relevant casts from the stated cast_groups will be included.

Examples

 # \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "ESSS"), end_moon = 515)
  ensemble_casts()
# }