Error if a function's request is deeper than can be handled

Produces an informative error message when a function that should only be called inside of other functions is called outside of a function (hence the request to the function is too deep for what it can handle).

error_if_deep(lev)

Arguments

lev
The number of frames back in the stack where the request needs to be able to be evaluated.

Value

Throws an error if the function is called in a place where it cannot operate and returns NULL otherwise.

Examples

 if (FALSE) {
 # will error:
 # error_if_deep(-10)
 }
 error_if_deep(0)