Exponential Smoothing State Space model for Portal Predictions

Fit an ESSS model in the portalcasting pipeline.

ESSS(tree = dirtree(), level = "All", quiet = FALSE)

Arguments

tree
dirtree-class directory tree list. See dirtree.

level
character value name of the type of plots included ("All" or "Controls").

quiet
logical value indicating if the function should be quiet.

Value

list of [1] "forecast" (the forecasted abundances) and [2] "all_model_aic" (the model AIC values).

Details

Model "ESSS" is a flexible exponential smoothing state space model with the possibility of multiplicative trends fit using ets in the forecast package (Hyndman et al. 2018).

References

Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting functions for time series and linear models. R package version 8.3.

Examples

# NOT RUN {
setup_dir()
ESSS()
# }