Create a control list for files

Most users will not want or need to change data file and folder names or saving conditions, but it is helpful to have them be flexible for certain circumstances, and this function gathers them into a list for higher-in-the-pipeline functions.

files_control(
  directory = "portalPredictions",
  raw_data = "PortalData",
  filename_moons = "moon_dates.csv",
  filename_config = "dir_config.yaml",
  filename_cov = "covariates.csv",
  filename_cov_casts = "covariate_casts.csv",
  filename_meta = "metadata.yaml",
  save = TRUE,
  overwrite = TRUE,
  cleanup = TRUE,
  source_name = "current_archive",
  append_cast_csv = TRUE,
  arg_checks = TRUE
)

Arguments

directory
character value of the directory name.

raw_data
character value indicating the name of the raw data directory. A standard portalcasting directory downloads the raw data files into from the PortalData repository, so raw_data = "PortalData".

filename_moons
character name of the file for saving the moons data.

filename_config
character filename of the directory configuration YAML.

filename_cov
character filename for saving the output.

filename_cov_casts
character filename for saving the covariate casts output.

filename_meta
character filename for saving the metadata.

save
logical indicator controlling if the output should be saved out.

overwrite
logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

cleanup
logical indicator of whether or not the tmp files should be cleaned up.

source_name
character value for the name to give the covariate forecast. Currently is "current_archive". Previous to "current_archive", the data were retroactively filled in and are given the source name "retroactive".

append_cast_csv
logical indicator controlling if the new cast covariates should be appended to the historical casts later use.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

Named list of names of the folders and files within the directory structure as well as saving strategies (directly as input).