Fill a Portalcasting Directory with Basic Components

Fill the directory with components including (optionally) the raw data (download_observations), portalPredictions archive (download_archive), climate forecasts (download_climate_forecasts), models (write_model).

fill_dir(
  main = ".",
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  models = prefab_models(),
  data_sets = prefab_data_sets(),
  quiet = FALSE,
  verbose = TRUE,
  controls_model = NULL,
  control_files = files_control(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_rodents = rodents_controls(),
  arg_checks = TRUE
)

fill_data(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_files = files_control(),
  data_sets = prefab_data_sets(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

fill_raw(
  main = ".",
  PortalData_version = "latest",
  PortalData_source = "gitub",
  portalPredictions_version = NULL,
  portalPredictions_source = "github",
  climate_forecast_version = Sys.Date(),
  climate_forecast_source = "NMME",
  quiet = FALSE,
  verbose = FALSE
)

fill_casts(main = ".", quiet = FALSE, verbose = FALSE)

fill_fits(main = ".", quiet = FALSE, verbose = FALSE)

fill_models(
  main = ".",
  models = prefab_models(),
  controls_model = NULL,
  control_files = files_control(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main

character value defining the main component of the portalcasting directory tree. Default value (".") puts the directory in the present location.

PortalData_source, PortalData_version

character values for the source and version of the Portal Data to download.

Default values retrieve the latest data from github.

See download_observations.

portalPredictions_source, portalPredictions_version

character values for the source and version of the archive to download.

Default values point to github, but version = NULL indicates no download.

See download_archive.

climate_forecast_source, climate_forecast_version

character values for the source and version of the climate forecasts to download.

Default values retrieve the current day's forecast from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts.

See download_climate_forecasts.

models

vector of models to be written via write_model.

data_sets

list of data_sets to be created using do.call on the defined functions.

quiet

logical indicator if messages should be quieted.

verbose

logical indicator if detailed messages should be printed.

controls_model

Additional controls for models not in the prefab set.
A list of a single model's script-writing controls or a list of lists, each of which is a single model's script-writing controls.
Presently, each model's script writing controls should include four elements:

name: a character value of the model name.

data_sets: a character vector of the data set names that the model is applied to.

covariatesTF: a logical indicator of if the model needs covariates.

lag: an integer-conformable value of the lag to use with the covariates or NA if covariatesTF = FALSE.

If only a single model is added, the name of the model from the element name will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the model and the name element.

control_files

list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

end_moon

integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

start_moon

integer (or integer numeric) newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01.

lead_time

integer (or integer numeric) value for the number of timesteps forward a cast will cover.

confidence_level

numeric confidence level used in summarizing model output. Must be between 0 and 1.

cast_date

Date from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using Sys.Date.

controls_rodents

Control list or list of lists (from rodents_controls) specifying the structuring of the rodents tables. See rodents_controls for details.

arg_checks

logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

NULL, invisible-ly.

Examples

 # \donttest{

  create_dir("./portalcasting")
  fill_dir("./portalcasting")

  create_dir("./pc")
  fill_raw("./pc")
  fill_casts("./pc")
  fill_fits("./pc")
  fill_models("./pc")
 # }