Run a single-species JAGS model

Provides an API to run.jags for single-species portalcasting models that requires the user only provide the JAGS model, initializer, monitor, and control list.

jags_ss(
  main = ".",
  data_set = "all",
  control_files = files_control(),
  control_runjags = runjags_control(),
  jags_model = NULL,
  monitor = NULL,
  inits = NULL,
  lag = NA,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main

character value of the name of the main component of the directory tree.

data_set

character value name of the rodent data set, such as ("all" or "controls").

control_files

list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

control_runjags

list of arguments passed to run.jags via runjags_control

jags_model

character value of the model. See Examples.

monitor

character vector of the names of the variables to monitor. See Examples and run.jags.

inits

function, character vector, or list of values (including random number generation components) for initializing runs of the model. See Examples and run.jags.

lag

integer (or integer numeric) of the lag time to use for the covariates.

quiet

logical value indicating if the function should be quiet.

verbose

logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

arg_checks

logical value of if the arguments should be checked using standard protocols via check_args.

Value

list of [1] model metadata list ("metadata"), [2] cast summary data.frame ("cast_tab"), [3] list of model fit objects ("model_fits"), and [4] list of model cast objects ("model_casts").

Details

The data structure constructed internally, which is available to the jags_model and inits components, contains the following elements: count, moon, and ntraps (all three include the forecast horizon as well, with NA for count, moon extended as expected, and ntraps using the max number of traps); N (the length of the count time series); past_count, past_moon, and past_ntraps, which provide all of the historic data that preceded start_moon (only including samples that are present; all NA counts are removed as are associated trap count sand moons); past_N (the length of the past_count time series); and if desired, covariates (table of all covariates for the span of moon). and past_covariates (table of all covariates for the span of past_moon).

References

Denwood, M. J. 2016. runjags: an R package providing interface utilities, model templates, parallel computing methods and additional distributions for MCMC models in JAGS. Journal of Statistical Software, 71:9. URL.

Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International URL.

Examples

 # \donttest{
  setup_dir()

  monitor <- c("mu", "tau")
  inits <- function(data = NULL){
    rngs <- c("base::Wichmann-Hill", "base::Marsaglia-Multicarry",
             "base::Super-Duper", "base::Mersenne-Twister")
    past_N <- data$past_N 
    past_count <- data$past_count 
    past_moon <- data$past_moon

    log_past_count <- log(past_count + 0.1)
    mean_log_past_count <- mean(log_past_count)
    sd_log_past_count <- max(c(sd(log_past_count) * sqrt(2), 0.01))
    diff_log_past_count <- rep(NA, past_N - 1)
    for(i in 1:(past_N - 1)){
      diff_count <- log_past_count[i + 1] - log_past_count[i]
      diff_time <- past_moon[i + 1] - past_moon[i] 
      diff_log_past_count[i] <- diff_count / diff_time
    }
    sd_diff_log_past_count <- max(c(sd(diff_log_past_count) * sqrt(2), 
                                  0.01))
    var_diff_log_past_count <- sd_diff_log_past_count^2
    precision_diff_log_past_count <- 1/(var_diff_log_past_count)
    rate <- 0.1
    shape <- precision_diff_log_past_count * rate

    function(chain = chain){
      list(.RNG.name = sample(rngs, 1),
           .RNG.seed = sample(1:1e+06, 1),
            mu = rnorm(1, mean_log_past_count, sd_log_past_count), 
            tau = rgamma(1, shape = shape, rate = rate))
    }
  }
  jags_model <- [1589 chars quoted with '"']

  jags_ss(jags_model = jags_model, monitor = monitor, inits = inits)
 # }