Measure error/fit metrics for forecasts

Summarize the cast-level errors or fits to the observations.
Presently included are root mean square error and coverage.

measure_cast_level_error(cast_tab = NULL, arg_checks = TRUE)

Arguments

cast_tab
A data.frame of a cast's output. See read_cast_tab.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

data.frame of metrics for each cast for each species.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Preparing directory for casting

#> ------------------------------------------------------------

#> Checking model availability

#>  *All requested models available*

#> ------------------------------------------------------------

#> Readying data for forecast origin newmoon 515

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 515

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 516

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 516

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 517

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 517

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 518

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 518

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 519

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 519

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 520

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 520

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Resetting data to the most up-to-date versions

#> ------------------------------------------------------------

#> Casting complete

#> ------------------------------------------------------------

  cast_tab <- read_cast_tab(cast_id = 1)
  cast_tab <- add_lead_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_obs_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_err_to_cast_tab(cast_tab = cast_tab)
  cast_tab <- add_covered_to_cast_tab(cast_tab = cast_tab)
  measure_cast_level_error(cast_tab = cast_tab)

#>    cast_id species     model         RMSE coverage
#> 1        1      BA AutoArima 1.155123e-05        1
#> 2        1      DM AutoArima 1.149331e+01        1
#> 3        1      DO AutoArima 2.031617e+00        1
#> 4        1      DS AutoArima 2.562863e-01        1
#> 5        1      NA AutoArima 3.418397e-01        1
#> 6        1      OL AutoArima 1.878825e+00        1
#> 7        1      OT AutoArima 1.596222e+00        1
#> 8        1      PB AutoArima 1.940570e+00        1
#> 9        1      PE AutoArima 2.618030e+00        1
#> 10       1      PF AutoArima 5.000000e-01        1
#> 11       1      PH AutoArima 4.406063e-02        1
#> 12       1      PL AutoArima 3.888322e-01        1
#> 13       1      PM AutoArima 0.000000e+00        1
#> 14       1      PP AutoArima 5.918862e+01        1
#> 15       1      RF AutoArima 2.364754e-02        1
#> 16       1      RM AutoArima 1.694697e+00        1
#> 17       1      RO AutoArima 2.887215e-01        1
#> 18       1      SF AutoArima 1.051358e+00        1
#> 19       1      SH AutoArima 1.253434e+00        1
#> 20       1      SO AutoArima 1.183600e-02        1
#> 21       1   total AutoArima 7.518123e+01        1

# }