Prepare options for a portalcasting model

Suite of functions used to generate model-writing options lists:

model_options is a basic template. Not to be confused with models_options, which creates the options list for the models subdirectory

AutoArima_options creates a list of control options for the AutoArima model

ESSS_options creates a list of control options for the ESSS model

nbGARCH_options creates a list of control options for the nbGARCH model

pevGARCH_options creates a list of control options for the pevGARCH model

model_options(tree = dirtree(), name = "AutoArima",
  covariates = FALSE, lag = NULL, quiet = FALSE)

AutoArima_options(tree = dirtree(), name = "AutoArima",
  covariates = FALSE, lag = NULL, quiet = FALSE)

ESSS_options(tree = dirtree(), name = "ESSS", covariates = FALSE,
  lag = NULL, quiet = FALSE)

nbGARCH_options(tree = dirtree(), name = "nbGARCH",
  covariates = FALSE, lag = NULL, quiet = FALSE)

pevGARCH_options(tree = dirtree(), name = "pevGARCH",
  covariates = TRUE, lag = 6, quiet = FALSE)

Arguments

tree
directory tree

name
character name of the model (MUST match the function name)

covariates
does the model require covariates

lag
lag time used for the covariates (NULL if covariates is false)

quiet
logical indicator if progress messages should be quieted

Value

model_options: a list of model options

AutoArima_options: a list of settings controlling the AutoArima model creation

ESSS_options: a list of settings controlling the ESSS model creation

nbGARCH_options: a list of settings controlling the nbGARCH model creation

pevGARCH_options: a list of settings controlling the pevGARCH model creation