Plot forecast or hindcast predictions for a given point in time across multiple species.

Plot the point value with confidence interval for a step in a forecast or hindcast across multiple species. Currently only reliable for forecasts.

plot_cast_point(tree = dirtree(), species = NULL, level = "Controls",
  cast_type = "forecasts", cast_date = NULL, model = "Ensemble",
  lead = 1, from_date = NULL, with_census = FALSE)

Arguments

tree
dirtree-class directory tree list. See dirtree.

species
character vector of the species codes (or "total" for the total across species) to be plotted or NULL (default) to plot all species and the total.

level
character value of the level of interest ("All" or "Controls").

cast_type
character value of the type of -cast of model. Used to select the file in the predictions subdirectory. Currently only reliably coded for "forecasts".

cast_date
Date the predictions were made. Used to select the file in the predictions subdirectory. If NULL (default), selects the most recent -casts (if with_census = FALSE) or the most recent -casts with a subsequent census (if with_census = TRUE).

model
character value of the name (or "Ensemble") of the model to be plotted.

lead
integer-conformable lead of the newmoon number used to select the data plotted.

from_date
Date to be used as a reference of when to count the lead from. If NULL (default), for cast_type = "forecasts", from_date = cast_date and plot_cast_point is not yet reliable for cast_type = "hindcasts".

with_census
logical toggle if the plot should include the observed data collected during the predicted census.

Examples

# NOT RUN {
setup_dir(all_options(download_existing_predictions = TRUE))
plot_cast_point()
plot_cast_point(with_census = TRUE)
# }