Visualize a Time Series Cast of a Species

Plot an observed timeseries and cast timeseries with a prediction interval.
Casts can be selected either by supplying a cast_id number or any combination of dataset, model, and end_moon, which filter the available casts in unison. This plot type can only handle output from a single cast, so if multiple casts still remain, the one with the highest number is selected. To be more certain about cast selection, use the cast_id input.

plot_cast_ts(
  main = ".",
  settings = directory_settings(),
  cast_id = NULL,
  cast_groups = NULL,
  dataset = NULL,
  model = NULL,
  end_moon = NULL,
  species = NULL,
  start_moon = 217,
  quiet = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

cast_id

integer (or integer numeric) value representing the cast of interest, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information.

cast_groups

integer (or integer numeric) value of the cast group to combine with an ensemble. If NULL (default), the most recent cast group is ensembled.

dataset

character value of the rodent data set to include Default value is NULL, which equates to no selection with respect to dataset.

model

character value of the name of the model to include. Default value is NULL, which equates to "Ensemble" or an unweighted combination of the most recent cast group's models via ensemble_casts.

end_moon

integer (or integer numeric) newmoon number of the forecast origin. Default value is NULL, which equates to no selection with respect to end_moon.

species

character value of the species codes (or "total" for the total across species) to be plotted. NULL (default) also gives the total.

start_moon

integer (or integer numeric) newmoon number for the beginning of the x-axis of the plot.
Does not influence the fit of the models, just the presentation.

quiet

logical indicator if progress messages should be quieted.

Value

NULL. Plot is generated.

Details

The resulting plot shows observations as a solid black line and predictions as a blue polygon with the bounds represent the error given by lower_pi and upper_pi and the bisecting blue line representing the estimate in the cast_table saved output from a model.
As of portalcasting v0.9.0, this represents the mean and the 95% prediction interval.
Observations that occurred after the cast are shown connected directly to the pre-cast observation data (as the black solid line).

Examples

 # \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  plot_cast_ts()
# }