Forecast Portal Rodents Models

Forecast the Portal rodent population data using the (updated if needed) data and models in a portalcasting directory.

portalcast wraps around cast to allow multiple runs of multiple models, with data preparation as needed between runs occurring via prepare_data.

cast runs a single cast of multiple models across data sets.

portalcast(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  end_moons = NULL,
  start_moon = 217,
  cast_date = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

cast(
  main = ".",
  models = prefab_models(),
  datasets = prefab_datasets(),
  end_moon = NULL,
  start_moon = 217,
  cast_date = Sys.Date(),
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

models

character vector of name(s) of model(s) to include.

datasets

character vector of dataset names to be created.

end_moons, end_moon

integer (or integer numeric) newmoon number(s) of the last sample(s) to be included. Default value is NULL, which equates to the most recently included sample.
end_moons allows for multiple moons, end_moon can only be one value.

start_moon

integer (or integer numeric) newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01.

cast_date

Date from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using Sys.Date.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

Value

Results are saved to files, NULL is returned invisible-ly.

Details

Multiple models can be run together on the multiple, varying data sets for a single new moon using cast. portalcast wraps around cast, providing updating, verification, and resetting utilities as well as facilitating multiple runs of cast across new moons (e.g., when using a rolling origin).

Examples

 # \donttest{
  setup_dir()
  portalcast(models = "ESSS")
  cast()
 # }