Specific message functions

Messaging functions: version_message creates a message for the local package version.

portalcast_welcome generates a welcome message for a portalcast run.

portalcast_goodbye makes a goodbye message for when models are done.

sandbox_welcome generates a special welcome message for sandboxes, displayed upon completion of setup_sandbox.

creation_message makes a message about the creation of the directory.

setup_completion_message creates a message for when the directory setting up has successfully finished.

fill_casts_message creates the final message to be used in fill_casts that relays the number, and optionally the specific names, of the files moved.

version_message(bline = TRUE, quiet = FALSE, arg_checks = TRUE)

portalcast_welcome(bline = TRUE, quiet = FALSE, arg_checks = TRUE)

portalcast_goodbye(bline = TRUE, quiet = FALSE, arg_checks = TRUE)

sandbox_welcome(main = ".", quiet = FALSE, arg_checks = TRUE)

creation_message(main = ".", bline = TRUE, quiet = FALSE,
  arg_checks = TRUE)

setup_completion_message(bline = TRUE, quiet = FALSE,
  arg_checks = TRUE)

data_readying_message(end_moon = NULL, bline = TRUE, quiet = FALSE,
  arg_checks = TRUE)

data_resetting_message(bline = TRUE, quiet = FALSE,
  arg_checks = TRUE)

models_running_message(end_moon = NULL, bline = TRUE, quiet = FALSE,
  arg_checks = TRUE)

model_running_message(model = NULL, quiet = FALSE, arg_checks = TRUE)

model_done_message(model = NULL, run_status = NULL, quiet = FALSE,
  arg_checks = TRUE)

fill_casts_message(files = NULL, movedTF = NULL, quiet = FALSE,
  verbose = FALSE, arg_checks = TRUE)

Arguments

bline
logical indicator if horizontal break lines should be included.

quiet
logical indicator if message should be quieted.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

main
character value of the name of the main component of the directory tree.

end_moon
integer (or integer numeric) newmoon number of the last sample(s) to be included.

model
character value of the model name or model script path. Will be stripped of all excess path components and the extension to give just the name of the model itself.

run_status
Output from the source-ing of a model script; a list if successful or NA if not.

files
character vector of the base file names of the files that may or may not have been moved.

movedTF
logical vector indicating if each of the files in files has been moved or not.

verbose
logical indicator of whether or not to print out all of the files and whether they were moved or not.

Value

Message is put to the console; NULL is returned.

Examples

 portalcast_welcome()
 portalcast_goodbye()
 sandbox_welcome()
 creation_message()
 setup_completion_message()
 data_readying_message(end_moon = 400)
 data_resetting_message()
 models_running_message(end_moon = 400)
 model_running_message("ESSS")
 model_done_message("ESSS", list("output"))
 model_done_message("ESSS", NA)
 fill_casts_message("xx.csv", TRUE)
 fill_casts_message(c("xx.csv", "yy.R"), c(TRUE, FALSE),
                          verbose = TRUE)