Functions for the prefab portalcasting models

Functions used for fitting the prefab models. For detailed descriptions of the models, see the model vignette.

AutoArima fits a flexible auto-regressive integrated moving average model using auto.arima.

NaiveArima fits a baseline random walk with no drift model as an ARIMA(0,1,0) using Arima.

ESSS fits a flexible exponential smoothing state space model with the possibility of multiplicative trends using ets.

nbGARCH fits a generalized autoregresive conditional heteroscedasticity model with overdispersion (i.e., a negative binomial response variable) using tsglm.

nbsGARCH fits a seasonal generalized autoregresive conditional heteroscedasticity model with overdispersion (i.e., a negative binomial response variable) using tsglm.

pevGARCH fits a set of generalized autoregressive conditional heteroscedasticity models with Poisson response variables and environmental covariates (max temp, mean temp, min temp, precipitation, and NDVI) using tsglm.

AutoArima(main = ".", data_set = "all",
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

NaiveArima(main = ".", data_set = "all",
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

ESSS(main = ".", data_set = "all_interp",
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

nbGARCH(main = ".", data_set = "all_interp",
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

nbsGARCH(main = ".", data_set = "all_interp",
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

pevGARCH(main = ".", data_set = "all_interp", lag = 6,
  control_files = files_control(), quiet = FALSE, verbose = FALSE,
  arg_checks = TRUE)

Arguments

main
character value of the name of the main component of the directory tree.

data_set
character value name of the rodent data set, such as ("all" or "controls").

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

quiet
logical value indicating if the function should be quiet.

verbose
logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args.

lag
integer (or integer numeric) of the lag time to use for the covariates.

Value

list of [1] model metadata list ("metadata"), [2] cast summary data.frame ("cast_tab"), [3] list of model fit objects ("model_fits"), and [4] list of model cast objects ("model_casts").

Details

AutoArima
Because the seasonality and sampling occurring with different frequencies, which auto.arima cannot accommodate, seasonal AutoArima models are not included.

pevGARCH
Environmental predictors have an optional lag time, which is currently structured to impact all covariates the same and defaults to 6.
All covariate models are fit and the best model (by AIC) is selected for use.

References

Hyndman R., Bergmeir C., Caceres G., Chhay L., O'Hara-Wild M., Petropoulos F., Razbash S., Wang E., and Yasmeen F. 2018. forecast: Forecasting functions for time series and linear models. R package version 8.3.

Liboschik T., Fokianos K., and Fried R. 2017. tscount: An R Package for Analysis of Count Time Series Following Generalized Linear Models. Journal of Statistical Software 82:5, 1-51. URL.

Examples

 # \donttest{
  setup_dir()
  AutoArima()
  NaiveArima()
  ESSS()
  nbGARCH()
  nbsGARCH()
  pevGARCH()
 # }