Provide the Names or Controls for the Prefabricated Models

Create a character vector of the names of the pre-fabricated (prefab) models or a list of their controls.

Usage

prefab_models_controls()

prefab_models()

Value

prefab_models: character vector of model names.
prefab_models_controls: list of model controls.

See also

Examples

  prefab_models_controls( )
  prefab_models( )