Prepare Covariate Data for Casting

Prepare and combine the historical and cast covariate data for a model run.

prep_historical_covariates and prep_forecast_covariates ready the historical and -casted covariates, respectively.

prep_covariates combines prep_historical_covariates and prep_forecast_covariates.

prep_covariates(
  main = ".",
  settings = directory_settings(),
  origin = Sys.Date(),
  quiet = TRUE,
  verbose = FALSE
)

prep_historic_covariates(
  main = ".",
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)

prep_forecast_covariates(
  main = ".",
  origin = Sys.Date(),
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

origin

Date for the forecast origin. Defaults to today's date.

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator if detailed messages should be printed.

Value

data.frame of historical and/or -casted covariates that is also saved out to settings$files$covariates if indicated by settings$save.

Details

prep_historical_covariates creates a data.frame of historical weather (weather) and NDVI data (ndvi). Automatically goes all the way back to the beginning of the available data, as the table is trimmed later as needed. NDVI data are generated by ndvi.

prep_forecast_covariates combines the NMME climate forecasts (download_climate_forecasts) with NDVI forecasts using a seasonal auto-ARIMA model on the most resolved and resources data available via (ndvi).