Prepare Lunar Data for the Portalcasting Repository

Get time information (calendar dates, census periods, and newmoon numbers) associated with trapping events (achieved and missed) based on a lunar survey schedule. If needed, additional moons will be added to both the in-use and resources versions of the data table.

add_future_newmoons adds future newmoon dates to the newmoon table, counting forward from origin. Because the newmoons table might not have the most recent newmoons, more rows than lead may need to be added to the table.

. get_cast_future_newmoons wraps around get_future_newmoons to include any additional newmoons needed to achieve the full lead from origin.

prepare_newmoons(
  main = ".",
  lead = 365,
  origin = Sys.Date(),
  settings = directory_settings(),
  quiet = TRUE,
  verbose = FALSE
)

add_future_newmoons(
  main = ".",
  newmoons = NULL,
  lead = 365,
  origin = Sys.Date(),
  settings = directory_settings()
)

forecast_future_newmoons(newmoons = NULL, lead = 365, origin = Sys.Date())

Arguments

main

character value of the name of the main component of the directory tree.

lead

integer (or integer numeric) value for the number of days forward a cast will cover.

origin

Date forecast origin, typically today's date (set using Sys.Date).

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

quiet

logical indicator controlling if messages are printed.

verbose

logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

moons

data.frame of moon data.

Value

Some version of a newmoons data.frame.

.

prepare_newmoons: fully appended and formatted data.frame (also saved out if settings$save = TRUE).

add_future_newmoons: fully appended and formatted data.frame.

forecast_future_newmoons: newmoons data.frame to append to the existing moons.

Details

Sometimes the resources newmoon data table is not fully up-to-date. Because the portalr functions weather and fcast_ndvi point to the resources newmoons data, that table needs to be updated to produce the correct current data table for casting.

Examples

 # \donttest{
  setup_dir()
  prepare_newmoons()
 # }