Prepare historical covariates data

prep_hist_covariates creates a data.frame of historical weather and NDVI data. Automatically goes all the way back to the beginning of the available data, as the table is automatically trimmed later.

prep_weather_data creates a data.frame of historical weather data using weather.

NDVI data are generated by ndvi.

prep_hist_covariates(main = ".", end_moon = NULL, quiet = TRUE,
  arg_checks = TRUE)

prep_weather_data(main = ".", arg_checks = TRUE)

Arguments

main
character value of the name of the main component of the directory tree.

end_moon
integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

quiet
logical indicator if progress messages should be quieted.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

Value

prep_hist_covariates: historical covariate data table as a data.frame.

prep_weather_data: data.frame of historical weather data.

Examples

 # \donttest{   
  create_dir()
  fill_raw()
  prep_hist_covariates()
  prep_weather_data()
 # }