Prepare lunar data for the portalcasting repository

A set of functions that get time information (calendar dates, census periods, and newmoon numbers) associated with trapping events (achieved and missed) based on a lunar survey schedule. If needed, additional moons will be added to both the in-use and raw versions of the data table.

add_future_moons adds future moon dates to the moon table, counting forward from cast_date. Because the moons table might not have the most recent moons, more rows than lead_time may need to be added to the table.

.

add_extra_future_moons adds more more moons to accomplish required data for lead_time. Because the moon table might not have the most recent moons, more rows than initially requested may need to be added to the table for it to cover lead_time .

format_moons formats the final output table with year and month columns formatted accordingly, and the moondate column formatted as a Date.

prep_moons(main = ".", lead_time = 12, cast_date = Sys.Date(),
  control_files = files_control(), quiet = TRUE, verbose = FALSE,
  arg_checks = TRUE)

format_moons(moons, arg_checks = TRUE)

add_future_moons(moons = NULL, lead_time = 12,
  cast_date = Sys.Date(), control_files = files_control(),
  arg_checks = TRUE)

add_extra_future_moons(moons, cast_date = Sys.Date(),
  arg_checks = TRUE)

Arguments

main
character value of the name of the main component of the directory tree.

lead_time
integer (or integer numeric) value for the number of timesteps forward a cast will cover.

cast_date
Date of the cast, typically today's date (set using Sys.Date).

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

quiet
logical indicator controlling if messages are printed.

verbose
logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

moons
Moons data.frame.

Value

All functions here return some version of a moons data.frame

. prep_moons, format_moons: fully appended and formatted data.frame (also saved out if save = TRUE).

add_future_moons and add_extra_future_moons: appropriately appended moons data.frame.

Details

Sometimes the raw moon data table is not fully up-to-date. Because the portalr functions weather and fcast_ndvi point to the raw moons data, that table needs to be updated to produce the correct current data table for casting.

Examples

 # \donttest{
  create_dir()
  fill_raw()
  prep_moons()
  pth <- file_path(files = "raw/PortalData/Rodents/moon_dates.csv")
  moons <- read.csv(pth, stringsAsFactors = FALSE)
  moons <- add_future_moons(moons)
  format_moons(moons)
 # }