Read in a data file and format it for specific class

Read in a specified data file.

Current options include "rodents" (produces a list), "rodents_table" (produces a rodents table), "covariates", "covariate_casts", "moons", and "metadata", which are available as calls to read_data with a specified data_name or as calls to the specific read_<data_name> functions (like read_moons).

If the requested data do not exist, an effort is made to prepare them using the associated prep_<data_name> functions (like prep_moons).

read_cov_casts reads in the current or (if no current version) local archived version of the covariate casts.

read_data(
  main = ".",
  data_name = NULL,
  data_set = "all",
  data_sets = c("all", "controls"),
  control_files = files_control(),
  arg_checks = TRUE
)

read_rodents_table(main = ".", data_set = "all", arg_checks = TRUE)

read_rodents(main = ".", data_sets = c("all", "controls"), arg_checks = TRUE)

read_covariates(main = ".", control_files = files_control(), arg_checks = TRUE)

read_covariate_casts(
  main = ".",
  control_files = files_control(),
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

read_moons(main = ".", control_files = files_control(), arg_checks = TRUE)

read_metadata(main = ".", control_files = files_control(), arg_checks = TRUE)

Arguments

main

character value of the name of the main component of the directory tree.

data_name

character representation of the data needed. Current options include "rodents", "rodents_table", "covariates", "covariate_forecasts", "moons", and "metadata".

data_set, data_sets

character representation of the grouping name(s) used to define the rodents. Standard options are "all" and "controls". data_set can only be length 1, data_sets is not restricted in length.

control_files

list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

arg_checks

logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator of whether or not to print out all of the information or just tidy messages.

Value

Data requested.

Examples

# \donttest{
 setup_dir()
 read_data(data_name = "rodents")
 read_data(data_name = "rodents_table")
 read_data(data_name = "rodents_table", data_set = "controls")
 read_data(data_name = "covariates")
 read_data(data_name = "covariate_casts")
 read_data(data_name = "moons")
 read_data(data_name = "metadata")

 read_rodents()
 read_rodents_table()
 read_covariates()
 read_covariate_casts()
 read_moons()
 read_metadata()
# }