Create control lists for generating rodents data tables

Given the number of arguments into summarize_rodent_data via prep_rodents_table, it helps to have a control list to organize them.

rodents_controls produces the lists for any user-requested datasets from the prefab datasets (prefab_data_sets) and any user-defined sets. rodents_control provides a template for the controls list which can be used to define the specific arguments for a user's novel data set.

rodents_controls(data_sets = NULL, controls_rodents = NULL, arg_checks = TRUE)

Arguments

data_sets
character value(s) of the rodent data set name(s) used to enforce certain arguments. Currently available prefab data sets are "all", "all_interp", "controls", "controls_interp", "exclosures_interp", "exclosures" and "dm_controls".

controls_rodents
Additional controls for datasets not in the prefab set.
A list of a single dataset's controls or a list of lists, each of which is a single dataset's controls.

Presently, each dataset's controls should include 18 elements, as our rodents_control template shows:

name: character value of the data set's name,

species: character-valued vector of species names to include.

total: logical value indicating if a total (sum across species) should be added or not. Only available if more than one species is included.

interpolate: logical value indicating if the NA values should be interpolated.

clean: logical indicator of if only the rodent data that passed QA/QC (clean = TRUE) or if all data (clean = FALSE) should be loaded.

type: character value of the rodent data set type, according to pre-existing definitions. An alternative toggle to species.
Either all species (type = "Rodents") or only granivoes (type = "Granivores").

level: character indicating the type of summary: "Plot", "Treatment", or "Site". Pipes directly to summarize_rodent_data.

plots: Specification of subset of plots. Can be a vector of numeric plots indicators or specific sets indicated by character values: "all" plots or "Longterm" plots (plots that have had the same treatment for the entire time series).

treatment: character indicating the specific treatment(s) to trim to if level = "Treatment": "control", "exclosure", "removal", or "spectabs"

min_plots: integer (or integer numeric) of the minimum number of plots surveyed for a survey to be used. Pipes directly to summarize_rodent_data.

min_traps: integer (or integer numeric) of the minimum number of traps collected for a plot to be used. Pipes directly to summarize_rodent_data.

output: character indicating the type of data: "abundance", "biomass", or "energy". Pipes directly to summarize_rodent_data.

fillweight: logical specifier of whether to fill in unknown weights with other records from that individual or species, where possible.

unknowns: logical indicator to either remove all individuals not identified to species (unknowns = FALSE) or sum them in an additional column (unknowns = TRUE).

time: character value specifying the format of the time index in the output. Options are "period" (sequential Portal surveys), "newmoon" (lunar cycle numbering), and "date" (calendar date).
The default time = "newmoon" produces an equispaced observation timestep, a common format format for discrete-time modeling.

na_drop: logical indicator of if NA values (representing insufficient sampling) should be dropped.

zero_drop: logical indicator of if 0 values (representing sufficient sampling but no detection) should be dropped.

effort: logical indicator of if the effort columns should be included in the output.

filename: character name of the file for saving the output.

If only a single dataset is added, the name of the set from the element data_set will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the dataset and the data_set element.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

Named list of length equal to the number of elements in data_sets and with elements that are each lists of those data_set's data-generating controls, for input as controls_rodents in prep_rodents.

Details

Any data set that is part of the prefab rodents data sets (c("all", "controls", "exclosures", "all_interp", "controls_interp", "exclosures_interp", "dm_controls")) has its controls already included internally via the non-exported function prefab_rodents_controls. Users only need to include controls for non-prefab data_sets.

Any user-defined data_sets that are not included in controls_rodents will throw an error.

If any user-defined controls_rodents duplicate any existing controls for the prefab data sets or if controls_r contains any duplicate named elements, an error will be thrown.

Users interested in adding data sets to the prefab rodent data sets should add the controls to the prefab_rodents_controls non-exported function found in the prepare_rodents.R script.

Examples

 rodents_controls(c("all", "controls"))
 all_PPonly <- rodents_control("all_PPonly", species = "PP")
 rodents_controls("all_PPonly", all_PPonly, arg_checks = FALSE)
 rodents_controls(c("all", "all_PPonly"), all_PPonly, arg_checks = FALSE)