Find casts that fit specifications

Determines the casts that match user specifications.
Functionally, a wrapper on read_casts_metadata with filtering for specifications that provides a simple user interface to the large set of available casts via the metadata.

select_casts(
  main = ".",
  cast_ids = NULL,
  cast_groups = NULL,
  end_moons = NULL,
  models = NULL,
  data_sets = NULL,
  quiet = FALSE,
  include_interp = FALSE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

cast_ids
integer (or integer numeric) values representing the casts of interest, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information.

cast_groups
integer (or integer numeric) value of the cast group to combine with an ensemble. If NULL (default), the most recent cast group is ensembled.

end_moons
integer (or integer numeric) newmoon numbers of the forecast origin. Default value is NULL, which equates to no selection with respect to end_moon.

models
character values of the names of the models to include. Default value is NULL, which equates to no selection with respect to model.

data_sets
character values of the rodent data sets to include Default value is NULL, which equates to no selection with respect to data_set.

quiet
logical indicator if progress messages should be quieted.

include_interp
logical indicator of if the basic data set names should also be inclusive of the associated interpolated data sets.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

data.frame of the cast_tab.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Preparing directory for casting

#> ------------------------------------------------------------

#> Checking model availability

#>  *All requested models available*

#> ------------------------------------------------------------

#> Readying data for forecast origin newmoon 515

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 515

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 516

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 516

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 517

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 517

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 518

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 518

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 519

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 519

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> Readying data for forecast origin newmoon 520

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 520

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running NaiveArima

#>   -NaiveArima for all

#>   -NaiveArima for controls

#>   -NaiveArima for exclosures

#>   -NaiveArima for dm_controls

#>   |++++| NaiveArima successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Resetting data to the most up-to-date versions

#> ------------------------------------------------------------

#> Casting complete

#> ------------------------------------------------------------

  select_casts(models = "AutoArima")

#>      cast_id cast_group  cast_date start_moon end_moon lead_time     model
#> 2          1          1 2019-11-15        217      524        12 AutoArima
#> 3          2          1 2019-11-15        217      524        12 AutoArima
#> 4          3          1 2019-11-15        217      524        12 AutoArima
#> 23        22          2 2019-11-21        217      524        12 AutoArima
#> 24        23          2 2019-11-21        217      524        12 AutoArima
#> 25        24          2 2019-11-21        217      524        12 AutoArima
#> 47        46          3 2019-11-22        217      524        12 AutoArima
#> 48        47          3 2019-11-22        217      524        12 AutoArima
#> 49        48          3 2019-11-22        217      524        12 AutoArima
#> 68        67          4 2019-11-28        217      525        12 AutoArima
#> 69        68          4 2019-11-28        217      525        12 AutoArima
#> 70        69          4 2019-11-28        217      525        12 AutoArima
#> 92        91          5 2019-11-29        217      525        12 AutoArima
#> 93        92          5 2019-11-29        217      525        12 AutoArima
#> 94        93          5 2019-11-29        217      525        12 AutoArima
#> 113      112          6 2019-12-05        217      525        12 AutoArima
#> 114      113          6 2019-12-05        217      525        12 AutoArima
#> 115      114          6 2019-12-05        217      525        12 AutoArima
#> 137      136          7 2019-12-06        217      525        12 AutoArima
#> 138      137          7 2019-12-06        217      525        12 AutoArima
#> 139      138          7 2019-12-06        217      525        12 AutoArima
#> 158      157          8 2019-12-19        217      525        12 AutoArima
#> 159      158          8 2019-12-19        217      525        12 AutoArima
#> 160      159          8 2019-12-19        217      525        12 AutoArima
#> 182      181          9 2019-12-20        217      525        12 AutoArima
#> 183      182          9 2019-12-20        217      525        12 AutoArima
#> 184      183          9 2019-12-20        217      525        12 AutoArima
#> 203      202         10 2019-12-27        217      526        12 AutoArima
#> 204      203         10 2019-12-27        217      526        12 AutoArima
#> 205      204         10 2019-12-27        217      526        12 AutoArima
#> 224      223         11 2020-01-02        217      526        12 AutoArima
#> 225      224         11 2020-01-02        217      526        12 AutoArima
#> 226      225         11 2020-01-02        217      526        12 AutoArima
#> 248      247         12 2020-01-03        217      526        12 AutoArima
#> 249      248         12 2020-01-03        217      526        12 AutoArima
#> 250      249         12 2020-01-03        217      526        12 AutoArima
#> 269      268         13 2020-01-09        217      526        12 AutoArima
#> 270      269         13 2020-01-09        217      526        12 AutoArima
#> 271      270         13 2020-01-09        217      526        12 AutoArima
#> 293      292         14 2020-01-10        217      526        12 AutoArima
#> 294      293         14 2020-01-10        217      526        12 AutoArima
#> 295      294         14 2020-01-10        217      526        12 AutoArima
#> 314      313         15 2020-01-16        217      526        12 AutoArima
#> 315      314         15 2020-01-16        217      526        12 AutoArima
#> 316      315         15 2020-01-16        217      526        12 AutoArima
#> 338      337         16 2020-01-17        217      526        12 AutoArima
#> 339      338         16 2020-01-17        217      526        12 AutoArima
#> 340      339         16 2020-01-17        217      526        12 AutoArima
#> 362      361         17 2020-01-23        217      526        12 AutoArima
#> 363      362         17 2020-01-23        217      526        12 AutoArima
#> 364      363         17 2020-01-23        217      526        12 AutoArima
#> 389      388         18 2020-01-24        217      526        12 AutoArima
#> 390      389         18 2020-01-24        217      526        12 AutoArima
#> 391      390         18 2020-01-24        217      526        12 AutoArima
#> 413      412         19 2020-01-29        217      527        12 AutoArima
#> 414      413         19 2020-01-29        217      527        12 AutoArima
#> 415      414         19 2020-01-29        217      527        12 AutoArima
#> 440      439         20 2020-01-31        217      527        12 AutoArima
#> 441      440         20 2020-01-31        217      527        12 AutoArima
#> 442      441         20 2020-01-31        217      527        12 AutoArima
#> 464      463         21 2020-02-06        217      527        12 AutoArima
#> 465      464         21 2020-02-06        217      527        12 AutoArima
#> 466      465         21 2020-02-06        217      527        12 AutoArima
#> 491      490         22 2020-02-07        217      527        12 AutoArima
#> 492      491         22 2020-02-07        217      527        12 AutoArima
#> 493      492         22 2020-02-07        217      527        12 AutoArima
#> 515      514         23 2020-02-13        217      527        12 AutoArima
#> 516      515         23 2020-02-13        217      527        12 AutoArima
#> 517      516         23 2020-02-13        217      527        12 AutoArima
#> 542      541         24 2020-02-14        217      527        12 AutoArima
#> 543      542         24 2020-02-14        217      527        12 AutoArima
#> 544      543         24 2020-02-14        217      527        12 AutoArima
#> 566      565         25 2020-02-20        217      527        12 AutoArima
#> 567      566         25 2020-02-20        217      527        12 AutoArima
#> 568      567         25 2020-02-20        217      527        12 AutoArima
#> 593      592         26 2020-02-27        217      528        12 AutoArima
#> 594      593         26 2020-02-27        217      528        12 AutoArima
#> 595      594         26 2020-02-27        217      528        12 AutoArima
#> 620      619         27 2020-03-06        217      528        12 AutoArima
#> 621      620         27 2020-03-06        217      528        12 AutoArima
#> 622      621         27 2020-03-06        217      528        12 AutoArima
#> 644      643         28 2020-03-12        217      528        12 AutoArima
#> 645      644         28 2020-03-12        217      528        12 AutoArima
#> 646      645         28 2020-03-12        217      528        12 AutoArima
#> 671      670         29 2020-03-19        217      528        12 AutoArima
#> 672      671         29 2020-03-19        217      528        12 AutoArima
#> 673      672         29 2020-03-19        217      528        12 AutoArima
#> 698      697         30 2020-04-02        217      529        12 AutoArima
#> 699      698         30 2020-04-02        217      529        12 AutoArima
#> 700      699         30 2020-04-02        217      529        12 AutoArima
#> 725      724         31 2020-04-09        217      529        12 AutoArima
#> 726      725         31 2020-04-09        217      529        12 AutoArima
#> 727      726         31 2020-04-09        217      529        12 AutoArima
#> 752      751         32 2020-05-07        217      530        12 AutoArima
#> 753      752         32 2020-05-07        217      530        12 AutoArima
#> 754      753         32 2020-05-07        217      530        12 AutoArima
#> 779      778         33 2020-05-14        217      530        12 AutoArima
#> 780      779         33 2020-05-14        217      530        12 AutoArima
#> 781      780         33 2020-05-14        217      530        12 AutoArima
#> 806      805         34 2020-05-21        217      530        12 AutoArima
#> 807      806         34 2020-05-21        217      530        12 AutoArima
#> 808      807         34 2020-05-21        217      530        12 AutoArima
#> 833      832         35 2020-05-28        217      531        12 AutoArima
#> 834      833         35 2020-05-28        217      531        12 AutoArima
#> 835      834         35 2020-05-28        217      531        12 AutoArima
#> 860      859         36 2020-06-04        217      531        12 AutoArima
#> 861      860         36 2020-06-04        217      531        12 AutoArima
#> 862      861         36 2020-06-04        217      531        12 AutoArima
#> 887      886         37 2020-06-11        217      531        12 AutoArima
#> 888      887         37 2020-06-11        217      531        12 AutoArima
#> 889      888         37 2020-06-11        217      531        12 AutoArima
#> 914      913         38 2020-06-18        217      531        12 AutoArima
#> 915      914         38 2020-06-18        217      531        12 AutoArima
#> 916      915         38 2020-06-18        217      531        12 AutoArima
#> 941      940         39 2020-07-02        217      532        12 AutoArima
#> 942      941         39 2020-07-02        217      532        12 AutoArima
#> 943      942         39 2020-07-02        217      532        12 AutoArima
#> 968      967         40 2020-07-09        217      532        12 AutoArima
#> 969      968         40 2020-07-09        217      532        12 AutoArima
#> 970      969         40 2020-07-09        217      532        12 AutoArima
#> 995      994         41 2020-07-16        217      532        12 AutoArima
#> 996      995         41 2020-07-16        217      532        12 AutoArima
#> 997      996         41 2020-07-16        217      532        12 AutoArima
#> 1022    1021         42 2020-07-23        217      533        12 AutoArima
#> 1023    1022         42 2020-07-23        217      533        12 AutoArima
#> 1024    1023         42 2020-07-23        217      533        12 AutoArima
#> 1049    1048         43 2020-07-30        217      533        12 AutoArima
#> 1050    1049         43 2020-07-30        217      533        12 AutoArima
#> 1051    1050         43 2020-07-30        217      533        12 AutoArima
#> 1076    1075         44 2020-08-06        217      533        12 AutoArima
#> 1077    1076         44 2020-08-06        217      533        12 AutoArima
#> 1078    1077         44 2020-08-06        217      533        12 AutoArima
#> 1103    1102         45 2020-08-13        217      533        12 AutoArima
#> 1104    1103         45 2020-08-13        217      533        12 AutoArima
#> 1105    1104         45 2020-08-13        217      533        12 AutoArima
#> 1130    1129         46 2020-08-20        217      534        12 AutoArima
#> 1131    1130         46 2020-08-20        217      534        12 AutoArima
#> 1132    1131         46 2020-08-20        217      534        12 AutoArima
#> 1157    1156         47 2020-08-27        217      534        12 AutoArima
#> 1158    1157         47 2020-08-27        217      534        12 AutoArima
#> 1159    1158         47 2020-08-27        217      534        12 AutoArima
#> 1184    1183         48 2020-09-03        217      534        12 AutoArima
#> 1185    1184         48 2020-09-03        217      534        12 AutoArima
#> 1186    1185         48 2020-09-03        217      534        12 AutoArima
#> 1211    1210         49 2020-09-10        217      534        12 AutoArima
#> 1212    1211         49 2020-09-10        217      534        12 AutoArima
#> 1213    1212         49 2020-09-10        217      534        12 AutoArima
#> 1238    1237         50 2020-09-17        217      534        12 AutoArima
#> 1239    1238         50 2020-09-17        217      534        12 AutoArima
#> 1240    1239         50 2020-09-17        217      534        12 AutoArima
#> 1265    1264         51 2020-10-01        217      535        12 AutoArima
#> 1266    1265         51 2020-10-01        217      535        12 AutoArima
#> 1267    1266         51 2020-10-01        217      535        12 AutoArima
#> 1292    1291         52 2020-10-15        217      535        12 AutoArima
#> 1293    1292         52 2020-10-15        217      535        12 AutoArima
#> 1294    1293         52 2020-10-15        217      535        12 AutoArima
#> 1319    1318         53 2020-10-22        217      536        12 AutoArima
#> 1320    1319         53 2020-10-22        217      536        12 AutoArima
#> 1321    1320         53 2020-10-22        217      536        12 AutoArima
#> 1346    1345         54 2020-10-29        217      536        12 AutoArima
#> 1347    1346         54 2020-10-29        217      536        12 AutoArima
#> 1348    1347         54 2020-10-29        217      536        12 AutoArima
#> 1370    1369         55 2020-12-03        217      537        12 AutoArima
#> 1371    1370         55 2020-12-03        217      537        12 AutoArima
#> 1372    1371         55 2020-12-03        217      537        12 AutoArima
#> 1394    1393         56 2020-12-10        217      537        12 AutoArima
#> 1395    1394         56 2020-12-10        217      537        12 AutoArima
#> 1396    1395         56 2020-12-10        217      537        12 AutoArima
#> 1418    1417         57 2020-12-17        217      538        12 AutoArima
#> 1419    1418         57 2020-12-17        217      538        12 AutoArima
#> 1420    1419         57 2020-12-17        217      538        12 AutoArima
#> 1442    1441         58 2021-01-14        217      539        12 AutoArima
#> 1443    1442         58 2021-01-14        217      539        12 AutoArima
#> 1444    1443         58 2021-01-14        217      539        12 AutoArima
#> 1466    1465         59 2021-01-21        217      539        12 AutoArima
#> 1467    1466         59 2021-01-21        217      539        12 AutoArima
#> 1468    1467         59 2021-01-21        217      539        12 AutoArima
#> 1490    1489         60 2021-01-28        217      539        12 AutoArima
#> 1491    1490         60 2021-01-28        217      539        12 AutoArima
#> 1492    1491         60 2021-01-28        217      539        12 AutoArima
#> 1514    1513         61 2021-02-05        217      539        12 AutoArima
#> 1515    1514         61 2021-02-05        217      539        12 AutoArima
#> 1516    1515         61 2021-02-05        217      539        12 AutoArima
#> 1538    1537         62 2021-02-25        217      540        12 AutoArima
#> 1539    1538         62 2021-02-25        217      540        12 AutoArima
#> 1540    1539         62 2021-02-25        217      540        12 AutoArima
#> 1562    1561         63 2021-02-26        217      515        12 AutoArima
#> 1563    1562         63 2021-02-26        217      515        12 AutoArima
#> 1564    1563         63 2021-02-26        217      515        12 AutoArima
#> 1565    1564         63 2021-02-26        217      515        12 AutoArima
#> 1570    1569         64 2021-02-26        217      516        12 AutoArima
#> 1571    1570         64 2021-02-26        217      516        12 AutoArima
#> 1572    1571         64 2021-02-26        217      516        12 AutoArima
#> 1573    1572         64 2021-02-26        217      516        12 AutoArima
#> 1578    1577         65 2021-02-26        217      517        12 AutoArima
#> 1579    1578         65 2021-02-26        217      517        12 AutoArima
#> 1580    1579         65 2021-02-26        217      517        12 AutoArima
#> 1581    1580         65 2021-02-26        217      517        12 AutoArima
#> 1586    1585         66 2021-02-26        217      518        12 AutoArima
#> 1587    1586         66 2021-02-26        217      518        12 AutoArima
#> 1588    1587         66 2021-02-26        217      518        12 AutoArima
#> 1589    1588         66 2021-02-26        217      518        12 AutoArima
#> 1594    1593         67 2021-02-26        217      519        12 AutoArima
#> 1595    1594         67 2021-02-26        217      519        12 AutoArima
#> 1596    1595         67 2021-02-26        217      519        12 AutoArima
#> 1597    1596         67 2021-02-26        217      519        12 AutoArima
#> 1602    1601         68 2021-02-26        217      520        12 AutoArima
#> 1603    1602         68 2021-02-26        217      520        12 AutoArima
#> 1604    1603         68 2021-02-26        217      520        12 AutoArima
#> 1605    1604         68 2021-02-26        217      520        12 AutoArima
#>         data_set portalcasting_version QAQC notes
#> 2            all                0.15.0 TRUE    NA
#> 3       controls                0.15.0 TRUE    NA
#> 4     exclosures                0.15.0 TRUE    NA
#> 23           all                0.15.0 TRUE    NA
#> 24      controls                0.15.0 TRUE    NA
#> 25    exclosures                0.15.0 TRUE    NA
#> 47           all                0.15.0 TRUE    NA
#> 48      controls                0.15.0 TRUE    NA
#> 49    exclosures                0.15.0 TRUE    NA
#> 68           all                0.15.0 TRUE    NA
#> 69      controls                0.15.0 TRUE    NA
#> 70    exclosures                0.15.0 TRUE    NA
#> 92           all                0.15.0 TRUE    NA
#> 93      controls                0.15.0 TRUE    NA
#> 94    exclosures                0.15.0 TRUE    NA
#> 113          all                0.15.0 TRUE    NA
#> 114     controls                0.15.0 TRUE    NA
#> 115   exclosures                0.15.0 TRUE    NA
#> 137          all                0.15.0 TRUE    NA
#> 138     controls                0.15.0 TRUE    NA
#> 139   exclosures                0.15.0 TRUE    NA
#> 158          all                0.15.1 TRUE    NA
#> 159     controls                0.15.1 TRUE    NA
#> 160   exclosures                0.15.1 TRUE    NA
#> 182          all                0.15.1 TRUE    NA
#> 183     controls                0.15.1 TRUE    NA
#> 184   exclosures                0.15.1 TRUE    NA
#> 203          all                0.15.1 TRUE    NA
#> 204     controls                0.15.1 TRUE    NA
#> 205   exclosures                0.15.1 TRUE    NA
#> 224          all                0.15.1 TRUE    NA
#> 225     controls                0.15.1 TRUE    NA
#> 226   exclosures                0.15.1 TRUE    NA
#> 248          all                0.15.1 TRUE    NA
#> 249     controls                0.15.1 TRUE    NA
#> 250   exclosures                0.15.1 TRUE    NA
#> 269          all                0.15.1 TRUE    NA
#> 270     controls                0.15.1 TRUE    NA
#> 271   exclosures                0.15.1 TRUE    NA
#> 293          all                0.15.1 TRUE    NA
#> 294     controls                0.15.1 TRUE    NA
#> 295   exclosures                0.15.1 TRUE    NA
#> 314          all                0.16.0 TRUE    NA
#> 315     controls                0.16.0 TRUE    NA
#> 316   exclosures                0.16.0 TRUE    NA
#> 338          all                0.17.1 TRUE    NA
#> 339     controls                0.17.1 TRUE    NA
#> 340   exclosures                0.17.1 TRUE    NA
#> 362          all                0.17.1 TRUE    NA
#> 363     controls                0.17.1 TRUE    NA
#> 364   exclosures                0.17.1 TRUE    NA
#> 389          all                0.17.1 TRUE    NA
#> 390     controls                0.17.1 TRUE    NA
#> 391   exclosures                0.17.1 TRUE    NA
#> 413          all                0.17.1 TRUE    NA
#> 414     controls                0.17.1 TRUE    NA
#> 415   exclosures                0.17.1 TRUE    NA
#> 440          all                0.17.1 TRUE    NA
#> 441     controls                0.17.1 TRUE    NA
#> 442   exclosures                0.17.1 TRUE    NA
#> 464          all                0.17.1 TRUE    NA
#> 465     controls                0.17.1 TRUE    NA
#> 466   exclosures                0.17.1 TRUE    NA
#> 491          all                0.17.1 TRUE    NA
#> 492     controls                0.17.1 TRUE    NA
#> 493   exclosures                0.17.1 TRUE    NA
#> 515          all                0.17.1 TRUE    NA
#> 516     controls                0.17.1 TRUE    NA
#> 517   exclosures                0.17.1 TRUE    NA
#> 542          all                0.17.1 TRUE    NA
#> 543     controls                0.17.1 TRUE    NA
#> 544   exclosures                0.17.1 TRUE    NA
#> 566          all                0.17.1 TRUE    NA
#> 567     controls                0.17.1 TRUE    NA
#> 568   exclosures                0.17.1 TRUE    NA
#> 593          all                0.17.1 TRUE    NA
#> 594     controls                0.17.1 TRUE    NA
#> 595   exclosures                0.17.1 TRUE    NA
#> 620          all                0.17.1 TRUE    NA
#> 621     controls                0.17.1 TRUE    NA
#> 622   exclosures                0.17.1 TRUE    NA
#> 644          all                0.17.1 TRUE    NA
#> 645     controls                0.17.1 TRUE    NA
#> 646   exclosures                0.17.1 TRUE    NA
#> 671          all                0.17.1 TRUE    NA
#> 672     controls                0.17.1 TRUE    NA
#> 673   exclosures                0.17.1 TRUE    NA
#> 698          all                0.17.1 TRUE    NA
#> 699     controls                0.17.1 TRUE    NA
#> 700   exclosures                0.17.1 TRUE    NA
#> 725          all                0.17.1 TRUE    NA
#> 726     controls                0.17.1 TRUE    NA
#> 727   exclosures                0.17.1 TRUE    NA
#> 752          all                0.17.1 TRUE    NA
#> 753     controls                0.17.1 TRUE    NA
#> 754   exclosures                0.17.1 TRUE    NA
#> 779          all                0.17.1 TRUE    NA
#> 780     controls                0.17.1 TRUE    NA
#> 781   exclosures                0.17.1 TRUE    NA
#> 806          all                0.17.1 TRUE    NA
#> 807     controls                0.17.1 TRUE    NA
#> 808   exclosures                0.17.1 TRUE    NA
#> 833          all                0.17.1 TRUE    NA
#> 834     controls                0.17.1 TRUE    NA
#> 835   exclosures                0.17.1 TRUE    NA
#> 860          all                0.17.1 TRUE    NA
#> 861     controls                0.17.1 TRUE    NA
#> 862   exclosures                0.17.1 TRUE    NA
#> 887          all                0.17.1 TRUE    NA
#> 888     controls                0.17.1 TRUE    NA
#> 889   exclosures                0.17.1 TRUE    NA
#> 914          all                0.17.1 TRUE    NA
#> 915     controls                0.17.1 TRUE    NA
#> 916   exclosures                0.17.1 TRUE    NA
#> 941          all                0.17.1 TRUE    NA
#> 942     controls                0.17.1 TRUE    NA
#> 943   exclosures                0.17.1 TRUE    NA
#> 968          all                0.17.1 TRUE    NA
#> 969     controls                0.17.1 TRUE    NA
#> 970   exclosures                0.17.1 TRUE    NA
#> 995          all                0.17.1 TRUE    NA
#> 996     controls                0.17.1 TRUE    NA
#> 997   exclosures                0.17.1 TRUE    NA
#> 1022         all                0.17.1 TRUE    NA
#> 1023    controls                0.17.1 TRUE    NA
#> 1024  exclosures                0.17.1 TRUE    NA
#> 1049         all                0.17.1 TRUE    NA
#> 1050    controls                0.17.1 TRUE    NA
#> 1051  exclosures                0.17.1 TRUE    NA
#> 1076         all                0.17.1 TRUE    NA
#> 1077    controls                0.17.1 TRUE    NA
#> 1078  exclosures                0.17.1 TRUE    NA
#> 1103         all                0.17.1 TRUE    NA
#> 1104    controls                0.17.1 TRUE    NA
#> 1105  exclosures                0.17.1 TRUE    NA
#> 1130         all                0.17.1 TRUE    NA
#> 1131    controls                0.17.1 TRUE    NA
#> 1132  exclosures                0.17.1 TRUE    NA
#> 1157         all                0.17.1 TRUE    NA
#> 1158    controls                0.17.1 TRUE    NA
#> 1159  exclosures                0.17.1 TRUE    NA
#> 1184         all                0.17.1 TRUE    NA
#> 1185    controls                0.17.1 TRUE    NA
#> 1186  exclosures                0.17.1 TRUE    NA
#> 1211         all                0.17.1 TRUE    NA
#> 1212    controls                0.17.1 TRUE    NA
#> 1213  exclosures                0.17.1 TRUE    NA
#> 1238         all                0.17.1 TRUE    NA
#> 1239    controls                0.17.1 TRUE    NA
#> 1240  exclosures                0.17.1 TRUE    NA
#> 1265         all                0.17.1 TRUE    NA
#> 1266    controls                0.17.1 TRUE    NA
#> 1267  exclosures                0.17.1 TRUE    NA
#> 1292         all                0.17.1 TRUE    NA
#> 1293    controls                0.17.1 TRUE    NA
#> 1294  exclosures                0.17.1 TRUE    NA
#> 1319         all                0.17.1 TRUE    NA
#> 1320    controls                0.17.1 TRUE    NA
#> 1321  exclosures                0.17.1 TRUE    NA
#> 1346         all                0.17.1 TRUE    NA
#> 1347    controls                0.17.1 TRUE    NA
#> 1348  exclosures                0.17.1 TRUE    NA
#> 1370         all                0.17.1 TRUE    NA
#> 1371    controls                0.17.1 TRUE    NA
#> 1372  exclosures                0.17.1 TRUE    NA
#> 1394         all                0.17.1 TRUE    NA
#> 1395    controls                0.17.1 TRUE    NA
#> 1396  exclosures                0.17.1 TRUE    NA
#> 1418         all                0.17.1 TRUE    NA
#> 1419    controls                0.17.1 TRUE    NA
#> 1420  exclosures                0.17.1 TRUE    NA
#> 1442         all                0.17.1 TRUE    NA
#> 1443    controls                0.17.1 TRUE    NA
#> 1444  exclosures                0.17.1 TRUE    NA
#> 1466         all                0.17.1 TRUE    NA
#> 1467    controls                0.17.1 TRUE    NA
#> 1468  exclosures                0.17.1 TRUE    NA
#> 1490         all                0.17.1 TRUE    NA
#> 1491    controls                0.17.1 TRUE    NA
#> 1492  exclosures                0.17.1 TRUE    NA
#> 1514         all                0.17.1 TRUE    NA
#> 1515    controls                0.17.1 TRUE    NA
#> 1516  exclosures                0.17.1 TRUE    NA
#> 1538         all                0.17.1 TRUE    NA
#> 1539    controls                0.17.1 TRUE    NA
#> 1540  exclosures                0.17.1 TRUE    NA
#> 1562         all                0.20.0 TRUE    NA
#> 1563    controls                0.20.0 TRUE    NA
#> 1564  exclosures                0.20.0 TRUE    NA
#> 1565 dm_controls                0.20.0 TRUE    NA
#> 1570         all                0.20.0 TRUE    NA
#> 1571    controls                0.20.0 TRUE    NA
#> 1572  exclosures                0.20.0 TRUE    NA
#> 1573 dm_controls                0.20.0 TRUE    NA
#> 1578         all                0.20.0 TRUE    NA
#> 1579    controls                0.20.0 TRUE    NA
#> 1580  exclosures                0.20.0 TRUE    NA
#> 1581 dm_controls                0.20.0 TRUE    NA
#> 1586         all                0.20.0 TRUE    NA
#> 1587    controls                0.20.0 TRUE    NA
#> 1588  exclosures                0.20.0 TRUE    NA
#> 1589 dm_controls                0.20.0 TRUE    NA
#> 1594         all                0.20.0 TRUE    NA
#> 1595    controls                0.20.0 TRUE    NA
#> 1596  exclosures                0.20.0 TRUE    NA
#> 1597 dm_controls                0.20.0 TRUE    NA
#> 1602         all                0.20.0 TRUE    NA
#> 1603    controls                0.20.0 TRUE    NA
#> 1604  exclosures                0.20.0 TRUE    NA
#> 1605 dm_controls                0.20.0 TRUE    NA

# }