Create and fill a forecasting directory

Combines create_dir and fill_dir to create a ready-to-run (via portalcast) directory where indicated.

setup_production creates a standard production directory.

setup_sandbox creates a sandboxing directory.

setup_dir(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  data_sets = prefab_data_sets(),
  bline = TRUE,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

setup_production(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = "latest",
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  quiet = FALSE,
  verbose = TRUE,
  data_sets = prefab_data_sets(),
  arg_checks = TRUE
)

setup_sandbox(
  main = ".",
  models = prefab_models(),
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  confidence_level = 0.95,
  cast_date = Sys.Date(),
  controls_model = NULL,
  controls_rodents = rodents_controls(),
  control_files = files_control(),
  PortalData_source = "gitub",
  PortalData_version = "latest",
  portalPredictions_source = "github",
  portalPredictions_version = NULL,
  climate_forecast_source = "NMME",
  climate_forecast_version = Sys.Date(),
  data_sets = prefab_data_sets(),
  quiet = FALSE,
  verbose = TRUE,
  arg_checks = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

models

character vector of name(s) of model(s) to include.

end_moon

integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

start_moon

integer (or integer numeric) newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01.

lead_time

integer (or integer numeric) value for the number of timesteps forward a cast will cover.

confidence_level

numeric confidence level used in summarizing model output. Must be between 0 and 1.

cast_date

Date from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using Sys.Date.

controls_model

Additional controls for models not in the prefab set.
A list of a single model's script-writing controls or a list of lists, each of which is a single model's script-writing controls.
Presently, each model's script writing controls should include three elements: name (a character value of the model name), covariates (a logical indicator of if the model needs covariates), and lag (an integer-conformable value of the lag to use with the covariates or NA if covariates = FALSE).
If only a single model is added, the name of the model from the element name will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the model and the name element.

controls_rodents

Control list or list of lists (from rodents_controls) specifying the structuring of the rodents tables. See rodents_controls for details.

control_files

list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

PortalData_source, PortalData_version

character values for the source and version of the Portal Data to download.

Default values retrieve the latest data from github.

See download_observations.

portalPredictions_source, portalPredictions_version

character values for the source and version of the archive to download.

Default values point to github, but version = NULL indicates no download.

See download_archive.

climate_forecast_source, climate_forecast_version

character values for the source and version of the climate forecasts to download.

Default values retrieve the current day's forecast from the Northwest Knowledge Network's North American Multi-Model Ensemble (NMME) climate forecasts.

See download_climate_forecasts.

data_sets

list of data_sets to be created using do.call on the defined functions.

bline

logical indicator if a horizontal break lines should be included in messages.

quiet

logical indicator if progress messages should be quieted.

verbose

logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

arg_checks

logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

Value

NULL

Examples

# \donttest{
 setup_dir()
 setup_sandbox()
 setup_production()
# }