Determine if Columns in a Table are Species Columns

Given a table, returns a logical vector indicating if each column is a species' column or not.

is_sp_col(rodents_tab = NULL, species = NULL, total = FALSE, nadot = FALSE)

species_from_table(rodents_tab = NULL, total = FALSE, nadot = FALSE)

Arguments

rodents_tab

data.frame of columns to be checked.

species

character vector indicating species names to use in determining if columns are species columns. Defaults to all_species.

total

logical indicator if the total should be included.

nadot

logical indicator if the dot should be added to the "NA" species name. Defaults to FALSE.

Value

is_sp_col: logical vector indicating if each column is a species' column or not.

species_from_table: character vector of species names from the column names in the rodents_tab data table.