Summarize a daily weather table by newmoons

Summarizes a daily weather table by newmoons. Taking the max date, min of the min temperatures, max of the max temperatures, mean of the mean temperatures, and sum of the precipitation.

summarize_daily_weather_by_moon(x)

Arguments

x
data.frame of daily weather, with columns named "date", "mintemp", "maxtemp", "meantemp", "precipitation", and "moon".

Value

data.frame of x summarized and arranged by x$moon.

Examples

 # \donttest{
  create_dir()
  fill_raw()
  raw_path <- raw_path()
  moons <- prep_moons()
  weather <- portalr::weather("daily", TRUE, raw_path)
  weather <- add_date_from_components(weather)
  weather <- add_moons_from_date(weather, moons)
  summarize_daily_weather_by_moon(weather)
 # }