Trim a moons table based on target moons

Some functions require that a data table of time (the moons) table only includes specific time stamps (newmoon numbers). This function is a simple utility to reduce the moons table to that as well as the columns.

trim_moons(
  moons = NULL,
  target_moons = NULL,
  retain_target_moons = TRUE,
  target_cols = c("moon", "moondate", "period", "censusdate"),
  arg_checks = TRUE
)

Arguments

moons
Moons data.frame. See prep_moons.

target_moons
integer-conformable newmoon numbers, usually made through target_moons of the moons to either drop or retain based on retain_target_moons.

retain_target_moons
logical value that dictates if target_moons are to be dropped (retain_target_moons = FALSE or retained (retain_target_moons = TRUE)

target_cols
character vector of columns to retain.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

Value

moons and a data.frame but trimmed.

Examples

 # \donttest{
  setup_dir()
  moons <- prep_moons()
  trim_moons(moons, 300:310)
 # }