Update models based on user input controls

Update model scripts based on the user-defined model control inputs. This allows users to define their own models or re-define prefab models within the portalcast pipeline.

update_models(
  main = ".",
  models = NULL,
  controls_model = NULL,
  update_prefab_models = FALSE,
  control_files = files_control(),
  bline = FALSE,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

models
character vector of name(s) of model(s), used to restrict the models to update from controls_model. The default (NULL) translates to all models in the controls_model list, which is recommended for general usage and is enforced within the main portalcast pipeline.

controls_model
Controls for models not in the prefab set or for overriding those in the prefab set.
A list of a single model's script-writing controls or a list of lists, each of which is a single model's script-writing controls.
Presently, each model's script writing controls should include four elements:

name: a character value of the model name.

data_sets: a character vector of the data set names that the model is applied to.

covariatesTF: a logical indicator of if the model needs covariates.

lag: an integer-conformable value of the lag to use with the covariates or NA if covariatesTF = FALSE.

If only a single model is added, the name of the model from the element name will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the model and the name element.

update_prefab_models
logical indicator if all of the models' scripts should be updated, even if they do not have an explicit change to their model options via controls_model. Default is FALSE, which leads to only the models in controls_model having their scripts re-written. Switching to TRUE results in the models listed in models having their scripts re-written.

This is particularly helpful when one is changing the global (with respect to the models) options main, quiet, verbose, or control_files.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

bline
logical indicator if horizontal break lines should be included in messaging.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

NULL.

Examples

 # \donttest{
  setup_dir()
  cm <- model_control(name = "AutoArima", data_sets = "all")
  update_models(controls_model = cm)
 # }