Obtain the URL for a Zenodo record to be downloaded

zenodo_url obtains the URL for a given Zenodo record, identified either by the concept record identifier (concept_rec_id) and version (rec_version) or record identifier (rec_id). (Note: if rec_id is used, it overrides concept_rec_id).

zenodo_versions: determines the available version numbers and the corresponding record identifier for each version available for a given Zenodo concept (group of records).

zenodo_url(concept_rec_id = NULL, rec_version = "latest",
  rec_id = NULL, arg_checks = TRUE)

zenodo_versions(concept_rec_id, arg_checks = TRUE)

Arguments

concept_rec_id
Concept record identifier, a character value corresponding to the Zenodo concept.

rec_version
character value of the version number or "latest" (default) for the data to be download.

rec_id
Optional input record identifier, a character value corresponding to the Zenodo record.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

zenodo_url: character value of the URL for the zip to be downloaded.

zenodo_versions: a data.frame of version number and record identifier for each version available.

Examples

 # \donttest{
   zenodo_versions("1215988")
   zenodo_url("1215988", "latest")
   zenodo_url("1215988", "1.71.0")
   zenodo_url(rec_id = "1217163")
 # }