Prepare options lists for a portalcasting directory

Suite of functions used to generate options lists that control the setup and use of a portalcasting directory:

all_options creates an all_options list that is a full hierarchy of portalcasting options lists: options_dir, options_PortalData, options_data, options_predictions, options_models, and options_cast.

dir_options creates a dir_options list of control options for the directory set-up.

PortalData_options creates a PortalData_options list of options for filling the PortalData subdirectory.

data_options creates a data_options list of lists of options for filling the data subdirectory.

moons_options creates a moons_options listof options for the moons data.

rodents_options creates a rodents_options list of options for the rodents data.

covariates_options creates a covariates_options listof options for the covariates data.

metadata_options creates a metadata_options listof options for the metadata creation.

predictions_options creates a predictions_options list of options for populating the predictions subdirectory

models_options creates a models_options list of options for populating the models subdirectory.

cast_options creates a cast_options list of control options for running portalcast.

all_options(base = ".", main = "", subs = subdirs(), quiet = FALSE,
  cast_date = today(), append_missing_to_raw = TRUE, m_save = TRUE,
  m_filename = "moons.csv", tmnt_type = NULL, start = 217,
  end = NULL, hind_step = 1, drop_spp = "PI", min_plots = 24,
  min_traps = 1, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml",
  download_existing_predictions = FALSE, models = model_names(),
  ensemble = TRUE, version = "latest", from_zenodo = TRUE,
  to_cleanup = c("tmp", "PortalData"),
  covariate_source = "retroactive",
  covariate_date_made = "2018-06-05 12:00:00 PST")

dir_options(base = ".", main = "", subs = subdirs(), quiet = FALSE,
  to_cleanup = c("tmp", "PortalData"))

PortalData_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, version = "latest", from_zenodo = TRUE)

data_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, cast_date = today(), append_missing_to_raw = TRUE,
  m_save = TRUE, m_filename = "moons.csv", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", r_save = TRUE, r_filename = "all.csv",
  cov_hist = TRUE, cov_fcast = TRUE, yr = as.numeric(format(today(),
  "%Y")), lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", c_save = TRUE,
  c_filename = "covariates.csv", cast_type = "forecasts",
  confidence_level = 0.9, meta_save = TRUE,
  meta_filename = "metadata.yaml", covariate_source = "retroactive",
  covariate_date_made = "2018-06-05 12:00:00 PST")

moons_options(n_future_moons = 12, cast_date = today(),
  append_missing_to_raw = TRUE, save = TRUE, filename = "moons.csv",
  tree = dirtree(), quiet = FALSE)

rodents_options(cast_type = "forecasts", tmnt_type = NULL,
  start = 217, end = NULL, hind_step = 1, drop_spp = "PI",
  min_plots = 24, level = "Site", treatment = NULL, plots = "all",
  output = "abundance", save = TRUE, filename = "all.csv",
  tree = dirtree(), quiet = FALSE)

covariates_options(cast_type = "forecasts", cov_hist = TRUE,
  cov_fcast = TRUE, cast_date = today(),
  yr = as.numeric(format(today(), "%Y")), start = 217, end = NULL,
  hind_step = 1, lead_time = 12, min_lag = 6, fcast_nms = NULL,
  nfcnm = 0, append_fcast_csv = TRUE,
  hist_fcast_file = "covariate_forecasts.csv",
  source_name = "current_archive", save = TRUE,
  filename = "covariates.csv", tree = dirtree(), quiet = FALSE)

metadata_options(cast_date = today(), cast_type = "forecasts",
  confidence_level = 0.9, lead_time = 12, save = TRUE,
  filename = "metadata.yaml", quiet = FALSE, tree = dirtree(),
  covariate_source = "retroactive",
  covariate_date_made = "2018-06-05 12:00:00 PST")

predictions_options(base = ".", main = "", subs = subdirs(),
  download_existing_predictions = FALSE, quiet = FALSE)

models_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, models = model_names())

cast_options(base = ".", main = "", subs = subdirs(),
  quiet = FALSE, models = model_names(), cast_type = "forecasts",
  cast_date = today(), ensemble = TRUE, start = 217, end = NULL,
  hind_step = 1, min_plots = 24, min_traps = 1)

Arguments

base
character name of the base folder where the directory should or does exist. Default "." (working directory). Will be created if it doesn't exist.

main
character name of the main folder within the base that contains the subdirectories. Default "" (no main level included).

subs
character vector naming the specific subdirectories within the portalcasting directory tree. Default subdirs() sets the subdirectories as "predictions", "models", "PortalData", "data", and "tmp". It is generally not advised to change the subdirectories.

quiet
logical indicator controlling if messages are generated.

cast_date
Date from which future is defined, typically today's date (using today).

append_missing_to_raw
logical indicator dictating if the missing moon dates should be appended to the raw data file (should be TRUE to allow the imported portalr functions to work properly).

m_save
logical indicator controlling if the moons data should be saved out.

m_filename
character name of the file for saving moons data.

tmnt_type
Treatment type: "all" or "controls".

start
integer (or integer numeric) newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01.

end
integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample. If cast_type is "hindcasts" and end is NULL, default becomes 490:403, corresponding to 2017-01-01 back to 2010-01-01.

hind_step
integer (or integer numeric) iteration parameter used to work across the input values for end. Default is 1, which is likely all it should be run using.

drop_spp
character-valued vector of species names to drop from the forecasting.

min_plots
integer (or integer numeric) of the minimum number of plots surveyed for a survey to be used.

min_traps
integer (or integer numeric) of the minimum number of traps trapped for a plot to be used.

level
character input for summarize_rodent_data, automatically set by tmnt_type.

treatment
character input for summarize_rodent_data, automatically set by tmnt_type.

plots
character input for summarize_rodent_data, automatically set by tmnt_type.

output
character input for summarize_rodent_data, automatically set by tmnt_type.

r_save
logical value indicating if the rodent data should be saved out.

r_filename
character name of the file to save the rodent data in.

cov_hist
logical indicator of whether or not historical covariates are to be included.

cov_fcast
logical indicator whether or not forecasted covariates are to be included.

yr
numeric value of the year of today's date.

lead_time
integer (or integer numeric) number of moons into the future the rodents are to be forecast.

min_lag
integer (or integer numeric) of the minimum covariate lag time used in any model.

fcast_nms
integer (or integer numeric) vector of newmoon numbers to be forecast for covariates.

nfcnm
integer (or integer numeric) number of forecast newmoons for covariates.

append_fcast_csv
logical indicator controlling if the new forecast should be appended to the historical forecasts for the purposes of hindcasting.

hist_fcast_file
character name of the file where the historical covariate forecasts are held.

source_name
character value for the name to give the covariate forecast. Currently is "current_archive". Previous to "current_archive", the data were retroactively filled in and are given the source name "retroactive".

c_save
logical indicator for if the covariate data should be saved out.

c_filename
character value for the name of the covariate file for the saving if c_save is TRUE.

cast_type
-cast type: "forecasts" or "hindcasts".

confidence_level
numeric confidence level used in summarizing model output. Must be between 0 and 1.

meta_save
logical indicator for if the metadata should be saved out

meta_filename
character value for the name of the metadata file for the saving if c_save is TRUE.

download_existing_predictions
logical indicator for if the existing predictions files should be retrieved from the portalPredictions repo.

models
character vector of class models of the names of models to include in a cast of the pipeline.

ensemble
logical indicator of whether to create an ensemble model.

version
character values of the version number or "latest" (default) for the Portal Data to be download.

from_zenodo
logical indicator of whether or not the Portal Data should come from Zenodo (come from GitHub if not).

to_cleanup
character vector of subdirectory names to cleanup after completion of casting.

covariate_source
character value for the name tagged to the covariate forecast. Possible values include "current_archive" and "retroactive" for hindcasts. Not presently needed for implementation in forecasts.

covariate_date_made
character value for the date-time tagged to the covariate forecast. Possible values depend on the dates available for covariates saved in the covariate_forecasts file for hindcasts. Not presently needed for implementation in forecasts.

n_future_moons
integer (or integer numeric) value for the number of future moons to add.

save
logical indicator if the specific data should be saved out.

filename
character name of the file for the saving.

tree
dirtree-class directory tree list. See dirtree.

Value

all_options: class-all_options list of control options lists (options_dir generated by dir_options, options_PortalData generated by PortalData_options, options_data generated by data_options, options_predictions generated by predictions_options, options_models generated by models_options, and options_cast generated by cast_options.

dir_options: a dir_options list of options for the directory structure creation.

PortalData_options: a PortalData_options list of settings controlling the population of the PortalData subdirectory.

data_options: data_options list of control options lists (moons generated by moons_options, rodents generated by rodents_options, covariates, generated by covariates_options, and metadata generated by metadata_options, as well as items cast_type, tree, and quiet for simple use with all or ancillary data..

moons_options: a moons_options list of settings controlling the moon data creation.

rodents_options: a rodents_options list of settings controlling the rodents data creation.

covariates_options: a covariates_options list of settings controlling the covariates data creation.

metadata_options: a metadata_options list of settings controlling the metadata creation.

predictions_options: a predictions_options list of settings controlling the population of the predictions subdirectory.

models_options: a models_options list of settings controlling the population of the models subdirectory.

cast_options: a cast_options list of settings controlling the running of portalcast.

Examples

# NOT RUN {
setup_dir()
all_options()
dir_options()
PortalData_options()
data_options()
moons_options()
rodents_options()
covariates_options()
metadata_options()
predictions_options()
models_options()
cast_options()
# }