Download raw components of a portalcasting directory

This suite of functions manages the downloading and (if needed) unzipping of raw files associated with the directory.

download downloads a file from a website into the directory, unzipping and cleaning up as needed.

download_message creates a customized download message.

download_url prepares the URL from the inputs, depending on the download type (see Details).

download_destin determines the download destination.

unzip_download unzips any compressed downloads.

unzip_destins determines the unzipping destinations for any compressed downloads.

download(
  name = NULL,
  type = NULL,
  url = NULL,
  concept_rec_id = NULL,
  rec_version = "latest",
  rec_id = NULL,
  sep_char = ".",
  return_version = TRUE,
  main = ".",
  sub = "raw",
  quiet = FALSE,
  verbose = FALSE,
  cleanup = TRUE,
  NULLname = FALSE,
  arg_checks = TRUE
)

unzip_download(
  name = NULL,
  zip_destin,
  main = ".",
  cleanup = TRUE,
  arg_checks = TRUE
)

unzip_destins(name = NULL, zip_destin, main = ".", arg_checks = TRUE)

download_message(
  name = NULL,
  type = NULL,
  url = NULL,
  rec_version = NULL,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

download_url(
  type = NULL,
  url = NULL,
  concept_rec_id = NULL,
  rec_version = "latest",
  rec_id = NULL,
  arg_checks = TRUE
)

download_destin(
  name = NULL,
  source_url,
  main = ".",
  sub = "raw",
  sep_char = ".",
  arg_checks = TRUE
)

Arguments

name
character value of the component's name, used to create the folder within the raw subdirectory. If left as NULL, name_from_url tries to obtain a name from the URL, unless NULLname is set to TRUE.

type
character value representing the type of input. Allowable types currently include a raw URL (type = "url"), which requires a non-NULL input for url, and from Zenodo (type = "zenodo"), which requires non-NULL input for either the concept record identifier (concept_rec_id) and version (rec_version) or record identifier (rec_id) (if rec_id is used, it overrides concept_rec_id.)

url
character value of the URL to be used if type = "url".

concept_rec_id
Concept record identifier, a character value corresponding to the Zenodo concept, used when type = "zenodo".

rec_version
character value of the version number or "latest" (default) for the data to be download. Used when type = "zenodo".

rec_id
Optional input record identifier, a character value corresponding to the Zenodo record. Used when type = "zenodo".

sep_char
character value of the separator that delineates the extension from the file path. Generally, this will be ".", but for some API URLs, the extension is actually a query component, so the separator may sometimes need to be "=".

return_version
logical indicator of whether or not to return the version name from the download URL.

main
character value of the name of the main component of the directory tree.

sub
character of the name of the subdirectory to download to.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

cleanup
logical indicator if any files put into the tmp subdirectory should be removed at the end of the process.

NULLname
logical indicator if name should be kept as NULL.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

zip_destin
character value of the destination of the download of the zip, which is to be unzipped.

source_url
character value of the URL from which the download should occur

Value

download: character of the name-version of the downloads if return_version = TRUE.

Details

If type = NULL, it is assumed to be a URL (i.e., type = "url").

Examples

 # \donttest{
  create_dir()
  download("PortalData", "zenodo", concept_rec_id = "1215988")
  source_url <- download_url(type = "zenodo", concept_rec_id = "1215988")
  destin <- download_destin(name = "PortalData", source_url = source_url)
  download_message(type = "zenodo", source_url, rec_version = "latest")
  download.file(source_url, destin, mode = "wb")
  unzip_destins("PortalData", destin)
  unzip_download("PortalData", destin)
 # }