Download and read covariate casts (presently only the NMME)

Given the details specified in control_cdl, download climate forecasts from the Northwest Knowledge Network's (NKN) simple API to the North American Multi-Model Ensemble (NMME) and them ready for analyses.

download_climate_casts: downloads the files from the server.

read_climate_casts: reads the downloaded files into R and does some minimal tidying: makes sure -9999s become NAs, makes sure the date column is named as such and is actually a Date, and converts the temperatures to C and the precipitation to mm.

climate_dl_control: manages the control list for the URL generation via NMME_urls.

download_climate_casts(
  main = ".",
  control_climate_dl = climate_dl_control(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)

climate_dl_control(
  start = Sys.Date(),
  end = as.Date("2050-01-01"),
  model = "ENSMEAN",
  lat = 31.9555,
  lon = -109.0744,
  freq = "daily",
  data = c(mintemp = "tasmin", meantemp = "tasmean", maxtemp = "tasmax", precipitation
    = "pr"),
  arg_checks = TRUE
)

read_climate_casts(
  main = ".",
  control_climate_dl = climate_dl_control(),
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

control_climate_dl
list of specifications for the download, which are sent to NMME_urls to create the specific URLs.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

start, end
Date for the start and end of the cast.

model
character value of the model, one of "ENSMEAN", (Multi-Model Mean), "CMC1" (CMC1-CanCM3), "CMC2" (CMC2-CanCM4), "CFCSv2" (NCEP-CFSv2), "GFDL" (GFDL-CM2.1), "GFDL-FLOR" (GFDL-FLOR), or "NCAR" (NCAR-CCSM4).

Presently can only take one value.

lat, lon
numeric latitude and longitude values used to downscale the model.

Presently can only take one value for each.

freq
character value of the frequency of the data, can be "daily" or "XmonthAverage", where "X" is a number between 1 and 7.

Presently can only take one value.

data
character value of the type of data, one of "tasmin" (minimum temperature), "tasmean" (mean temperature), "tasmax" (maximum temperature), "pr" (precipitation), "dps" (dew point), "rsds" (shortwave radiation; sun intensity), "was" (wind speed).

Value

Named character vector of URLs, or NULL if data, freq, or model is NULL.

download_climate_casts: downloads the files and returns a character vector of URLs where the downloads came from, as generated by NMME_urls.

read_climate_casts: data.frame of the downloaded data formatted for casting.

climate_dl_control: a list of control arguments to pass to NMME_urls to generate the URLs for downloading.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  download_climate_casts()

#>                                                                                                                                                                                                                                                                                                tasmin 
#>   "https://tds-proxy.nkn.uidaho.edu/thredds/ncss/NWCSC_INTEGRATED_SCENARIOS_ALL_CLIMATE/bcsd-nmme/dailyForecasts/bcsd_nmme_metdata_ENSMEAN_forecast_tasmin_daily.nc?var=tasmin&latitude=31.9555&longitude=-109.0744&time_start=2021-02-25T00%3A00%3A00Z&time_end=2050-01-01T00%3A00%3A00Z&accept=csv" 
#>                                                                                                                                                                                                                                                                                               tasmean 
#> "https://tds-proxy.nkn.uidaho.edu/thredds/ncss/NWCSC_INTEGRATED_SCENARIOS_ALL_CLIMATE/bcsd-nmme/dailyForecasts/bcsd_nmme_metdata_ENSMEAN_forecast_tasmean_daily.nc?var=tasmean&latitude=31.9555&longitude=-109.0744&time_start=2021-02-25T00%3A00%3A00Z&time_end=2050-01-01T00%3A00%3A00Z&accept=csv" 
#>                                                                                                                                                                                                                                                                                                tasmax 
#>   "https://tds-proxy.nkn.uidaho.edu/thredds/ncss/NWCSC_INTEGRATED_SCENARIOS_ALL_CLIMATE/bcsd-nmme/dailyForecasts/bcsd_nmme_metdata_ENSMEAN_forecast_tasmax_daily.nc?var=tasmax&latitude=31.9555&longitude=-109.0744&time_start=2021-02-25T00%3A00%3A00Z&time_end=2050-01-01T00%3A00%3A00Z&accept=csv" 
#>                                                                                                                                                                                                                                                                                                    pr 
#>           "https://tds-proxy.nkn.uidaho.edu/thredds/ncss/NWCSC_INTEGRATED_SCENARIOS_ALL_CLIMATE/bcsd-nmme/dailyForecasts/bcsd_nmme_metdata_ENSMEAN_forecast_pr_daily.nc?var=pr&latitude=31.9555&longitude=-109.0744&time_start=2021-02-25T00%3A00%3A00Z&time_end=2050-01-01T00%3A00%3A00Z&accept=csv" 

  read_climate_casts()

#>           date    mintemp  meantemp  maxtemp precipitation
#> 1   2021-02-28  3.0856260 14.076807 25.06799    0.00000000
#> 2   2021-03-01  3.6756219 14.226833 24.77801    0.00000000
#> 3   2021-03-02  4.0556272 14.471825 24.88799    0.00000000
#> 4   2021-03-03  3.1856325 12.986810 22.78799    0.00000000
#> 5   2021-03-04 -1.4243529  6.896814 15.21798    0.00000000
#> 6   2021-03-05 -5.2043830  6.781824 18.76800    0.00000000
#> 7   2021-03-06 -0.8343569 10.791834 22.41799    0.00000000
#> 8   2021-03-07  2.9156431 13.886804 24.85800    0.00000000
#> 9   2021-03-08  3.7456301 14.456813 25.16799    0.00000000
#> 10  2021-03-09  3.5256280 13.871820 24.21801    0.00000000
#> 11  2021-03-10  5.8456357 14.941796 24.03799    0.00000000
#> 12  2021-03-11  3.7956174 14.586817 25.37799    0.00000000
#> 13  2021-03-12  6.5756480 17.091823 27.60800    0.00000000
#> 14  2021-03-13  6.8056276 17.196804 27.58801    0.00000000
#> 15  2021-03-14  5.1756223 16.691797 28.20800    0.00000000
#> 16  2021-03-15  6.1456235 16.806819 27.46801    0.00000000
#> 17  2021-03-16  5.1056459 16.756829 28.40802    0.00000000
#> 18  2021-03-17  4.4456418 12.836817 21.22799    0.00000000
#> 19  2021-03-18  0.9056346 12.846828 24.78799    0.00000000
#> 20  2021-03-19  3.4956296 15.661829 27.82800    0.00000000
#> 21  2021-03-20  6.5456178 17.161829 27.77801    0.00000000
#> 22  2021-03-21  4.5356390 14.761806 24.98800    0.00000000
#> 23  2021-03-22  1.5856255 11.666834 21.74801    0.00000000
#> 24  2021-03-23  1.7756293 13.371819 24.96801    0.00000000
#> 25  2021-03-24  3.5956362 14.446833 25.29803    0.00000000
#> 26  2021-03-25  4.2956183 14.521815 24.74801    0.00000000
#> 27  2021-03-26  2.6456239 12.096827 21.54803    0.00000000
#> 28  2021-03-27  0.1756223 11.486842 22.79803    0.00000000
#> 29  2021-03-28  2.6256349 14.176814 25.72799    0.00000000
#> 30  2021-03-29  2.6056459 12.896813 23.18801    0.00000000
#> 31  2021-03-30  2.5631354 13.925898 25.28866    0.00000000
#> 32  2021-03-31  8.2637575 18.385646 28.50753    0.00000000
#> 33  2021-04-01  8.9437506 18.440637 27.93752    0.00000000
#> 34  2021-04-02  8.1537416 18.265652 28.37752    0.00000000
#> 35  2021-04-03  8.7137392 18.735623 28.75753    0.00000000
#> 36  2021-04-04  9.2037286 19.095637 28.98755    0.00000000
#> 37  2021-04-05  8.4537294 16.540612 24.62752    0.00000000
#> 38  2021-04-06  3.8737424 11.440637 19.00753    0.00000000
#> 39  2021-04-07  3.0137571 13.830651 24.64751    0.00000000
#> 40  2021-04-08  6.9437493 17.325647 27.70757    0.00000000
#> 41  2021-04-09  8.7437375 19.385647 30.02758    0.00000000
#> 42  2021-04-10  8.6437310 19.115626 29.58752    0.00000000
#> 43  2021-04-11  9.5837360 19.765650 29.94754    0.00000000
#> 44  2021-04-12 10.8437432 20.550622 30.25753    0.00000000
#> 45  2021-04-13 11.2437397 21.360618 31.47753    0.00000000
#> 46  2021-04-14 10.3137461 18.335656 26.35754    0.00000000
#> 47  2021-04-15  3.6837408 13.250635 22.81753    0.00000000
#> 48  2021-04-16  4.8937310 15.145624 25.39752    0.00000000
#> 49  2021-04-17  5.7837465 16.230676 26.67758    0.00000000
#> 50  2021-04-18  7.1137322 17.030635 26.94753    0.00000000
#> 51  2021-04-19  5.2437380 13.930628 22.61752    0.00000000
#> 52  2021-04-20  1.4137522 14.025659 26.63754    0.00000000
#> 53  2021-04-21  7.3337343 18.675622 30.01754    0.00000000
#> 54  2021-04-22 11.3937336 21.135644 30.87752    0.00000000
#> 55  2021-04-23  9.7237481 20.535639 31.34753    0.00000000
#> 56  2021-04-24 12.0637470 20.810632 29.55752    0.00000000
#> 57  2021-04-25  9.4037416 18.380640 27.35754    0.00000000
#> 58  2021-04-26  1.0637453 12.635646 24.20758    0.00000000
#> 59  2021-04-27  9.8137453 19.735650 29.65756    0.00000000
#> 60  2021-04-28 10.4237323 21.425620 32.42755    0.00000000
#> 61  2021-04-29 10.3955036 21.008814 31.62213    0.00000000
#> 62  2021-04-30  9.2771551 19.214469 29.15179    0.00000000
#> 63  2021-05-01  7.7571657 18.204490 28.65179    0.00000000
#> 64  2021-05-02  9.8271730 20.984489 32.14181    0.00000000
#> 65  2021-05-03 15.0071674 24.339502 33.67181    0.00000000
#> 66  2021-05-04 13.8271734 24.329495 34.83181    0.00000000
#> 67  2021-05-05 14.7471576 24.679468 34.61181    0.00000000
#> 68  2021-05-06 14.1971673 23.464500 32.73181    0.00000000
#> 69  2021-05-07 14.1372002 23.219507 32.30181    0.00000000
#> 70  2021-05-08 12.3771625 22.439478 32.50179    0.18964913
#> 71  2021-05-09 12.2171572 21.859461 31.50180    0.04004995
#> 72  2021-05-10 13.4771686 21.794489 30.11181    0.50180492
#> 73  2021-05-11 11.5771739 20.479486 29.38180    0.25090232
#> 74  2021-05-12 11.9971551 22.139490 32.28179    0.00000000
#> 75  2021-05-13 13.4271791 23.774499 34.12179    0.00000000
#> 76  2021-05-14 14.0171772 24.199490 34.38183    0.08422294
#> 77  2021-05-15 15.7871649 25.029505 34.27181    0.29566428
#> 78  2021-05-16 17.0071665 25.739466 34.47179    0.00000000
#> 79  2021-05-17 16.1671702 24.924461 33.68178    0.00000000
#> 80  2021-05-18 13.5871527 22.564477 31.54180    1.52426226
#> 81  2021-05-19 12.4371592 20.809470 29.18179    0.04240584
#> 82  2021-05-20 11.6671711 20.734490 29.80181    0.00000000
#> 83  2021-05-21 11.7171563 21.124475 30.53182    0.00000000
#> 84  2021-05-22 12.2271686 21.819483 31.41180    0.00000000
#> 85  2021-05-23 12.7071783 21.349479 29.99179    0.00000000
#> 86  2021-05-24  9.6971978 19.334496 28.97179    0.00000000
#> 87  2021-05-25  9.3571726 19.619471 29.88180    0.00000000
#> 88  2021-05-26 11.9771661 22.364468 32.75180    0.00000000
#> 89  2021-05-27 13.8171641 24.374474 34.93178    0.00000000
#> 90  2021-05-28 15.3971820 25.709500 36.02181    0.00000000
#> 91  2021-05-29 15.3072145 26.029502 36.75179    0.00000000
#> 92  2021-05-30 17.0840708 27.603539 38.12304    0.00000000
#> 93  2021-05-31 12.4751239 23.293788 34.11248    0.00000000
#> 94  2021-06-01 14.9651125 25.843777 36.72247    0.00000000
#> 95  2021-06-02 17.5051202 27.878808 38.25247    0.00000000
#> 96  2021-06-03 17.7551227 28.013819 38.27248    0.00000000
#> 97  2021-06-04 17.7551524 28.268823 38.78249    0.00000000
#> 98  2021-06-05 18.1051297 28.388816 38.67248    0.00000000
#> 99  2021-06-06 19.0851381 28.898828 38.71252    0.00000000
#> 100 2021-06-07 18.5151333 28.508814 38.50246    0.00000000
#> 101 2021-06-08 19.2251248 29.003809 38.78249    0.00000000
#> 102 2021-06-09 18.2251231 28.253810 38.28249    0.00000000
#> 103 2021-06-10 18.0251736 27.138816 36.25246    0.00000000
#> 104 2021-06-11 16.9151264 26.003808 35.09252    0.50837252
#> 105 2021-06-12 17.8351381 26.078822 34.32247    1.10701554
#> 106 2021-06-13 17.4251344 25.573815 33.72247    0.24943562
#> 107 2021-06-14 17.1251170 26.003808 34.88247    0.00000000
#> 108 2021-06-15 16.4051141 26.433801 36.46249    0.00000000
#> 109 2021-06-16 17.2351625 27.438838 37.64251    0.00000000
#> 110 2021-06-17 18.0351173 28.133816 38.23248    0.00000000
#> 111 2021-06-18 18.6351267 28.848809 39.06249    0.00000000
#> 112 2021-06-19 19.4251336 29.558801 39.69247    0.09977488
#> 113 2021-06-20 19.9951426 28.088828 36.18249    0.42997873
#> 114 2021-06-21 17.7951304 26.853845 35.91253    0.10215045
#> 115 2021-06-22 16.4251349 25.583827 34.74248    0.87658536
#> 116 2021-06-23 16.3451174 26.133813 35.92248    1.03574852
#> 117 2021-06-24 16.7551210 25.748804 34.74248    0.78631391
#> 118 2021-06-25 16.1751344 24.758814 33.34249    0.15441293
#> 119 2021-06-26 16.3051393 24.668816 33.03249    0.19717314
#> 120 2021-06-27 14.8751153 24.293789 33.71249    0.00000000
#> 121 2021-06-28 18.1451331 27.353812 36.56249    1.25430065
#> 122 2021-06-29 20.7710520 27.335476 33.89993    3.65599959
#> 123 2021-06-30 19.6926795 29.245877 38.79910    0.00000000
#> 124 2021-07-01 19.7527398 29.170922 38.58910    0.00000000
#> 125 2021-07-02 19.1527049 27.900903 36.64910    0.00000000
#> 126 2021-07-03 18.6026849 28.615901 38.62911    0.00000000
#> 127 2021-07-04 19.4526927 29.050929 38.64916    1.30006959
#> 128 2021-07-05 20.0426907 29.420895 38.79910    0.38489321
#> 129 2021-07-06 20.5327394 30.030908 39.52911    0.00000000
#> 130 2021-07-07 20.3526857 29.295892 38.23910    0.00000000
#> 131 2021-07-08 18.3826870 27.430903 36.47912    0.00000000
#> 132 2021-07-09 19.5726861 28.570887 37.56911    0.00000000
#> 133 2021-07-10 20.6026883 29.605921 38.60915    0.24376786
#> 134 2021-07-11 21.2427012 29.950892 38.65911    0.00000000
#> 135 2021-07-12 20.8026971 29.185910 37.56911    0.00000000
#> 136 2021-07-13 20.2826775 28.755917 37.22915    2.51887938
#> 137 2021-07-14 20.4627016 28.470908 36.47912    1.21881565
#> 138 2021-07-15 19.2526796 27.505883 35.75912    0.53457156
#> 139 2021-07-16 18.2427004 26.675928 35.10913    0.88524671
#> 140 2021-07-17 17.3126772 25.710899 34.10912    2.35637132
#> 141 2021-07-18 19.2026944 26.005885 32.80910   16.58437258
#> 142 2021-07-19 17.0327399 23.885922 30.73910    4.01993957
#> 143 2021-07-20 17.8526857 25.280906 32.70910    4.17389433
#> 144 2021-07-21 17.4826898 24.895901 32.30910    0.67142035
#> 145 2021-07-22 17.1227052 25.500912 33.87911    0.00000000
#> 146 2021-07-23 16.6326862 25.585904 34.53912    0.00000000
#> 147 2021-07-24 17.6126946 26.740901 35.86910    0.00000000
#> 148 2021-07-25 18.8026979 26.125912 33.44912    4.50318739
#> 149 2021-07-26 17.1726947 24.660916 32.14910    5.66640180
#> 150 2021-07-27 17.0726861 24.595913 32.11910    4.78543798
#> 151 2021-07-28 17.8427039 25.340907 32.83910    2.63434555
#> 152 2021-07-29 17.6927439 25.730951 33.76916    8.86951978
#> 153 2021-07-30 16.9021246 25.285483 33.66887    0.00000000
#> 154 2021-07-31 15.6188922 22.944421 30.26995    0.00000000
#> 155 2021-08-01 18.3189053 25.399441 32.47997    1.71793467
#> 156 2021-08-02 19.3688880 25.559413 31.74996    4.70217559
#> 157 2021-08-03 18.6689419 23.644435 28.61996   10.90532393
#> 158 2021-08-04 18.1889004 25.194452 32.20000    0.00000000
#> 159 2021-08-05 19.9589199 26.634454 33.30999    0.80583716
#> 160 2021-08-06 20.4288907 27.104424 33.77996    0.82797539
#> 161 2021-08-07 22.1188906 28.419427 34.71996    0.00000000
#> 162 2021-08-08 21.8488990 28.064448 34.27999    8.29300534
#> 163 2021-08-09 20.7788934 27.284448 33.78997    0.00000000
#> 164 2021-08-10 20.4288907 26.569422 32.70995    3.44472108
#> 165 2021-08-11 19.4188775 26.444422 33.46996    0.00000000
#> 166 2021-08-12 19.9588860 25.849448 31.74001    0.70400106
#> 167 2021-08-13 19.3189028 26.224450 33.13000    0.00000000
#> 168 2021-08-14 19.2289056 25.489434 31.74996    0.63758613
#> 169 2021-08-15 19.1589016 24.079429 28.99999    6.50423220
#> 170 2021-08-16 18.2088894 24.759424 31.30999    4.03359925
#> 171 2021-08-17 19.6888987 26.769435 33.84997    1.37257730
#> 172 2021-08-18 20.8289422 27.379447 33.92995    2.51048524
#> 173 2021-08-19 20.8288829 27.234463 33.64001    2.39536614
#> 174 2021-08-20 20.3288820 26.599418 32.86996    0.95194981
#> 175 2021-08-21 21.4388953 27.159449 32.87997    3.82107173
#> 176 2021-08-22 20.4188792 27.294430 34.16997    0.56674369
#> 177 2021-08-23 20.2688853 25.854424 31.43996    5.07852662
#> 178 2021-08-24 18.1189219 24.859458 31.59997    0.00000000
#> 179 2021-08-25 18.5089027 25.359429 32.20996    0.00000000
#> 180 2021-08-26 18.7088860 26.294429 33.87997    0.00000000
#> 181 2021-08-27 20.1889292 27.769436 35.34997    0.00000000
#> 182 2021-08-28 19.8888821 26.779446 33.66998    0.00000000
#> 183 2021-08-29 17.7088886 24.769431 31.82995    0.21695872
#> 184 2021-08-30 16.5057331 24.455287 32.40481    0.00000000
#> 185 2021-08-31 16.6461436 24.992362 33.33858    0.00000000
#> 186 2021-09-01 17.0961253 25.332362 33.56862    7.81488454
#> 187 2021-09-02 17.3961427 23.262384 29.12863   11.00299315
#> 188 2021-09-03 14.3561342 21.702355 29.04861    2.39395040
#> 189 2021-09-04 15.0561778 23.862389 32.66860    0.00000000
#> 190 2021-09-05 16.2461493 25.197351 34.14859    1.31782018
#> 191 2021-09-06 16.8861347 25.167380 33.44860    0.00000000
#> 192 2021-09-07 17.0561769 25.722377 34.38861    0.00000000
#> 193 2021-09-08 17.4461322 25.477380 33.50860    1.28904675
#> 194 2021-09-09 16.8461249 25.292384 33.73861    1.25451838
#> 195 2021-09-10 16.9061534 24.897363 32.88860    0.00000000
#> 196 2021-09-11 16.3961432 24.522366 32.64858    5.62234108
#> 197 2021-09-12 15.0261476 23.222377 31.41860    0.00000000
#> 198 2021-09-13 14.8861334 22.587365 30.28860    0.85744363
#> 199 2021-09-14 13.3061473 21.672389 30.03860    0.00000000
#> 200 2021-09-15 14.5661248 23.172357 31.77859    0.00000000
#> 201 2021-09-16 14.8061456 23.342370 31.87859    0.00000000
#> 202 2021-09-17 13.8661448 22.882351 31.89858    0.00000000
#> 203 2021-09-18 14.1661326 23.017358 31.86858    0.00000000
#> 204 2021-09-19 14.1761420 22.967368 31.75860    0.00000000
#> 205 2021-09-20 15.2661387 22.792384 30.31860    0.00000000
#> 206 2021-09-21 12.6961305 21.227383 29.75860    0.00000000
#> 207 2021-09-22 13.2461505 22.107383 30.96865    0.00000000
#> 208 2021-09-23 14.1561529 22.607388 31.05859    1.03008490
#> 209 2021-09-24 15.5061298 23.427361 31.34860    0.38555755
#> 210 2021-09-25 14.9361504 23.492398 32.04861    0.00000000
#> 211 2021-09-26 14.5061281 23.662381 32.81860    0.00000000
#> 212 2021-09-27 15.3161261 24.527367 33.73861    0.00000000
#> 213 2021-09-28 15.3761546 24.527401 33.67864    0.00000000
#> 214 2021-09-29 15.5640496 24.261865 32.95965    0.00000000

 # }