Combine (ensemble) casts

Combine multiple casts' output into a single ensemble. Presently, only a general average ensemble is available.

ensemble_casts(
  main = ".",
  method = "unwtavg",
  cast_groups = NULL,
  cast_ids = NULL,
  cast_tab = NULL,
  end_moon = NULL,
  models = NULL,
  data_set = NULL,
  include_interp = TRUE,
  species = NULL,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

method
character value of the name of the ensemble method to use. Presently, only "unwtavg" (unweighted average) is allowed.

cast_groups
integer (or integer numeric) value of the cast group to combine with an ensemble. If NULL (default), the most recent cast group is ensembled.

cast_ids
integer (or integer numeric) values representing the casts of interest for restricting ensembling, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information.

cast_tab
Optional data.frame of cast table outputs. If not input, will be loaded.

end_moon
integer (or integer numeric) newmoon number of the forecast origin. Default value is NULL, which equates to no selection with respect to end_moon.

models
character value(s) of the name of the model to include. Default value is NULL, which equates to no selection with respect to model. NULL translates to all models in the table.

data_set
character value of the rodent data set to include Default value is NULL, which equates to the first data set encountered.

include_interp
logical indicator of if the basic data set names should also be inclusive of the associated interpolated data sets.

species
character vector of the species code(s) or "total" for the total across species) to be plotted NULL translates to the species defined by evalplot_species.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

NULL. Plot is generated.

Details

A pre-loaded table of casts can be input, but if not (default), the table will be efficiently (as defined by the inputs) loaded and trimmed.
The casts can be trimmed specifically using the cast_ids input, otherwise, all relevant casts from the stated cast_groups will be included.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  portalcast(models = c("AutoArima", "ESSS"), end_moon = 515)

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Preparing directory for casting

#> ------------------------------------------------------------

#> Checking model availability

#>  *All requested models available*

#> ------------------------------------------------------------

#> Readying data for forecast origin newmoon 515

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Running models for forecast origin newmoon 515

#>  -Running AutoArima

#>   -AutoArima for all

#>   -AutoArima for controls

#>   -AutoArima for exclosures

#>   -AutoArima for dm_controls

#>   |++++| AutoArima successful |++++|

#>  -Running ESSS

#>   -ESSS for all_interp

#>   -ESSS for controls_interp

#>   -ESSS for exclosures_interp

#>   -ESSS for dm_controls_interp

#>   |++++| ESSS successful |++++|

#> ------------------------------------------------------------

#> Clearing tmp subdirectory

#> ------------------------------------------------------------

#> Resetting data to the most up-to-date versions

#> ------------------------------------------------------------

#> Casting complete

#> ------------------------------------------------------------

  ensemble_casts()

#>                model moon species     estimate          var      lower_pi
#> 1   ensemble_unwtavg  527      BA   0.25624896 4.558914e+00    -1.8248489
#> 2   ensemble_unwtavg  528      BA   0.52011488 2.025593e+01    -3.8665845
#> 3   ensemble_unwtavg  529      BA   0.58762365 2.768499e+01    -4.5408008
#> 4   ensemble_unwtavg  530      BA   0.62535855 3.133970e+01    -4.8310793
#> 5   ensemble_unwtavg  531      BA   0.63809927 3.555555e+01    -5.1737656
#> 6   ensemble_unwtavg  532      BA   0.63115526 3.725295e+01    -5.3178196
#> 7   ensemble_unwtavg  533      BA   0.60647376 3.889298e+01    -5.4720402
#> 8   ensemble_unwtavg  534      BA   0.57806428 3.883875e+01    -5.4962106
#> 9   ensemble_unwtavg  535      BA   0.55702702 4.035085e+01    -5.6343629
#> 10  ensemble_unwtavg  536      BA   0.54763229 4.203221e+01    -5.7714343
#> 11  ensemble_unwtavg  537      BA   0.55690820 4.638984e+01    -6.0816418
#> 12  ensemble_unwtavg  538      BA   0.58408135 4.959042e+01    -6.2796563
#> 13  ensemble_unwtavg  539      BA   0.62458779 5.406785e+01    -6.5423122
#> 14  ensemble_unwtavg  527      DM  49.69937320 2.168539e+04   -93.8315444
#> 15  ensemble_unwtavg  528      DM 117.24049499 4.988723e+05  -571.1842759
#> 16  ensemble_unwtavg  529      DM 122.72093565 5.529140e+05  -602.0330330
#> 17  ensemble_unwtavg  530      DM 124.80785148 5.782536e+05  -616.3675149
#> 18  ensemble_unwtavg  531      DM 124.77123642 5.781740e+05  -616.3531046
#> 19  ensemble_unwtavg  532      DM 123.45030368 5.665819e+05  -610.2068717
#> 20  ensemble_unwtavg  533      DM 122.05620975 5.544436e+05  -603.6995391
#> 21  ensemble_unwtavg  534      DM 120.83472379 5.438789e+05  -597.9733102
#> 22  ensemble_unwtavg  535      DM 120.38551681 5.400650e+05  -595.8977791
#> 23  ensemble_unwtavg  536      DM 120.54032702 5.414496e+05  -596.6605595
#> 24  ensemble_unwtavg  537      DM 121.50286229 5.498972e+05  -601.2711920
#> 25  ensemble_unwtavg  538      DM 122.85089546 5.618067e+05  -607.7080424
#> 26  ensemble_unwtavg  539      DM 124.22069549 5.740663e+05  -614.2662487
#> 27  ensemble_unwtavg  527      DO   2.01602587 7.994124e+01    -6.6985698
#> 28  ensemble_unwtavg  528      DO   5.29130011 1.144212e+03   -27.6784086
#> 29  ensemble_unwtavg  529      DO   7.72472342 2.372103e+03   -39.7463014
#> 30  ensemble_unwtavg  530      DO   9.12093703 3.288044e+03   -46.7686099
#> 31  ensemble_unwtavg  531      DO   9.87271597 3.842311e+03   -50.5441317
#> 32  ensemble_unwtavg  532      DO  10.29240790 4.181760e+03   -52.7367273
#> 33  ensemble_unwtavg  533      DO  10.54737441 4.407322e+03   -54.1593178
#> 34  ensemble_unwtavg  534      DO  10.81389187 4.636276e+03   -55.5522279
#> 35  ensemble_unwtavg  535      DO  11.15181051 4.938256e+03   -57.3415702
#> 36  ensemble_unwtavg  536      DO  11.59389526 5.334030e+03   -59.5912736
#> 37  ensemble_unwtavg  537      DO  12.13563347 5.840289e+03   -62.3511087
#> 38  ensemble_unwtavg  538      DO  12.74412524 6.431612e+03   -65.4225638
#> 39  ensemble_unwtavg  539      DO  13.42469162 7.123304e+03   -68.8379273
#> 40  ensemble_unwtavg  527      DS   0.16966715 2.251399e+00    -1.2928065
#> 41  ensemble_unwtavg  528      DS   0.29397501 7.018971e+00    -2.2882764
#> 42  ensemble_unwtavg  529      DS   0.30963793 8.007526e+00    -2.4484683
#> 43  ensemble_unwtavg  530      DS   0.31996280 8.910481e+00    -2.5894972
#> 44  ensemble_unwtavg  531      DS   0.33713427 9.959029e+00    -2.7387521
#> 45  ensemble_unwtavg  532      DS   0.36142738 1.122904e+01    -2.9046993
#> 46  ensemble_unwtavg  533      DS   0.38647214 1.255369e+01    -3.0669328
#> 47  ensemble_unwtavg  534      DS   0.41415519 1.401142e+01    -3.2342483
#> 48  ensemble_unwtavg  535      DS   0.44751928 1.571119e+01    -3.4158507
#> 49  ensemble_unwtavg  536      DS   0.48225145 1.756992e+01    -3.6032629
#> 50  ensemble_unwtavg  537      DS   0.51541484 1.948661e+01    -3.7871753
#> 51  ensemble_unwtavg  538      DS   0.55357459 2.172984e+01    -3.9899203
#> 52  ensemble_unwtavg  539      DS   0.60149584 2.457447e+01    -4.2302477
#> 53  ensemble_unwtavg  527      NA   0.13537888 2.415381e+00    -1.3794187
#> 54  ensemble_unwtavg  528      NA   0.24444166 6.491205e+00    -2.2388313
#> 55  ensemble_unwtavg  529      NA   0.27621711 7.694873e+00    -2.4275079
#> 56  ensemble_unwtavg  530      NA   0.31134118 9.039140e+00    -2.6190483
#> 57  ensemble_unwtavg  531      NA   0.36321911 1.089333e+01    -2.8537144
#> 58  ensemble_unwtavg  532      NA   0.41294122 1.290646e+01    -3.0886493
#> 59  ensemble_unwtavg  533      NA   0.46047389 1.630450e+01    -3.4751682
#> 60  ensemble_unwtavg  534      NA   0.44494526 1.611042e+01    -3.4672029
#> 61  ensemble_unwtavg  535      NA   0.39295770 1.426092e+01    -3.2877856
#> 62  ensemble_unwtavg  536      NA   0.35802452 1.398524e+01    -3.2869697
#> 63  ensemble_unwtavg  537      NA   0.32432997 1.384386e+01    -3.3021932
#> 64  ensemble_unwtavg  538      NA   0.30072077 1.400835e+01    -3.3472825
#> 65  ensemble_unwtavg  539      NA   0.30361919 1.472517e+01    -3.4365561
#> 66  ensemble_unwtavg  527      OL   0.58778239 1.191369e+01    -2.7764424
#> 67  ensemble_unwtavg  528      OL   1.19355659 7.004947e+01    -6.9640777
#> 68  ensemble_unwtavg  529      OL   1.27086298 8.200343e+01    -7.5554194
#> 69  ensemble_unwtavg  530      OL   1.20078408 7.996766e+01    -7.5152516
#> 70  ensemble_unwtavg  531      OL   1.13716873 7.940217e+01    -7.5479945
#> 71  ensemble_unwtavg  532      OL   1.09735897 8.008757e+01    -7.6252090
#> 72  ensemble_unwtavg  533      OL   1.11471225 8.566802e+01    -7.9066302
#> 73  ensemble_unwtavg  534      OL   1.19330210 9.783377e+01    -8.4473452
#> 74  ensemble_unwtavg  535      OL   1.29874941 1.125945e+02    -9.0436268
#> 75  ensemble_unwtavg  536      OL   1.43634742 1.315158e+02    -9.7413107
#> 76  ensemble_unwtavg  537      OL   1.59291315 1.564673e+02   -10.5990477
#> 77  ensemble_unwtavg  538      OL   1.75470384 1.819356e+02   -11.3921128
#> 78  ensemble_unwtavg  539      OL   1.81431428 1.960873e+02   -11.8342373
#> 79  ensemble_unwtavg  527      OT   8.21912786 6.321584e+02   -16.2870034
#> 80  ensemble_unwtavg  528      OT  22.05741438 1.779176e+04  -107.9509227
#> 81  ensemble_unwtavg  529      OT  21.23608570 1.654906e+04  -104.1497570
#> 82  ensemble_unwtavg  530      OT  21.17572279 1.651689e+04  -104.0881773
#> 83  ensemble_unwtavg  531      OT  21.41075016 1.695348e+04  -105.4979127
#> 84  ensemble_unwtavg  532      OT  21.72875265 1.755560e+04  -107.4138926
#> 85  ensemble_unwtavg  533      OT  22.45832740 1.886433e+04  -111.4114356
#> 86  ensemble_unwtavg  534      OT  23.27653311 2.039650e+04  -115.9235874
#> 87  ensemble_unwtavg  535      OT  24.16745985 2.214838e+04  -120.8875725
#> 88  ensemble_unwtavg  536      OT  24.91517821 2.380588e+04  -125.4696039
#> 89  ensemble_unwtavg  537      OT  25.39261325 2.505342e+04  -128.8823035
#> 90  ensemble_unwtavg  538      OT  25.47983926 2.561059e+04  -130.5011094
#> 91  ensemble_unwtavg  539      OT  25.39179569 2.595241e+04  -131.6266416
#> 92  ensemble_unwtavg  527      PB   0.96784795 4.457523e+01    -5.5395683
#> 93  ensemble_unwtavg  528      PB   2.44996307 3.159100e+02   -14.8738496
#> 94  ensemble_unwtavg  529      PB   3.62213169 6.362471e+02   -20.9631230
#> 95  ensemble_unwtavg  530      PB   4.46287470 9.495847e+02   -25.5721954
#> 96  ensemble_unwtavg  531      PB   5.36716169 1.341140e+03   -30.3271383
#> 97  ensemble_unwtavg  532      PB   6.35337054 1.843612e+03   -35.4967367
#> 98  ensemble_unwtavg  533      PB   6.85009242 2.146172e+03   -38.3036783
#> 99  ensemble_unwtavg  534      PB   7.31463792 2.445301e+03   -40.8832464
#> 100 ensemble_unwtavg  535      PB   7.44419730 2.556701e+03   -41.8393261
#> 101 ensemble_unwtavg  536      PB   7.20235424 2.460385e+03   -41.1439590
#> 102 ensemble_unwtavg  537      PB   6.99701144 2.380127e+03   -40.5542331
#> 103 ensemble_unwtavg  538      PB   6.45904998 2.131669e+03   -38.5419039
#> 104 ensemble_unwtavg  539      PB   6.17434284 2.035406e+03   -37.7987846
#> 105 ensemble_unwtavg  527      PE   0.59298821 1.181573e+01    -2.7573772
#> 106 ensemble_unwtavg  528      PE   3.82475499 5.716721e+02   -19.4795040
#> 107 ensemble_unwtavg  529      PE   3.18670142 4.052397e+02   -16.4341467
#> 108 ensemble_unwtavg  530      PE   3.15869235 4.039615e+02   -16.4311889
#> 109 ensemble_unwtavg  531      PE   3.32286877 4.473957e+02   -17.2932882
#> 110 ensemble_unwtavg  532      PE   3.12834742 4.041105e+02   -16.4651447
#> 111 ensemble_unwtavg  533      PE   2.94651597 3.678284e+02   -15.7467185
#> 112 ensemble_unwtavg  534      PE   2.90461407 3.660134e+02   -15.7424417
#> 113 ensemble_unwtavg  535      PE   3.12984483 4.260246e+02   -16.9878923
#> 114 ensemble_unwtavg  536      PE   3.37250361 4.985442e+02   -18.3902388
#> 115 ensemble_unwtavg  537      PE   3.73807532 6.078644e+02   -20.2925544
#> 116 ensemble_unwtavg  538      PE   3.77619872 6.356648e+02   -20.7978036
#> 117 ensemble_unwtavg  539      PE   3.83892071 6.742654e+02   -21.4702103
#> 118 ensemble_unwtavg  527      PF   0.53229335 1.679991e+01    -3.4626930
#> 119 ensemble_unwtavg  528      PF   1.15970695 8.326824e+01    -7.7343828
#> 120 ensemble_unwtavg  529      PF   1.46575105 1.268770e+02    -9.5130084
#> 121 ensemble_unwtavg  530      PF   1.70437338 1.699532e+02   -11.0021425
#> 122 ensemble_unwtavg  531      PF   1.89168191 2.072855e+02   -12.1411791
#> 123 ensemble_unwtavg  532      PF   2.04041198 2.390123e+02   -13.0281530
#> 124 ensemble_unwtavg  533      PF   2.17156066 2.705685e+02   -13.8609092
#> 125 ensemble_unwtavg  534      PF   2.28442152 2.990060e+02   -14.5695313
#> 126 ensemble_unwtavg  535      PF   2.33261574 3.149682e+02   -14.9653551
#> 127 ensemble_unwtavg  536      PF   2.37949898 3.315639e+02   -15.3683385
#> 128 ensemble_unwtavg  537      PF   2.45398026 3.548013e+02   -15.9052469
#> 129 ensemble_unwtavg  538      PF   2.56671263 3.863833e+02   -16.5922049
#> 130 ensemble_unwtavg  539      PF   2.71589014 4.267757e+02   -17.4195730
#> 131 ensemble_unwtavg  527      PH   0.00000000 0.000000e+00     0.0000000
#> 132 ensemble_unwtavg  528      PH   0.00000000 0.000000e+00     0.0000000
#> 133 ensemble_unwtavg  529      PH   0.00000000 0.000000e+00     0.0000000
#> 134 ensemble_unwtavg  530      PH   0.00000000 0.000000e+00     0.0000000
#> 135 ensemble_unwtavg  531      PH   0.00000000 0.000000e+00     0.0000000
#> 136 ensemble_unwtavg  532      PH   0.00000000 0.000000e+00     0.0000000
#> 137 ensemble_unwtavg  533      PH   0.00000000 0.000000e+00     0.0000000
#> 138 ensemble_unwtavg  534      PH   0.00000000 0.000000e+00     0.0000000
#> 139 ensemble_unwtavg  535      PH   0.00000000 0.000000e+00     0.0000000
#> 140 ensemble_unwtavg  536      PH   0.00000000 0.000000e+00     0.0000000
#> 141 ensemble_unwtavg  537      PH   0.00000000 0.000000e+00     0.0000000
#> 142 ensemble_unwtavg  538      PH   0.00000000 0.000000e+00     0.0000000
#> 143 ensemble_unwtavg  539      PH   0.00000000 0.000000e+00     0.0000000
#> 144 ensemble_unwtavg  527      PL   0.04545367 1.762449e-01    -0.3637317
#> 145 ensemble_unwtavg  528      PL   0.08562199 7.158249e-01    -0.7390195
#> 146 ensemble_unwtavg  529      PL   0.08762807 9.417802e-01    -0.8582531
#> 147 ensemble_unwtavg  530      PL   0.08505477 1.023029e+00    -0.9007837
#> 148 ensemble_unwtavg  531      PL   0.08136672 8.625819e-01    -0.8238696
#> 149 ensemble_unwtavg  532      PL   0.07526725 9.243926e-01    -0.8618416
#> 150 ensemble_unwtavg  533      PL   0.07244641 1.006195e+00    -0.9052473
#> 151 ensemble_unwtavg  534      PL   0.07028862 1.091984e+00    -0.9482322
#> 152 ensemble_unwtavg  535      PL   0.06959639 1.185370e+00    -0.9915831
#> 153 ensemble_unwtavg  536      PL   0.06947105 1.281579e+00    -1.0339329
#> 154 ensemble_unwtavg  537      PL   0.07055200 1.383802e+00    -1.0760134
#> 155 ensemble_unwtavg  538      PL   0.07227330 1.489422e+00    -1.1172437
#> 156 ensemble_unwtavg  539      PL   0.07673686 1.610327e+00    -1.1601181
#> 157 ensemble_unwtavg  527      PM   0.23707744 3.779874e+00    -1.6578844
#> 158 ensemble_unwtavg  528      PM   0.45950305 1.476774e+01    -3.2860745
#> 159 ensemble_unwtavg  529      PM   0.51343184 1.893381e+01    -3.7276902
#> 160 ensemble_unwtavg  530      PM   0.53976541 2.216896e+01    -4.0494084
#> 161 ensemble_unwtavg  531      PM   0.55352154 2.493493e+01    -4.3135290
#> 162 ensemble_unwtavg  532      PM   0.55282957 2.683859e+01    -4.4965919
#> 163 ensemble_unwtavg  533      PM   0.54060172 2.798558e+01    -4.6155889
#> 164 ensemble_unwtavg  534      PM   0.52807525 2.883906e+01    -4.7061493
#> 165 ensemble_unwtavg  535      PM   0.51614498 2.935363e+01    -4.7645702
#> 166 ensemble_unwtavg  536      PM   0.51210561 2.987449e+01    -4.8152543
#> 167 ensemble_unwtavg  537      PM   0.51744126 3.171575e+01    -4.9716356
#> 168 ensemble_unwtavg  538      PM   0.53030470 3.384195e+01    -5.1397793
#> 169 ensemble_unwtavg  539      PM   0.54828525 3.683013e+01    -5.3668333
#> 170 ensemble_unwtavg  527      PP   3.12740676 5.667739e+02   -20.0767999
#> 171 ensemble_unwtavg  528      PP   8.75814309 4.031912e+03   -53.1314054
#> 172 ensemble_unwtavg  529      PP  17.46160623 1.304255e+04   -93.8506502
#> 173 ensemble_unwtavg  530      PP  30.15119490 3.599091e+04  -154.7578679
#> 174 ensemble_unwtavg  531      PP  45.21814050 7.884427e+04  -228.4642623
#> 175 ensemble_unwtavg  532      PP  59.96765164 1.376239e+05  -301.6159438
#> 176 ensemble_unwtavg  533      PP  71.31054305 1.944537e+05  -358.4929265
#> 177 ensemble_unwtavg  534      PP  79.51593472 2.423894e+05  -400.3485227
#> 178 ensemble_unwtavg  535      PP  78.24521841 2.366513e+05  -395.9053390
#> 179 ensemble_unwtavg  536      PP  72.38422757 2.056553e+05  -369.6254415
#> 180 ensemble_unwtavg  537      PP  66.46418027 1.777221e+05  -344.4324398
#> 181 ensemble_unwtavg  538      PP  57.47907906 1.383856e+05  -305.1038182
#> 182 ensemble_unwtavg  539      PP  48.04338548 1.027218e+05  -264.3436828
#> 183 ensemble_unwtavg  527      RF   0.00000000 0.000000e+00     0.0000000
#> 184 ensemble_unwtavg  528      RF   0.00000000 0.000000e+00     0.0000000
#> 185 ensemble_unwtavg  529      RF   0.00000000 0.000000e+00     0.0000000
#> 186 ensemble_unwtavg  530      RF   0.00000000 0.000000e+00     0.0000000
#> 187 ensemble_unwtavg  531      RF   0.00000000 0.000000e+00     0.0000000
#> 188 ensemble_unwtavg  532      RF   0.00000000 0.000000e+00     0.0000000
#> 189 ensemble_unwtavg  533      RF   0.00000000 0.000000e+00     0.0000000
#> 190 ensemble_unwtavg  534      RF   0.00000000 0.000000e+00     0.0000000
#> 191 ensemble_unwtavg  535      RF   0.00000000 0.000000e+00     0.0000000
#> 192 ensemble_unwtavg  536      RF   0.00000000 0.000000e+00     0.0000000
#> 193 ensemble_unwtavg  537      RF   0.00000000 0.000000e+00     0.0000000
#> 194 ensemble_unwtavg  538      RF   0.00000000 0.000000e+00     0.0000000
#> 195 ensemble_unwtavg  539      RF   0.00000000 0.000000e+00     0.0000000
#> 196 ensemble_unwtavg  527      RM   0.73230658 1.747406e+01    -3.3420470
#> 197 ensemble_unwtavg  528      RM   1.59266544 1.285151e+02    -9.4567384
#> 198 ensemble_unwtavg  529      RM   1.89577802 1.774864e+02   -11.0892921
#> 199 ensemble_unwtavg  530      RM   1.56689579 1.329932e+02    -9.6733670
#> 200 ensemble_unwtavg  531      RM   1.38048540 1.084409e+02    -8.7693340
#> 201 ensemble_unwtavg  532      RM   1.48201088 1.242730e+02    -9.3835001
#> 202 ensemble_unwtavg  533      RM   1.70576273 1.567886e+02   -10.4987090
#> 203 ensemble_unwtavg  534      RM   2.15925870 2.308628e+02   -12.6501874
#> 204 ensemble_unwtavg  535      RM   2.75843035 3.556138e+02   -15.6218058
#> 205 ensemble_unwtavg  536      RM   4.32848994 8.291339e+02   -23.7370984
#> 206 ensemble_unwtavg  537      RM   3.55464390 5.928096e+02   -20.1765407
#> 207 ensemble_unwtavg  538      RM   2.01888481 2.328896e+02   -12.8554253
#> 208 ensemble_unwtavg  539      RM   1.96094786 2.355668e+02   -12.9986115
#> 209 ensemble_unwtavg  527      RO   0.01841184 2.337485e-01    -0.4528217
#> 210 ensemble_unwtavg  528      RO   0.04027774 5.928892e-01    -0.7102186
#> 211 ensemble_unwtavg  529      RO   0.06106298 1.370058e+00    -1.0797943
#> 212 ensemble_unwtavg  530      RO   0.11625737 1.606666e+00    -1.1191911
#> 213 ensemble_unwtavg  531      RO   0.08466793 1.470713e+00    -1.0973545
#> 214 ensemble_unwtavg  532      RO   0.16558267 3.382688e+00    -1.6270561
#> 215 ensemble_unwtavg  533      RO   0.08921344 2.162712e+00    -1.3441660
#> 216 ensemble_unwtavg  534      RO   0.08971037 2.384360e+00    -1.4153284
#> 217 ensemble_unwtavg  535      RO   0.09210539 2.729330e+00    -1.5181317
#> 218 ensemble_unwtavg  536      RO   0.10270118 2.693387e+00    -1.4968979
#> 219 ensemble_unwtavg  537      RO   0.12082412 3.066583e+00    -1.5860014
#> 220 ensemble_unwtavg  538      RO   0.21338457 4.732704e+00    -1.9070090
#> 221 ensemble_unwtavg  539      RO   0.29666175 6.874086e+00    -2.2587993
#> 222 ensemble_unwtavg  527      SF   0.25365427 3.622372e+00    -1.6014072
#> 223 ensemble_unwtavg  528      SF   0.58531517 2.052886e+01    -3.8308385
#> 224 ensemble_unwtavg  529      SF   0.72728725 3.049210e+01    -4.6548589
#> 225 ensemble_unwtavg  530      SF   0.76249952 3.438567e+01    -4.9529519
#> 226 ensemble_unwtavg  531      SF   0.76867199 3.523940e+01    -5.0172968
#> 227 ensemble_unwtavg  532      SF   0.77311024 3.746410e+01    -5.1927003
#> 228 ensemble_unwtavg  533      SF   0.73868655 3.569471e+01    -5.0845405
#> 229 ensemble_unwtavg  534      SF   0.74950412 3.735137e+01    -5.2073242
#> 230 ensemble_unwtavg  535      SF   0.79567742 4.165323e+01    -5.4948367
#> 231 ensemble_unwtavg  536      SF   0.87413472 4.791850e+01    -5.8729071
#> 232 ensemble_unwtavg  537      SF   0.99159603 5.759214e+01    -6.4051961
#> 233 ensemble_unwtavg  538      SF   1.17455690 7.452093e+01    -7.2394124
#> 234 ensemble_unwtavg  539      SF   1.45973073 1.057272e+02    -8.5622894
#> 235 ensemble_unwtavg  527      SH   0.21869388 2.496287e+00    -1.3212649
#> 236 ensemble_unwtavg  528      SH   0.39288741 1.039135e+01    -2.7490523
#> 237 ensemble_unwtavg  529      SH   0.49633243 1.540751e+01    -3.3295187
#> 238 ensemble_unwtavg  530      SH   0.40490606 1.227037e+01    -3.0093070
#> 239 ensemble_unwtavg  531      SH   0.41994077 1.397452e+01    -3.2236552
#> 240 ensemble_unwtavg  532      SH   0.54316394 1.970437e+01    -3.7833993
#> 241 ensemble_unwtavg  533      SH   0.48328427 1.784711e+01    -3.6343313
#> 242 ensemble_unwtavg  534      SH   0.51423615 1.994116e+01    -3.8382461
#> 243 ensemble_unwtavg  535      SH   0.50108891 2.062318e+01    -3.9251982
#> 244 ensemble_unwtavg  536      SH   0.55411555 2.377743e+01    -4.1986268
#> 245 ensemble_unwtavg  537      SH   0.60429803 2.710169e+01    -4.4698135
#> 246 ensemble_unwtavg  538      SH   0.67471845 3.157489e+01    -4.8021555
#> 247 ensemble_unwtavg  539      SH   0.77435474 3.826572e+01    -5.2549430
#> 248 ensemble_unwtavg  527      SO   0.00000000 0.000000e+00     0.0000000
#> 249 ensemble_unwtavg  528      SO   0.00000000 0.000000e+00     0.0000000
#> 250 ensemble_unwtavg  529      SO   0.00000000 0.000000e+00     0.0000000
#> 251 ensemble_unwtavg  530      SO   0.00000000 0.000000e+00     0.0000000
#> 252 ensemble_unwtavg  531      SO   0.00000000 0.000000e+00     0.0000000
#> 253 ensemble_unwtavg  532      SO   0.00000000 0.000000e+00     0.0000000
#> 254 ensemble_unwtavg  533      SO   0.00000000 0.000000e+00     0.0000000
#> 255 ensemble_unwtavg  534      SO   0.00000000 0.000000e+00     0.0000000
#> 256 ensemble_unwtavg  535      SO   0.00000000 0.000000e+00     0.0000000
#> 257 ensemble_unwtavg  536      SO   0.00000000 0.000000e+00     0.0000000
#> 258 ensemble_unwtavg  537      SO   0.00000000 0.000000e+00     0.0000000
#> 259 ensemble_unwtavg  538      SO   0.00000000 0.000000e+00     0.0000000
#> 260 ensemble_unwtavg  539      SO   0.00000000 0.000000e+00     0.0000000
#> 261 ensemble_unwtavg  527   total  66.66369096 3.972702e+04  -127.6058759
#> 262 ensemble_unwtavg  528   total 173.20351332 1.092426e+06  -845.5236520
#> 263 ensemble_unwtavg  529   total 193.52108520 1.371691e+06  -948.0156237
#> 264 ensemble_unwtavg  530   total 202.67247813 1.503623e+06  -992.5019734
#> 265 ensemble_unwtavg  531   total 211.94751918 1.643680e+06 -1037.6507868
#> 266 ensemble_unwtavg  532   total 219.70397635 1.765494e+06 -1075.3710316
#> 267 ensemble_unwtavg  533   total 222.16344991 1.805133e+06 -1087.3694368
#> 268 ensemble_unwtavg  534   total 222.20702862 1.806082e+06 -1087.6698510
#> 269 ensemble_unwtavg  535   total 223.19356716 1.822289e+06 -1092.5474618
#> 270 ensemble_unwtavg  536   total 219.87697886 1.769208e+06 -1076.5596275
#> 271 ensemble_unwtavg  537   total 218.55133330 1.748758e+06 -1070.3707858
#> 272 ensemble_unwtavg  538   total 216.12292414 1.711161e+06 -1058.8683664
#> 273 ensemble_unwtavg  539   total 214.75591366 1.690388e+06 -1052.4727547
#>         upper_pi obs       error end_moon lead data_set cast_id covered
#> 1      2.3373468   0  0.25624896      527    0 controls 9999413    TRUE
#> 2      4.9068142  NA          NA      527    1 controls 9999413    TRUE
#> 3      5.7160481  NA          NA      527    2 controls 9999413    TRUE
#> 4      6.0817964  NA          NA      527    3 controls 9999413    TRUE
#> 5      6.4499642  NA          NA      527    4 controls 9999413    TRUE
#> 6      6.5801301  NA          NA      527    5 controls 9999413    TRUE
#> 7      6.6849877  NA          NA      527    6 controls 9999413    TRUE
#> 8      6.6523392  NA          NA      527    7 controls 9999413    TRUE
#> 9      6.7484169  NA          NA      527    8 controls 9999413    TRUE
#> 10     6.8666989  NA          NA      527    9 controls 9999413    TRUE
#> 11     7.1954582  NA          NA      527   10 controls 9999413    TRUE
#> 12     7.4478190  NA          NA      527   11 controls 9999413    TRUE
#> 13     7.7914878  NA          NA      527   12 controls 9999413    TRUE
#> 14   193.2302908  24 25.69937320      527    0 controls 9999413    TRUE
#> 15   805.6652659  NA          NA      527    1 controls 9999413    TRUE
#> 16   847.4749043  NA          NA      527    2 controls 9999413    TRUE
#> 17   865.9832179  NA          NA      527    3 controls 9999413    TRUE
#> 18   865.8955774  NA          NA      527    4 controls 9999413    TRUE
#> 19   857.1074791  NA          NA      527    5 controls 9999413    TRUE
#> 20   847.8119586  NA          NA      527    6 controls 9999413    TRUE
#> 21   839.6427578  NA          NA      527    7 controls 9999413    TRUE
#> 22   836.6688127  NA          NA      527    8 controls 9999413    TRUE
#> 23   837.7412136  NA          NA      527    9 controls 9999413    TRUE
#> 24   844.2769166  NA          NA      527   10 controls 9999413    TRUE
#> 25   853.4098333  NA          NA      527   11 controls 9999413    TRUE
#> 26   862.7076397  NA          NA      527   12 controls 9999413    TRUE
#> 27    10.7306215   0  2.01602587      527    0 controls 9999413    TRUE
#> 28    38.2610089  NA          NA      527    1 controls 9999413    TRUE
#> 29    55.1957483  NA          NA      527    2 controls 9999413    TRUE
#> 30    65.0104840  NA          NA      527    3 controls 9999413    TRUE
#> 31    70.2895636  NA          NA      527    4 controls 9999413    TRUE
#> 32    73.3215431  NA          NA      527    5 controls 9999413    TRUE
#> 33    75.2540666  NA          NA      527    6 controls 9999413    TRUE
#> 34    77.1800117  NA          NA      527    7 controls 9999413    TRUE
#> 35    79.6451913  NA          NA      527    8 controls 9999413    TRUE
#> 36    82.7790641  NA          NA      527    9 controls 9999413    TRUE
#> 37    86.6223757  NA          NA      527   10 controls 9999413    TRUE
#> 38    90.9108143  NA          NA      527   11 controls 9999413    TRUE
#> 39    95.6873106  NA          NA      527   12 controls 9999413    TRUE
#> 40     1.6321408   0  0.16966715      527    0 controls 9999413    TRUE
#> 41     2.8762264  NA          NA      527    1 controls 9999413    TRUE
#> 42     3.0677442  NA          NA      527    2 controls 9999413    TRUE
#> 43     3.2294228  NA          NA      527    3 controls 9999413    TRUE
#> 44     3.4130207  NA          NA      527    4 controls 9999413    TRUE
#> 45     3.6275540  NA          NA      527    5 controls 9999413    TRUE
#> 46     3.8398770  NA          NA      527    6 controls 9999413    TRUE
#> 47     4.0625586  NA          NA      527    7 controls 9999413    TRUE
#> 48     4.3108893  NA          NA      527    8 controls 9999413    TRUE
#> 49     4.5677658  NA          NA      527    9 controls 9999413    TRUE
#> 50     4.8180050  NA          NA      527   10 controls 9999413    TRUE
#> 51     5.0970695  NA          NA      527   11 controls 9999413    TRUE
#> 52     5.4332393  NA          NA      527   12 controls 9999413    TRUE
#> 53     1.6501764   0  0.13537888      527    0 controls 9999413    TRUE
#> 54     2.7277146  NA          NA      527    1 controls 9999413    TRUE
#> 55     2.9799421  NA          NA      527    2 controls 9999413    TRUE
#> 56     3.2417307  NA          NA      527    3 controls 9999413    TRUE
#> 57     3.5801526  NA          NA      527    4 controls 9999413    TRUE
#> 58     3.9145317  NA          NA      527    5 controls 9999413    TRUE
#> 59     4.3961160  NA          NA      527    6 controls 9999413    TRUE
#> 60     4.3570934  NA          NA      527    7 controls 9999413    TRUE
#> 61     4.0737010  NA          NA      527    8 controls 9999413    TRUE
#> 62     4.0030187  NA          NA      527    9 controls 9999413    TRUE
#> 63     3.9508532  NA          NA      527   10 controls 9999413    TRUE
#> 64     3.9487241  NA          NA      527   11 controls 9999413    TRUE
#> 65     4.0437945  NA          NA      527   12 controls 9999413    TRUE
#> 66     3.9520072   0  0.58778239      527    0 controls 9999413    TRUE
#> 67     9.3511909  NA          NA      527    1 controls 9999413    TRUE
#> 68    10.0971454  NA          NA      527    2 controls 9999413    TRUE
#> 69     9.9168197  NA          NA      527    3 controls 9999413    TRUE
#> 70     9.8223319  NA          NA      527    4 controls 9999413    TRUE
#> 71     9.8199270  NA          NA      527    5 controls 9999413    TRUE
#> 72    10.1360547  NA          NA      527    6 controls 9999413    TRUE
#> 73    10.8339494  NA          NA      527    7 controls 9999413    TRUE
#> 74    11.6411256  NA          NA      527    8 controls 9999413    TRUE
#> 75    12.6140056  NA          NA      527    9 controls 9999413    TRUE
#> 76    13.7848740  NA          NA      527   10 controls 9999413    TRUE
#> 77    14.9015204  NA          NA      527   11 controls 9999413    TRUE
#> 78    15.4628659  NA          NA      527   12 controls 9999413    TRUE
#> 79    32.7252591   6  2.21912786      527    0 controls 9999413    TRUE
#> 80   152.0657515  NA          NA      527    1 controls 9999413    TRUE
#> 81   146.6219284  NA          NA      527    2 controls 9999413    TRUE
#> 82   146.4396229  NA          NA      527    3 controls 9999413    TRUE
#> 83   148.3194130  NA          NA      527    4 controls 9999413    TRUE
#> 84   150.8713979  NA          NA      527    5 controls 9999413    TRUE
#> 85   156.3280905  NA          NA      527    6 controls 9999413    TRUE
#> 86   162.4766536  NA          NA      527    7 controls 9999413    TRUE
#> 87   169.2224922  NA          NA      527    8 controls 9999413    TRUE
#> 88   175.2999604  NA          NA      527    9 controls 9999413    TRUE
#> 89   179.6675300  NA          NA      527   10 controls 9999413    TRUE
#> 90   181.4607880  NA          NA      527   11 controls 9999413    TRUE
#> 91   182.4102330  NA          NA      527   12 controls 9999413    TRUE
#> 92     7.4752642   0  0.96784795      527    0 controls 9999413    TRUE
#> 93    19.7737758  NA          NA      527    1 controls 9999413    TRUE
#> 94    28.2073864  NA          NA      527    2 controls 9999413    TRUE
#> 95    34.4979448  NA          NA      527    3 controls 9999413    TRUE
#> 96    41.0614617  NA          NA      527    4 controls 9999413    TRUE
#> 97    48.2034778  NA          NA      527    5 controls 9999413    TRUE
#> 98    52.0038632  NA          NA      527    6 controls 9999413    TRUE
#> 99    55.5125223  NA          NA      527    7 controls 9999413    TRUE
#> 100   56.7277206  NA          NA      527    8 controls 9999413    TRUE
#> 101   55.5486675  NA          NA      527    9 controls 9999413    TRUE
#> 102   54.5482560  NA          NA      527   10 controls 9999413    TRUE
#> 103   51.4600038  NA          NA      527   11 controls 9999413    TRUE
#> 104   50.1474703  NA          NA      527   12 controls 9999413    TRUE
#> 105    3.9433536   2 -1.40701179      527    0 controls 9999413    TRUE
#> 106   27.1290140  NA          NA      527    1 controls 9999413    TRUE
#> 107   22.8075496  NA          NA      527    2 controls 9999413    TRUE
#> 108   22.7485736  NA          NA      527    3 controls 9999413    TRUE
#> 109   23.9390257  NA          NA      527    4 controls 9999413    TRUE
#> 110   22.7218395  NA          NA      527    5 controls 9999413    TRUE
#> 111   21.6397504  NA          NA      527    6 controls 9999413    TRUE
#> 112   21.5516698  NA          NA      527    7 controls 9999413    TRUE
#> 113   23.2475820  NA          NA      527    8 controls 9999413    TRUE
#> 114   25.1352460  NA          NA      527    9 controls 9999413    TRUE
#> 115   27.7687050  NA          NA      527   10 controls 9999413    TRUE
#> 116   28.3502010  NA          NA      527   11 controls 9999413    TRUE
#> 117   29.1480517  NA          NA      527   12 controls 9999413    TRUE
#> 118    4.5272797   0  0.53229335      527    0 controls 9999413    TRUE
#> 119   10.0537967  NA          NA      527    1 controls 9999413    TRUE
#> 120   12.4445105  NA          NA      527    2 controls 9999413    TRUE
#> 121   14.4108893  NA          NA      527    3 controls 9999413    TRUE
#> 122   15.9245429  NA          NA      527    4 controls 9999413    TRUE
#> 123   17.1089770  NA          NA      527    5 controls 9999413    TRUE
#> 124   18.2040305  NA          NA      527    6 controls 9999413    TRUE
#> 125   19.1383743  NA          NA      527    7 controls 9999413    TRUE
#> 126   19.6305865  NA          NA      527    8 controls 9999413    TRUE
#> 127   20.1273365  NA          NA      527    9 controls 9999413    TRUE
#> 128   20.8132074  NA          NA      527   10 controls 9999413    TRUE
#> 129   21.7256301  NA          NA      527   11 controls 9999413    TRUE
#> 130   22.8513532  NA          NA      527   12 controls 9999413    TRUE
#> 131    0.0000000   0  0.00000000      527    0 controls 9999419    TRUE
#> 132    0.0000000  NA          NA      527    1 controls 9999419    TRUE
#> 133    0.0000000  NA          NA      527    2 controls 9999419    TRUE
#> 134    0.0000000  NA          NA      527    3 controls 9999419    TRUE
#> 135    0.0000000  NA          NA      527    4 controls 9999419    TRUE
#> 136    0.0000000  NA          NA      527    5 controls 9999419    TRUE
#> 137    0.0000000  NA          NA      527    6 controls 9999419    TRUE
#> 138    0.0000000  NA          NA      527    7 controls 9999419    TRUE
#> 139    0.0000000  NA          NA      527    8 controls 9999419    TRUE
#> 140    0.0000000  NA          NA      527    9 controls 9999419    TRUE
#> 141    0.0000000  NA          NA      527   10 controls 9999419    TRUE
#> 142    0.0000000  NA          NA      527   11 controls 9999419    TRUE
#> 143    0.0000000  NA          NA      527   12 controls 9999419    TRUE
#> 144    0.4546391   0  0.04545367      527    0 controls 9999413    TRUE
#> 145    0.9102635  NA          NA      527    1 controls 9999413    TRUE
#> 146    1.0335092  NA          NA      527    2 controls 9999413    TRUE
#> 147    1.0708932  NA          NA      527    3 controls 9999413    TRUE
#> 148    0.9866030  NA          NA      527    4 controls 9999413    TRUE
#> 149    1.0123761  NA          NA      527    5 controls 9999413    TRUE
#> 150    1.0501401  NA          NA      527    6 controls 9999413    TRUE
#> 151    1.0888094  NA          NA      527    7 controls 9999413    TRUE
#> 152    1.1307759  NA          NA      527    8 controls 9999413    TRUE
#> 153    1.1728750  NA          NA      527    9 controls 9999413    TRUE
#> 154    1.2171174  NA          NA      527   10 controls 9999413    TRUE
#> 155    1.2617903  NA          NA      527   11 controls 9999413    TRUE
#> 156    1.3135919  NA          NA      527   12 controls 9999413    TRUE
#> 157    2.1320392   0  0.23707744      527    0 controls 9999413    TRUE
#> 158    4.2050806  NA          NA      527    1 controls 9999413    TRUE
#> 159    4.7545538  NA          NA      527    2 controls 9999413    TRUE
#> 160    5.1289392  NA          NA      527    3 controls 9999413    TRUE
#> 161    5.4205721  NA          NA      527    4 controls 9999413    TRUE
#> 162    5.6022511  NA          NA      527    5 controls 9999413    TRUE
#> 163    5.6967923  NA          NA      527    6 controls 9999413    TRUE
#> 164    5.7622998  NA          NA      527    7 controls 9999413    TRUE
#> 165    5.7968602  NA          NA      527    8 controls 9999413    TRUE
#> 166    5.8394655  NA          NA      527    9 controls 9999413    TRUE
#> 167    6.0065181  NA          NA      527   10 controls 9999413    TRUE
#> 168    6.2003887  NA          NA      527   11 controls 9999413    TRUE
#> 169    6.4634038  NA          NA      527   12 controls 9999413    TRUE
#> 170   26.3316134   0  3.12740676      527    0 controls 9999413    TRUE
#> 171   70.6476916  NA          NA      527    1 controls 9999413    TRUE
#> 172  128.7738626  NA          NA      527    2 controls 9999413    TRUE
#> 173  215.0602577  NA          NA      527    3 controls 9999413    TRUE
#> 174  318.9005433  NA          NA      527    4 controls 9999413    TRUE
#> 175  421.5512470  NA          NA      527    5 controls 9999413    TRUE
#> 176  501.1140126  NA          NA      527    6 controls 9999413    TRUE
#> 177  559.3803921  NA          NA      527    7 controls 9999413    TRUE
#> 178  552.3957758  NA          NA      527    8 controls 9999413    TRUE
#> 179  514.3938966  NA          NA      527    9 controls 9999413    TRUE
#> 180  477.3608004  NA          NA      527   10 controls 9999413    TRUE
#> 181  420.0619763  NA          NA      527   11 controls 9999413    TRUE
#> 182  360.4304538  NA          NA      527   12 controls 9999413    TRUE
#> 183    0.0000000   0  0.00000000      527    0 controls 9999419    TRUE
#> 184    0.0000000  NA          NA      527    1 controls 9999419    TRUE
#> 185    0.0000000  NA          NA      527    2 controls 9999419    TRUE
#> 186    0.0000000  NA          NA      527    3 controls 9999419    TRUE
#> 187    0.0000000  NA          NA      527    4 controls 9999419    TRUE
#> 188    0.0000000  NA          NA      527    5 controls 9999419    TRUE
#> 189    0.0000000  NA          NA      527    6 controls 9999419    TRUE
#> 190    0.0000000  NA          NA      527    7 controls 9999419    TRUE
#> 191    0.0000000  NA          NA      527    8 controls 9999419    TRUE
#> 192    0.0000000  NA          NA      527    9 controls 9999419    TRUE
#> 193    0.0000000  NA          NA      527   10 controls 9999419    TRUE
#> 194    0.0000000  NA          NA      527   11 controls 9999419    TRUE
#> 195    0.0000000  NA          NA      527   12 controls 9999419    TRUE
#> 196    4.8066601   0  0.73230658      527    0 controls 9999413    TRUE
#> 197   12.6420693  NA          NA      527    1 controls 9999413    TRUE
#> 198   14.8808482  NA          NA      527    2 controls 9999413    TRUE
#> 199   12.8071586  NA          NA      527    3 controls 9999413    TRUE
#> 200   11.5303048  NA          NA      527    4 controls 9999413    TRUE
#> 201   12.3475219  NA          NA      527    5 controls 9999413    TRUE
#> 202   13.9102345  NA          NA      527    6 controls 9999413    TRUE
#> 203   16.9687048  NA          NA      527    7 controls 9999413    TRUE
#> 204   21.1386665  NA          NA      527    8 controls 9999413    TRUE
#> 205   32.3940783  NA          NA      527    9 controls 9999413    TRUE
#> 206   27.2858285  NA          NA      527   10 controls 9999413    TRUE
#> 207   16.8931950  NA          NA      527   11 controls 9999413    TRUE
#> 208   16.9205072  NA          NA      527   12 controls 9999413    TRUE
#> 209    0.4896454   0  0.01841184      527    0 controls 9999413    TRUE
#> 210    0.7907740  NA          NA      527    1 controls 9999413    TRUE
#> 211    1.2019202  NA          NA      527    2 controls 9999413    TRUE
#> 212    1.3517058  NA          NA      527    3 controls 9999413    TRUE
#> 213    1.2666903  NA          NA      527    4 controls 9999413    TRUE
#> 214    1.9582214  NA          NA      527    5 controls 9999413    TRUE
#> 215    1.5225929  NA          NA      527    6 controls 9999413    TRUE
#> 216    1.5947492  NA          NA      527    7 controls 9999413    TRUE
#> 217    1.7023425  NA          NA      527    8 controls 9999413    TRUE
#> 218    1.7023003  NA          NA      527    9 controls 9999413    TRUE
#> 219    1.8276497  NA          NA      527   10 controls 9999413    TRUE
#> 220    2.3337782  NA          NA      527   11 controls 9999413    TRUE
#> 221    2.8521228  NA          NA      527   12 controls 9999413    TRUE
#> 222    2.1087157   0  0.25365427      527    0 controls 9999413    TRUE
#> 223    5.0014689  NA          NA      527    1 controls 9999413    TRUE
#> 224    6.1094334  NA          NA      527    2 controls 9999413    TRUE
#> 225    6.4779509  NA          NA      527    3 controls 9999413    TRUE
#> 226    6.5546408  NA          NA      527    4 controls 9999413    TRUE
#> 227    6.7389207  NA          NA      527    5 controls 9999413    TRUE
#> 228    6.5619136  NA          NA      527    6 controls 9999413    TRUE
#> 229    6.7063325  NA          NA      527    7 controls 9999413    TRUE
#> 230    7.0861916  NA          NA      527    8 controls 9999413    TRUE
#> 231    7.6211765  NA          NA      527    9 controls 9999413    TRUE
#> 232    8.3883881  NA          NA      527   10 controls 9999413    TRUE
#> 233    9.5885262  NA          NA      527   11 controls 9999413    TRUE
#> 234   11.4817508  NA          NA      527   12 controls 9999413    TRUE
#> 235    1.7586526   0  0.21869388      527    0 controls 9999413    TRUE
#> 236    3.5348271  NA          NA      527    1 controls 9999413    TRUE
#> 237    4.3221836  NA          NA      527    2 controls 9999413    TRUE
#> 238    3.8191191  NA          NA      527    3 controls 9999413    TRUE
#> 239    4.0635367  NA          NA      527    4 controls 9999413    TRUE
#> 240    4.8697272  NA          NA      527    5 controls 9999413    TRUE
#> 241    4.6008998  NA          NA      527    6 controls 9999413    TRUE
#> 242    4.8667184  NA          NA      527    7 controls 9999413    TRUE
#> 243    4.9273760  NA          NA      527    8 controls 9999413    TRUE
#> 244    5.3068579  NA          NA      527    9 controls 9999413    TRUE
#> 245    5.6784095  NA          NA      527   10 controls 9999413    TRUE
#> 246    6.1515924  NA          NA      527   11 controls 9999413    TRUE
#> 247    6.8036525  NA          NA      527   12 controls 9999413    TRUE
#> 248    0.0000000   0  0.00000000      527    0 controls 9999419    TRUE
#> 249    0.0000000  NA          NA      527    1 controls 9999419    TRUE
#> 250    0.0000000  NA          NA      527    2 controls 9999419    TRUE
#> 251    0.0000000  NA          NA      527    3 controls 9999419    TRUE
#> 252    0.0000000  NA          NA      527    4 controls 9999419    TRUE
#> 253    0.0000000  NA          NA      527    5 controls 9999419    TRUE
#> 254    0.0000000  NA          NA      527    6 controls 9999419    TRUE
#> 255    0.0000000  NA          NA      527    7 controls 9999419    TRUE
#> 256    0.0000000  NA          NA      527    8 controls 9999419    TRUE
#> 257    0.0000000  NA          NA      527    9 controls 9999419    TRUE
#> 258    0.0000000  NA          NA      527   10 controls 9999419    TRUE
#> 259    0.0000000  NA          NA      527   11 controls 9999419    TRUE
#> 260    0.0000000  NA          NA      527   12 controls 9999419    TRUE
#> 261  260.9332579  32 34.66369096      527    0 controls 9999413    TRUE
#> 262 1191.9306786  NA          NA      527    1 controls 9999413    TRUE
#> 263 1335.0577941  NA          NA      527    2 controls 9999413    TRUE
#> 264 1397.8469297  NA          NA      527    3 controls 9999413    TRUE
#> 265 1461.5458251  NA          NA      527    4 controls 9999413    TRUE
#> 266 1514.7789843  NA          NA      527    5 controls 9999413    TRUE
#> 267 1531.6963366  NA          NA      527    6 controls 9999413    TRUE
#> 268 1532.0839082  NA          NA      527    7 controls 9999413    TRUE
#> 269 1538.9345961  NA          NA      527    8 controls 9999413    TRUE
#> 270 1516.3135853  NA          NA      527    9 controls 9999413    TRUE
#> 271 1507.4734524  NA          NA      527   10 controls 9999413    TRUE
#> 272 1491.1142147  NA          NA      527   11 controls 9999413    TRUE
#> 273 1481.9845821  NA          NA      527   12 controls 9999413    TRUE

# }