Produce the control lists for models

The models are written and managed using generalized functions that take a few arguments that vary among models, and so this function produces the appropriate list of lists of model script-writing control arguments. The specific models must be input by the user, with the default being NULL, which returns. NULL. The prefab models (prefab_models) have their control lists included already, but any user-created model needs to have controls provided. See Details.
Users adding models to the prefab set should add their script-writing controls lists to the prefab_controls object in the source code of the function.

extract_min_lag provides a simple function to extract the minimum non-0 lag time used across models from the controls lists.

model_controls(
  models = NULL,
  controls_model = NULL,
  quiet = FALSE,
  arg_checks = TRUE
)

extract_min_lag(
  models = prefab_models(),
  controls_model = NULL,
  quiet = FALSE,
  arg_checks = TRUE
)

extract_data_sets(
  models = prefab_models(),
  controls_model = NULL,
  quiet = FALSE,
  arg_checks = TRUE
)

Arguments

models
character vector of name(s) of model(s) whose script-writing controls need to be included in the list used to write the scripts.

controls_model
Additional controls for models not in the prefab set.
A list of a single model's script-writing controls or a list of lists, each of which is a single model's script-writing controls.
Presently, each model's script writing controls should include four elements:

name: a character value of the model name.

data_sets: a character vector of the data set names that the model is applied to.

covariatesTF: a logical indicator of if the model needs covariates.

lag: an integer-conformable value of the lag to use with the covariates or NA if covariatesTF = FALSE.

In addition, some models require additional specific elements:

max_E: integer (or integer numeric) maximum embedding dimension to search amongst for EDM models. See simplex for more information.

control_runjags: list of arguments passed to run.jags via runjags_control.

If only a single model is added, the name of the model from the element name will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the model and the name element.
See Details and Examples.

quiet
logical indicator if progress messages should be quieted.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

model_controls: named list of length equal to the number of elements in models and with elements that are each lists of those models's script-writing controls.

extract_min_lag: numeric value of the minimum non-0 lag from any included model or NA if no models have lags.

extract_data_sets: character vector of the data set names from all included models.

Details

Any model that is part of the prefab set (c("AutoArima", "NaiveArima", "ESSS", "nbGARCH", "nbsGARCH", "pevGARCH", "simplexEDM", "GPEDM", "jags_RW")) has its script-writing controls already included internally via the non-exported function prefab_model_controls. Users only need to include controls for non-prefab models.

Any user-defined models that are not included in controls_m will throw an error.

If any user-defined controls_model duplicate any existing controls for the prefab models or if controls_model contains any duplicate named elements, an error will be thrown unless arg_checks = FALSE, in which case, the first user-input element for any and all conflicting copies is selected. This override allows users to test an existing prefab model with different configurations, such as on a new data set, without requiring a new model.

Users interested in adding models to the prefab set should add the controls to the prefab_model_controls non-exported function found in the prepare_models.R script.

Examples

 model_controls(prefab_models())
 controls <- list(name = "xx", data_sets = prefab_data_sets(), 
                  interpolate = FALSE, covariatesTF = FALSE, lag = NA)
 model_controls("xx", controls)
 model_controls(c(prefab_models(), "xx"), controls)
 model_controls(c("xx", "ESSS"), controls)
 extract_min_lag()
 extract_min_lag("AutoArima")
 extract_data_sets()
 extract_data_sets("AutoArima")