Prepare options for a portalcasting model

Suite of functions used to generate model-writing options lists:

model_options is a basic template that provides validation of all of the inputs (specific models' functions can thus just wrap around this generalized function). Not to be confused with models_options, which creates the options list for the models subdirectory.

AutoArima_options creates a model_options list of control options for the model script controlling the AutoArima model.

ESSS_options creates a model_options list of control options for the model script controlling the ESSS model.

nbGARCH_options creates a model_options list of control options for the model script controlling the nbGARCH model.

pevGARCH_options creates a model_options list of control options for the model script controlling the pevGARCH model.

nbsGARCH_options creates a model_options list of control options for the model script controlling the nbsGARCH model.

model_options(tree = dirtree(), model = "AutoArima",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

AutoArima_options(tree = dirtree(), model = "AutoArima",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

ESSS_options(tree = dirtree(), model = "ESSS",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

nbGARCH_options(tree = dirtree(), model = "nbGARCH",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

pevGARCH_options(tree = dirtree(), model = "pevGARCH",
  mod_covariates = TRUE, lag = 6, quiet = FALSE)

nbsGARCH_options(tree = dirtree(), model = "nbsGARCH",
  mod_covariates = FALSE, lag = NULL, quiet = FALSE)

Arguments

tree
dirtree-class directory tree list. See dirtree.

model
character value of the name of the model.

mod_covariates
logical indicator for if the model requires covariates.

lag
integer (or integer numeric) lag time used for the covariates or NULL if covariates is FALSE.

quiet
logical indicator controlling if messages are printed.

Value

model_options: a class-model_options list of options used to set up a general model script.

AutoArima_options: a model_options list of settings controlling the AutoArima model creation.

ESSS_options: a model_options list of settings controlling the ESSS model creation.

nbGARCH_options: a model_options list of settings controlling the nbGARCH model creation.

pevGARCH_options: a model_options list of settings controlling the pevGARCH model creation.

nbsGARCH_options: a model_options list of settings controlling the nbsGARCH model creation.