Plot Predictions for a Given Point in Time Across Multiple Species

Plot the point value with confidence interval for a time point across multiple species. Casts can be selected either by supplying a cast_id number or any combination of dataset, model, and end_moon, which filter the available casts in unison. This plot type can only handle output from a single cast, so if multiple casts still remain, the one with the highest number is selected. To be more certain about cast selection, use the cast_id input.

plot_cast_point(
  main = ".",
  cast_id = NULL,
  cast_groups = NULL,
  dataset = NULL,
  model = NULL,
  end_moon = NULL,
  species = NULL,
  highlight_sp = NULL,
  moon = NULL,
  with_census = FALSE,
  settings = directory_settings(),
  quiet = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

cast_id

integer (or integer numeric) value representing the cast of interest, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information.

cast_groups

integer (or integer numeric) value of the cast group to combine with an ensemble. If NULL (default), the most recent cast group is ensembled.

dataset

character value of the rodent data set to include Default value is NULL, which equates to no selection with respect to dataset.

model

character value of the name of the model to include. Default value is NULL, which equates to no selection with respect to model. Also available is "Ensemble", which combines the models via ensemble_casts.

end_moon

integer (or integer numeric) newmoon number of the forecast origin. Default value is NULL, which equates to no selection with respect to end_moon.

species

character vector of the species codes (or "total" for the total across species) to be plotted or NULL (default) to plot all species in dataset.

highlight_sp

character vector of the species codes (or "total" for the total across species) to be highlighted or NULL (default) to not highlight anything.

moon

integer (or integer numeric) newmoon number for the plot.

with_census

logical toggle if the plot should include the observed data collected during the predicted census.

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

quiet

logical indicator if progress messages should be quieted.

Value

NULL. Plot is generated.

Details

The resulting plot shows predictions as points (open white circles) with error, where the point represents the estimate and the bounds of the error are lower_pi and upper_pi in the cast_table saved output from a model.
As of portalcasting v0.9.0, this represents the mean and the 95% prediction interval. If with_census = TRUE, the observations from the associated moon are plotted as blue filled squares.

Examples

 # \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  plot_cast_point()
# }