Functions for the prefab portalcasting models

Functions used for fitting the prefab models. For detailed descriptions of the models, see the model vignette.

AutoArima fits a flexible auto-regressive integrated moving average model using auto.arima.

NaiveArima fits a baseline random walk with no drift model as an ARIMA(0,1,0) using Arima.

ESSS fits a flexible exponential smoothing state space model with the possibility of multiplicative trends using ets.

nbGARCH fits a generalized autoregresive conditional heteroscedasticity model with overdispersion (i.e., a negative binomial response variable) using tsglm.

nbsGARCH fits a seasonal generalized autoregresive conditional heteroscedasticity model with overdispersion (i.e., a negative binomial response variable) using tsglm.

pevGARCH fits a set of generalized autoregressive conditional heteroscedasticity models with Poisson response variables and environmental covariates (max temp, mean temp, min temp, precipitation, and NDVI) using tsglm.

jags_RW fits a log-scale density random walk with a Poisson observation process using JAGS (Just Another Gibbs Sampler; Plummer 2003) hierarchical Bayesian inference.

jags_logistic fits a log-scale density logistic-growth (r-K) model with a Poisson observation process using JAGS (Just Another Gibbs Sampler; Plummer 2003) hierarchical Bayesian inference.

jags_logistic(
  main = ".",
  dataset = "dm_controls",
  settings = directory_settings(),
  control_runjags = runjags_control(),
  quiet = FALSE,
  verbose = FALSE
)

jags_RW(
  main = ".",
  dataset = "all",
  settings = directory_settings(),
  control_runjags = runjags_control(),
  quiet = FALSE,
  verbose = FALSE
)

AutoArima(
  main = ".",
  dataset = "all",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

NaiveArima(
  main = ".",
  dataset = "all",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

ESSS(
  main = ".",
  dataset = "all_interp",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

nbGARCH(
  main = ".",
  dataset = "all_interp",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

nbsGARCH(
  main = ".",
  dataset = "all_interp",
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

pevGARCH(
  main = ".",
  dataset = "all_interp",
  lag = 6,
  settings = directory_settings(),
  quiet = FALSE,
  verbose = FALSE
)

Arguments

main

character value of the name of the main component of the directory tree.

dataset

character value name of the rodent data set, such as ("all" or "controls").

settings

list of controls for the directory, with defaults set in directory_settings that should generally not need to be altered.

control_runjags

list of arguments passed to run.jags via runjags_control.

quiet

logical value indicating if the function should be quiet.

verbose

logical indicator of whether or not to print out all of the information or not (and thus just the tidy messages).

lag

integer (or integer numeric) of the lag time to use for the covariates.

Value

list of [1] model metadata list ("metadata"), cast summary data.frame ("cast_tab"), list of model fit objects ("model_fits"), and list of model cast objects ("model_casts").

Details

AutoArima
Because the seasonality and sampling occurring with different frequencies, which auto.arima cannot accommodate, seasonal AutoArima models are not included.

pevGARCH
Environmental predictors have an optional lag time, which is currently structured to impact all covariates the same and defaults to 6.
All covariate models are fit and the best model (by AIC) is selected for use.

References

Hyndman, R., Bergmeir, C., Caceres, G., Chhay, L., O'Hara-Wild, M., Petropoulos, F., Razbash, S., Wang, E., and Yasmeen, F. 2018. forecast: Forecasting functions for time series and linear models. R package version 8.3. URL.

Liboschik, T., Fokianos, K., and Fried, R. 2017. tscount: An R Package for Analysis of Count Time Series Following Generalized Linear Models. Journal of Statistical Software 82:5, 1-51. URL.

Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003). ISSN 1609-395X. URL.

Examples

 # \donttest{
  setup_dir()
  AutoArima()
  NaiveArima()
  ESSS()
  nbGARCH()
  nbsGARCH()
  pevGARCH()
  jags_RW()
  jags_logistic()
 # }