Prepare a model-running metadata list

Sets up the metadata used for casting, in particular the matching of time period across the data sets. This should always be run after prep_moons, prep_rodents, and prep_covariates before any model is run.

prep_metadata(
  main = ".",
  models = prefab_models(),
  data_sets = NULL,
  moons = NULL,
  rodents = NULL,
  covariates = NULL,
  end_moon = NULL,
  start_moon = 217,
  lead_time = 12,
  min_lag = NULL,
  cast_date = Sys.Date(),
  confidence_level = 0.95,
  controls_model = NULL,
  controls_rodents = NULL,
  control_files = files_control(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

models
character vector of name(s) of model(s) to include.

data_sets
character vector of the rodent data set names that the model is applied to.

moons
Moons data.frame. See prep_moons.

rodents
Rodents list. See prep_rodents,

covariates
Covariates data.frame. See prep_covariates.

end_moon
integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

start_moon
integer (or integer numeric) newmoon number of the first sample to be included. Default value is 217, corresponding to 1995-01-01.

lead_time
integer (or integer numeric) value for the number of timesteps forward a cast will cover.

min_lag
integer (or integer numeric) of the minimum covariate lag time used in any model.

cast_date
Date from which future is defined (the origin of the cast). In the recurring forecasting, is set to today's date using Sys.Date.

confidence_level
numeric confidence level used in summarizing model output. Must be between 0 and 1.

controls_model
Additional controls for models not in the prefab set.
A list of a single model's script-writing controls or a list of lists, each of which is a single model's script-writing controls.
Presently, each model's script writing controls should include three elements: name (a character value of the model name), covariates (a logical indicator of if the model needs covariates), and lag (an integer-conformable value of the lag to use with the covariates or NA if covariates = FALSE).
If only a single model is added, the name of the model from the element name will be used to name the model's list in the larger list. If multiple models are added, each element list must be named according to the model and the name element.

controls_rodents
Control list or list of control lists from rodents_controls specifying the structuring of the rodents tables. See rodents_controls for details.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

list of casting metadata, which is also saved out as a YAML file (.yaml) if desired.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  moons <- prep_moons()
  rodents <- prep_rodents(moons = moons)
  covariates <- prep_covariates(moons = moons)
  prep_metadata(moons = moons, rodents = rodents, covariates = covariates)

#> $cast_group
#> [1] 63
#> 
#> $models
#> [1] "AutoArima"  "ESSS"       "NaiveArima" "nbGARCH"    "nbsGARCH"  
#> [6] "pevGARCH"   "simplexEDM" "GPEDM"      "jags_RW"   
#> 
#> $data_sets
#> [1] "all"                "controls"           "exclosures"        
#> [4] "dm_controls"        "all_interp"         "controls_interp"   
#> [7] "exclosures_interp"  "dm_controls_interp"
#> 
#> $directory_configuration
#> $directory_configuration$setup_date
#> [1] "2021-02-26"
#> 
#> $directory_configuration$setup_R_version
#> $directory_configuration$setup_R_version$platform
#> [1] "x86_64-pc-linux-gnu"
#> 
#> $directory_configuration$setup_R_version$arch
#> [1] "x86_64"
#> 
#> $directory_configuration$setup_R_version$os
#> [1] "linux-gnu"
#> 
#> $directory_configuration$setup_R_version$system
#> [1] "x86_64, linux-gnu"
#> 
#> $directory_configuration$setup_R_version$status
#> [1] ""
#> 
#> $directory_configuration$setup_R_version$major
#> [1] "4"
#> 
#> $directory_configuration$setup_R_version$minor
#> [1] "0.3"
#> 
#> $directory_configuration$setup_R_version$year
#> [1] "2020"
#> 
#> $directory_configuration$setup_R_version$month
#> [1] "10"
#> 
#> $directory_configuration$setup_R_version$day
#> [1] "10"
#> 
#> $directory_configuration$setup_R_version$`svn rev`
#> [1] "79318"
#> 
#> $directory_configuration$setup_R_version$language
#> [1] "R"
#> 
#> $directory_configuration$setup_R_version$version.string
#> [1] "R version 4.0.3 (2020-10-10)"
#> 
#> $directory_configuration$setup_R_version$nickname
#> [1] "Bunny-Wunnies Freak Out"
#> 
#> 
#> $directory_configuration$setup_portalcasting_version
#> [1] "0.20.0"
#> 
#> $directory_configuration$directory_tree
#> $directory_configuration$directory_tree$main
#> [1] "."
#> 
#> $directory_configuration$directory_tree$subs
#> [1] "casts"  "models" "raw"    "data"   "tmp"   
#> 
#> 
#> $directory_configuration$downloads_versions
#> [1] "PortalData-2.80.0"            "portalPredictions-2021-02-25"
#> 
#> 
#> $controls_rodents
#> $controls_rodents$all
#> $controls_rodents$all$name
#> [1] "all"
#> 
#> $controls_rodents$all$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$all$total
#> [1] TRUE
#> 
#> $controls_rodents$all$interpolate
#> [1] FALSE
#> 
#> $controls_rodents$all$clean
#> [1] FALSE
#> 
#> $controls_rodents$all$type
#> [1] "Rodents"
#> 
#> $controls_rodents$all$level
#> [1] "Site"
#> 
#> $controls_rodents$all$plots
#> [1] "all"
#> 
#> $controls_rodents$all$treatment
#> NULL
#> 
#> $controls_rodents$all$min_plots
#> [1] 24
#> 
#> $controls_rodents$all$min_traps
#> [1] 1
#> 
#> $controls_rodents$all$output
#> [1] "abundance"
#> 
#> $controls_rodents$all$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$all$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$all$time
#> [1] "newmoon"
#> 
#> $controls_rodents$all$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$all$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$all$effort
#> [1] TRUE
#> 
#> $controls_rodents$all$filename
#> [1] "rodents_all.csv"
#> 
#> 
#> $controls_rodents$all_interp
#> $controls_rodents$all_interp$name
#> [1] "all_interp"
#> 
#> $controls_rodents$all_interp$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$all_interp$total
#> [1] TRUE
#> 
#> $controls_rodents$all_interp$interpolate
#> [1] TRUE
#> 
#> $controls_rodents$all_interp$clean
#> [1] FALSE
#> 
#> $controls_rodents$all_interp$type
#> [1] "Rodents"
#> 
#> $controls_rodents$all_interp$level
#> [1] "Site"
#> 
#> $controls_rodents$all_interp$plots
#> [1] "all"
#> 
#> $controls_rodents$all_interp$treatment
#> NULL
#> 
#> $controls_rodents$all_interp$min_plots
#> [1] 24
#> 
#> $controls_rodents$all_interp$min_traps
#> [1] 1
#> 
#> $controls_rodents$all_interp$output
#> [1] "abundance"
#> 
#> $controls_rodents$all_interp$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$all_interp$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$all_interp$time
#> [1] "newmoon"
#> 
#> $controls_rodents$all_interp$na_drop
#> [1] TRUE
#> 
#> $controls_rodents$all_interp$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$all_interp$effort
#> [1] TRUE
#> 
#> $controls_rodents$all_interp$filename
#> [1] "rodents_all_interp.csv"
#> 
#> 
#> $controls_rodents$controls
#> $controls_rodents$controls$name
#> [1] "controls"
#> 
#> $controls_rodents$controls$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$controls$total
#> [1] TRUE
#> 
#> $controls_rodents$controls$interpolate
#> [1] FALSE
#> 
#> $controls_rodents$controls$clean
#> [1] FALSE
#> 
#> $controls_rodents$controls$type
#> [1] "Rodents"
#> 
#> $controls_rodents$controls$level
#> [1] "Treatment"
#> 
#> $controls_rodents$controls$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$controls$treatment
#> [1] "control"
#> 
#> $controls_rodents$controls$min_plots
#> [1] 24
#> 
#> $controls_rodents$controls$min_traps
#> [1] 1
#> 
#> $controls_rodents$controls$output
#> [1] "abundance"
#> 
#> $controls_rodents$controls$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$controls$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$controls$time
#> [1] "newmoon"
#> 
#> $controls_rodents$controls$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$controls$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$controls$effort
#> [1] TRUE
#> 
#> $controls_rodents$controls$filename
#> [1] "rodents_controls.csv"
#> 
#> 
#> $controls_rodents$controls_interp
#> $controls_rodents$controls_interp$name
#> [1] "controls_interp"
#> 
#> $controls_rodents$controls_interp$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$controls_interp$total
#> [1] TRUE
#> 
#> $controls_rodents$controls_interp$interpolate
#> [1] TRUE
#> 
#> $controls_rodents$controls_interp$clean
#> [1] FALSE
#> 
#> $controls_rodents$controls_interp$type
#> [1] "Rodents"
#> 
#> $controls_rodents$controls_interp$level
#> [1] "Treatment"
#> 
#> $controls_rodents$controls_interp$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$controls_interp$treatment
#> [1] "control"
#> 
#> $controls_rodents$controls_interp$min_plots
#> [1] 24
#> 
#> $controls_rodents$controls_interp$min_traps
#> [1] 1
#> 
#> $controls_rodents$controls_interp$output
#> [1] "abundance"
#> 
#> $controls_rodents$controls_interp$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$controls_interp$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$controls_interp$time
#> [1] "newmoon"
#> 
#> $controls_rodents$controls_interp$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$controls_interp$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$controls_interp$effort
#> [1] TRUE
#> 
#> $controls_rodents$controls_interp$filename
#> [1] "rodents_controls_interp.csv"
#> 
#> 
#> $controls_rodents$exclosures
#> $controls_rodents$exclosures$name
#> [1] "exclosures"
#> 
#> $controls_rodents$exclosures$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$exclosures$total
#> [1] TRUE
#> 
#> $controls_rodents$exclosures$interpolate
#> [1] FALSE
#> 
#> $controls_rodents$exclosures$clean
#> [1] FALSE
#> 
#> $controls_rodents$exclosures$type
#> [1] "Rodents"
#> 
#> $controls_rodents$exclosures$level
#> [1] "Treatment"
#> 
#> $controls_rodents$exclosures$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$exclosures$treatment
#> [1] "exclosure"
#> 
#> $controls_rodents$exclosures$min_plots
#> [1] 24
#> 
#> $controls_rodents$exclosures$min_traps
#> [1] 1
#> 
#> $controls_rodents$exclosures$output
#> [1] "abundance"
#> 
#> $controls_rodents$exclosures$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$exclosures$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$exclosures$time
#> [1] "newmoon"
#> 
#> $controls_rodents$exclosures$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$exclosures$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$exclosures$effort
#> [1] TRUE
#> 
#> $controls_rodents$exclosures$filename
#> [1] "rodents_exclosures.csv"
#> 
#> 
#> $controls_rodents$exclosures_interp
#> $controls_rodents$exclosures_interp$name
#> [1] "exclosures_interp"
#> 
#> $controls_rodents$exclosures_interp$species
#>  [1] "BA" "DM" "DO" "DS" "NA" "OL" "OT" "PB" "PE" "PF" "PH" "PL" "PM" "PP" "RF"
#> [16] "RM" "RO" "SF" "SH" "SO"
#> 
#> $controls_rodents$exclosures_interp$total
#> [1] TRUE
#> 
#> $controls_rodents$exclosures_interp$interpolate
#> [1] TRUE
#> 
#> $controls_rodents$exclosures_interp$clean
#> [1] FALSE
#> 
#> $controls_rodents$exclosures_interp$type
#> [1] "Rodents"
#> 
#> $controls_rodents$exclosures_interp$level
#> [1] "Treatment"
#> 
#> $controls_rodents$exclosures_interp$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$exclosures_interp$treatment
#> [1] "exclosure"
#> 
#> $controls_rodents$exclosures_interp$min_plots
#> [1] 24
#> 
#> $controls_rodents$exclosures_interp$min_traps
#> [1] 1
#> 
#> $controls_rodents$exclosures_interp$output
#> [1] "abundance"
#> 
#> $controls_rodents$exclosures_interp$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$exclosures_interp$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$exclosures_interp$time
#> [1] "newmoon"
#> 
#> $controls_rodents$exclosures_interp$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$exclosures_interp$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$exclosures_interp$effort
#> [1] TRUE
#> 
#> $controls_rodents$exclosures_interp$filename
#> [1] "rodents_exclosures_interp.csv"
#> 
#> 
#> $controls_rodents$dm_controls
#> $controls_rodents$dm_controls$name
#> [1] "dm_controls"
#> 
#> $controls_rodents$dm_controls$species
#> [1] "DM"
#> 
#> $controls_rodents$dm_controls$total
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$interpolate
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$clean
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$type
#> [1] "Rodents"
#> 
#> $controls_rodents$dm_controls$level
#> [1] "Treatment"
#> 
#> $controls_rodents$dm_controls$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$dm_controls$treatment
#> [1] "control"
#> 
#> $controls_rodents$dm_controls$min_plots
#> [1] 24
#> 
#> $controls_rodents$dm_controls$min_traps
#> [1] 1
#> 
#> $controls_rodents$dm_controls$output
#> [1] "abundance"
#> 
#> $controls_rodents$dm_controls$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$time
#> [1] "newmoon"
#> 
#> $controls_rodents$dm_controls$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$dm_controls$effort
#> [1] TRUE
#> 
#> $controls_rodents$dm_controls$filename
#> [1] "rodents_dm_controls.csv"
#> 
#> 
#> $controls_rodents$dm_controls_interp
#> $controls_rodents$dm_controls_interp$name
#> [1] "dm_controls_interp"
#> 
#> $controls_rodents$dm_controls_interp$species
#> [1] "DM"
#> 
#> $controls_rodents$dm_controls_interp$total
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls_interp$interpolate
#> [1] TRUE
#> 
#> $controls_rodents$dm_controls_interp$clean
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls_interp$type
#> [1] "Rodents"
#> 
#> $controls_rodents$dm_controls_interp$level
#> [1] "Treatment"
#> 
#> $controls_rodents$dm_controls_interp$plots
#> [1] "Longterm"
#> 
#> $controls_rodents$dm_controls_interp$treatment
#> [1] "control"
#> 
#> $controls_rodents$dm_controls_interp$min_plots
#> [1] 24
#> 
#> $controls_rodents$dm_controls_interp$min_traps
#> [1] 1
#> 
#> $controls_rodents$dm_controls_interp$output
#> [1] "abundance"
#> 
#> $controls_rodents$dm_controls_interp$fillweight
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls_interp$unknowns
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls_interp$time
#> [1] "newmoon"
#> 
#> $controls_rodents$dm_controls_interp$na_drop
#> [1] FALSE
#> 
#> $controls_rodents$dm_controls_interp$zero_drop
#> [1] TRUE
#> 
#> $controls_rodents$dm_controls_interp$effort
#> [1] TRUE
#> 
#> $controls_rodents$dm_controls_interp$filename
#> [1] "rodents_dm_controls_interp.csv"
#> 
#> 
#> 
#> $cast_type
#> [1] "forecast"
#> 
#> $start_moon
#> [1] 217
#> 
#> $end_moon
#> [1] 540
#> 
#> $last_moon
#> [1] 540
#> 
#> $lead_time
#> [1] 12
#> 
#> $min_lag
#> [1] 6
#> 
#> $cast_date
#> [1] "2021-02-26"
#> 
#> $covariate_cast_moons
#> [1] 547 548 549 550 551 552
#> 
#> $covariate_cast_months
#> [1]  9 10 11 12  1  2
#> 
#> $covariate_cast_years
#> [1] 2021 2021 2021 2021 2022 2022
#> 
#> $rodent_cast_moons
#>  [1] 528 529 530 531 532 533 534 535 536 537 538 539 540 541 542 543 544 545 546
#> [20] 547 548 549 550 551 552
#> 
#> $rodent_cast_months
#>  [1]  2  3  4  5  6  7  8  9 10 11 12  1  2  3  4  5  6  7  8  9 10 11 12  1  2
#> 
#> $rodent_cast_years
#>  [1] 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2020 2021 2021 2021 2021
#> [16] 2021 2021 2021 2021 2021 2021 2021 2021 2022 2022
#> 
#> $confidence_level
#> [1] 0.95
#> 

 # }