Prepare historical covariates data

prep_hist_covariates creates a data.frame of historical weather and NDVI data. Automatically goes all the way back to the beginning of the available data, as the table is automatically trimmed later.

NDVI data are generated by ndvi.

prep_hist_covariates(
  main = ".",
  moons = NULL,
  control_files = files_control(),
  quiet = TRUE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

moons
Lunar data data.frame. See prep_moons.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

quiet
logical indicator if progress messages should be quieted.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

Value

prep_hist_covariates: daily historical covariate data table as a data.frame.

Examples

 # \donttest{   
  create_dir()

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

  fill_raw()

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

  prep_hist_covariates()

#>     moon mintemp maxtemp   meantemp precipitation      ndvi source
#> 1     32  -9.400   21.50  5.3823529     3.8100000        NA   hist
#> 2     33 -10.000   23.00  5.1666667    38.8620000        NA   hist
#> 3     34   1.000   24.50 10.6517241     7.8740000        NA   hist
#> 4     35  -2.000   29.00 11.6896552     0.0000000        NA   hist
#> 5     36   3.000   31.00 18.6700000     1.0160000        NA   hist
#> 6     37   2.000   39.00 25.6344828     8.3820000        NA   hist
#> 7     38  18.000   41.00 28.3551724    13.4620000        NA   hist
#> 8     39  18.000   39.00 25.7100000    75.4380000        NA   hist
#> 9     40  15.000   35.00 22.0200000    47.4980000        NA   hist
#> 10    41  13.000   36.00 18.7551724    39.6240000        NA   hist
#> 11    42  -2.000   29.00  9.8333333     1.7780000        NA   hist
#> 12    43  -7.800   28.00  4.4333333    14.4780000        NA   hist
#> 13    44  -4.000   24.00  3.6266667     2.2860000        NA   hist
#> 14    45  -7.200   19.50  3.9241379    30.2260000        NA   hist
#> 15    46  -1.700   26.00  8.7166667    35.8140000        NA   hist
#> 16    47  -1.700   26.00 10.9551724    16.5100000        NA   hist
#> 17    48   7.000   32.00 19.4310345     2.7940000        NA   hist
#> 18    49   1.700   32.50 22.8724138     3.0480000        NA   hist
#> 19    50   6.100   43.90 28.7758621    77.7240000 0.2770000   hist
#> 20    51  11.100   39.40 24.0793103    71.1200000 0.2980000   hist
#> 21    52   8.300   40.60 24.0066667    23.1140000 0.9130000   hist
#> 22    53   8.300   37.80 19.3655172    53.8480000 0.7660000   hist
#> 23    54  -3.900   34.40 14.6300000    29.9720000 0.6490000   hist
#> 24    55  -2.000   33.90  7.9433333     1.0160000 0.5110000   hist
#> 25    56  -4.500   22.00  4.9033333    33.5280000 0.5160000   hist
#> 26    57  -6.500   19.00  4.2655172    25.1460000 0.3950000   hist
#> 27    58  -6.000   24.00  6.9233333     6.8580000 0.4590000   hist
#> 28    59  -4.000   25.00 11.5900000    25.9080000 0.6490000   hist
#> 29    60  -1.700   27.50 17.0068966     4.5720000 0.8470000   hist
#> 30    61   5.000   32.00 20.7965517     9.6520000 0.7240000   hist
#> 31    62   8.500   35.50 27.4233333     1.0160000 0.6250000   hist
#> 32    63   8.900   37.50 27.1896552    25.4000000 0.5490000   hist
#> 33    64  15.000   35.00 23.9689655    41.1480000 0.6970000   hist
#> 34    65  10.000   36.00 22.3900000    96.5200000 0.6650000   hist
#> 35    66   2.000   34.00 17.4482759    13.7160000 0.6880000   hist
#> 36    67  -2.500   29.00 11.4733333    21.3360000 0.4650000   hist
#> 37    68  -4.000   19.00  4.3366667   128.7340000 0.3720000   hist
#> 38    69 -15.000   18.00  0.4241379    57.6640000 0.1760000   hist
#> 39    70  -7.000   17.50  5.1333333    30.2400000 0.3730000   hist
#> 40    71  -6.700   21.50 10.0896552    12.9540000 0.4170000   hist
#> 41    72  -4.000   24.00  9.0600000    30.2260000 0.5680000   hist
#> 42    73  -3.500   28.00 18.8827586     0.0000000 0.6600000   hist
#> 43    74   4.000   33.00 23.3310345     0.2540000 0.6130000   hist
#> 44    75  10.500   37.00 27.8133333    30.9880000 0.5440000   hist
#> 45    76  14.000   34.50 24.9965517    61.2140000 0.6380000   hist
#> 46    77  15.600   36.00 23.4586207     5.0800000 0.6210000   hist
#> 47    78  12.500   34.50 19.1689655   193.1300000 0.5620000   hist
#> 48    79   8.000   28.50 13.3586207    41.1480000 0.6620000   hist
#> 49    80  -5.600   23.00  6.7966667    49.2760000 0.6460000   hist
#> 50    81  -5.000   21.00  4.9620690     8.6360000 0.5100000   hist
#> 51    82  -4.500   18.00  4.4366667    21.1280000 0.5950000   hist
#> 52    83  -4.500   21.00  7.8833333     0.0000000 0.6460000   hist
#> 53    84  -6.000   24.00 13.6333333     0.0000000 0.8200000   hist
#> 54    85   1.000   28.00 17.3758621     3.5560000 0.5076992   hist
#> 55    86   1.100   35.50 25.8300000     2.5400000 0.9818707   hist
#> 56    87   8.900   34.00 24.4793103    21.3360000 0.9549770   hist
#> 57    88  11.700   35.00 23.7566667    39.6240000 0.6825208   hist
#> 58    89  11.100   32.50 22.6655172    78.7400000 0.6620000   hist
#> 59    90  10.600   32.00 21.2275862    21.0820000 0.6940000   hist
#> 60    91   0.000   26.50 13.4800000    56.1340000 0.6860000   hist
#> 61    92  -1.000   26.00  7.8896552     2.0320000 0.9466146   hist
#> 62    93  -8.300   18.00  3.4633333    72.3900000 0.7472264   hist
#> 63    94  -5.500   17.00  2.3068966    30.7340000 0.5906795   hist
#> 64    95  -6.000   23.00  4.1900000    28.9620000 0.4500000   hist
#> 65    96  -1.500   25.00 10.1433333    31.4960000 0.5250000   hist
#> 66    97  -1.000   28.50 16.4172414     1.2700000 0.8160252   hist
#> 67    98   2.000   30.00 18.9000000    26.1620000 0.7778059   hist
#> 68    99   8.500   37.00 26.0827586     0.0000000 0.5660000   hist
#> 69   100  14.000   36.00 27.8517241    46.2280000 0.5000000   hist
#> 70   101   8.300   32.50 23.4133333   163.0680000 0.8626387   hist
#> 71   102  12.500   33.00 22.8678571    20.8280000 0.9625480   hist
#> 72   103   6.000   30.00 15.9655172    99.5680000 1.2861275   hist
#> 73   104   1.100   26.50 10.9133333    63.2460000 1.1106031   hist
#> 74   105  -4.400   19.00  5.3689655    10.9220000 1.1461421   hist
#> 75   106 -11.700   22.80  1.9766667     9.1440000 0.6418051   hist
#> 76   107  -4.700   28.60  5.7344828    23.8760000 0.9297407   hist
#> 77   108  -3.300   25.00 10.9233333     1.2700000 0.6844325   hist
#> 78   109  -5.000   28.00 13.5689655    33.2880000 1.0149348   hist
#> 79   110   0.600   30.50 19.2766667     1.5240000 0.7800948   hist
#> 80   111   7.500   33.00 23.1333333    10.4140000 1.0345345   hist
#> 81   112  14.500   35.00 25.4379310    18.2880000 0.9982919   hist
#> 82   113  13.500   36.50 24.5733333    57.4040000 1.0478334   hist
#> 83   114  14.000   34.50 22.1333333    67.0560000 1.1884980   hist
#> 84   115   5.600   31.50 19.1482759    36.5760000 1.1399296   hist
#> 85   116  -0.600   28.50 11.9965517    62.4840000 0.9764966   hist
#> 86   117  -7.200   18.50  6.1500000     6.0960000 0.6411063   hist
#> 87   118  -3.000   18.00  2.3241379    11.1760000 0.4619318   hist
#> 88   119 -16.700   20.00  2.3633333     0.5080000 0.7600049   hist
#> 89   120  -5.000   21.10  5.8103448    25.6760000 0.5423683   hist
#> 90   121  -1.700   23.90  9.8566667     3.5560000 0.6651834   hist
#> 91   122   1.100   29.40 16.0172414    17.7800000 1.1087014   hist
#> 92   123   7.200   28.90 18.2366667     9.6520000 1.1447261   hist
#> 93   124   8.900   34.40 24.9379310     7.1120000 1.1770269   hist
#> 94   125  15.000   37.80 27.5766667    18.2880000 0.9652167   hist
#> 95   126  15.600   34.40 22.3300000   106.4260000 1.2852309   hist
#> 96   127   6.700   32.80 19.2482759    45.7200000 1.1177855   hist
#> 97   128   3.900   30.60 16.0433333    29.7180000 1.2243503   hist
#> 98   129   1.700   30.00  9.4655172     7.6200000 0.5690000   hist
#> 99   130  -2.800   25.00  3.9900000    38.3540000 0.4560000   hist
#> 100  131  -3.900   18.30  2.7448276    30.4800000 0.4280000   hist
#> 101  132 -14.400   19.40  4.4133333    24.6380000 0.5126441   hist
#> 102  133  -8.300   28.90  9.8275862    17.7800000 0.6974632   hist
#> 103  134   6.700   32.20 15.7241379    16.5100000 1.1376963   hist
#> 104  135   6.100   36.70 19.9034483     6.3500000 0.8756719   hist
#> 105  136   9.400   36.10 25.7300000     0.2540000 0.6360686   hist
#> 106  137  10.000   36.10 25.1586207    68.0720000 0.4890000   hist
#> 107  138  18.300   35.60 22.7200000   128.2700000 1.2015553   hist
#> 108  139  12.200   36.70 20.7034483    64.0080000 1.2573622   hist
#> 109  140  10.000   30.60 16.3533333    38.8620000 1.6814254   hist
#> 110  141   8.900   28.30 11.6533333    50.5460000 1.0098085   hist
#> 111  142   0.600   25.00  3.1068966     8.8900000 0.5410000   hist
#> 112  143  -5.000   20.00  1.7413793    42.1640000 0.5490000   hist
#> 113  144  -2.800   19.40  2.7033333     8.6360000 0.4430000   hist
#> 114  145   1.100   26.10  8.7551724     7.9240000 0.4090000   hist
#> 115  146  -4.400   30.00 16.2896552    14.2240000 0.5240000   hist
#> 116  147   6.700   33.90 22.1100000     0.0000000 0.9063593   hist
#> 117  148   7.800   35.60 24.3241379     2.7940000 1.1259902   hist
#> 118  149  14.140   41.65 28.3407416     0.0000000 0.9339190   hist
#> 119  150  16.680   41.19 26.6381806    93.2180000 1.2652717   hist
#> 120  151  14.110   35.27 23.9656034    24.1300000 1.2820681   hist
#> 121  152  10.140   34.58 22.9893056    15.4940000 1.0807375   hist
#> 122  153   3.814   33.00 17.9765861    28.7020000 1.4170832   hist
#> 123  154  -0.727   26.42 12.4115514     0.0000000 0.7630000   hist
#> 124  155  -8.380   22.53  5.9203003     0.0000000 0.6110000   hist
#> 125  156  -5.596   22.40  5.9698181    12.7000000 0.5090000   hist
#> 126  157  -8.650   24.87  7.3587389    16.7640000 0.4850000   hist
#> 127  158  -3.912   28.48 13.2235445     8.6360000 0.5480782   hist
#> 128  159   0.559   31.55 17.3918376     3.5560000 0.9862556   hist
#> 129  160   3.877   35.65 20.9731542     3.5560000 0.9720651   hist
#> 130  161   9.240   40.64 27.1081178     7.8740000 0.9390966   hist
#> 131  162  17.340   42.64 26.9062931    70.8660000 0.7466631   hist
#> 132  163  14.800   35.35 22.6538472    83.0580000 0.7190000   hist
#> 133  164  13.390   36.74 24.7725144     7.1120000 1.5577639   hist
#> 134  165   3.365   31.31 19.0411194    53.0860000 1.1029142   hist
#> 135  166  -0.158   28.92 14.0908083    30.4800000 0.9667798   hist
#> 136  167  -3.977   25.33  9.2617342    31.7500000 0.4950000   hist
#> 137  168 -10.160   19.47  5.6794014    26.9240000 0.4210000   hist
#> 138  169  -8.110   21.54  7.4220486    27.4320000 0.1740000   hist
#> 139  170  -2.467   25.59 10.7348181    22.0980000 0.7770000   hist
#> 140  171  -4.625   29.03 12.7237270    13.2080000 0.9289255   hist
#> 141  172   3.385   31.72 18.5229181     0.0000000 1.2289724   hist
#> 142  173   7.270   34.60 22.3005556    12.1920000 1.1315811   hist
#> 143  174  13.810   38.96 26.5267816    27.6860000 1.0610986   hist
#> 144  175  16.490   35.89 24.8137644    55.3720000 1.1253633   hist
#> 145  176  14.410   35.43 24.4948194    36.0680000 1.3542689   hist
#> 146  177  11.070   32.61 21.2087069    34.2900000 1.2254894   hist
#> 147  178  -6.100   32.23 14.2900696    18.7000000 1.1576739   hist
#> 148  179 -13.300   22.80  6.7655172    32.6000000 0.4820000   hist
#> 149  180  -2.151   19.91  7.0303792   104.3940000 0.3610000   hist
#> 150  181  -6.991   16.35  4.8948222    26.6700000 0.3980000   hist
#> 151  182  -1.127   22.14  9.4986875    20.5740000 0.6510000   hist
#> 152  183   0.001   24.37 11.8828276    34.5440000 0.9297542   hist
#> 153  184   3.955   33.42 19.1004875     3.8100000 0.9471633   hist
#> 154  185   8.520   31.13 18.8003736    66.2940000 0.5334728   hist
#> 155  186  10.580   38.72 25.7697361    10.1600000 1.5134892   hist
#> 156  187  15.320   40.11 26.2771983    52.0700000 1.2094399   hist
#> 157  188  14.220   37.32 24.6574425    54.8640000 1.0955714   hist
#> 158  189   3.900   33.69 24.2765093    42.9080000 1.2636798   hist
#> 159  190   8.100   32.57 20.2094828     5.0800000 1.4297017   hist
#> 160  191  -3.364   27.91 11.6844597    11.9380000 0.8767308   hist
#> 161  192  -5.831   19.71  4.9478491    45.4660000 0.6370650   hist
#> 162  193  -3.447   19.76  8.7309278    60.7060000 0.6499825   hist
#> 163  194  -1.565   20.56  8.0388222    17.5260000 0.9009513   hist
#> 164  195  -2.189   27.85 11.9288319     1.7780000 1.1255200   hist
#> 165  196  -0.220   31.12 15.1786466     5.3340000 1.5448334   hist
#> 166  197   7.890   33.08 21.6803056     0.7620000 1.0673743   hist
#> 167  198   8.650   39.71 25.0129598    10.4140000 1.1512953   hist
#> 168  199  16.750   40.23 27.2280278    16.5100000 1.0772255   hist
#> 169  200  16.990   39.11 26.8926724     0.0000000 1.0528663   hist
#> 170  201  12.680   36.10 23.2730603    47.4980000 0.9779916   hist
#> 171  202   9.960   35.02 22.0568056     2.0320000 1.3249546   hist
#> 172  203  -1.485   27.25 13.3276178     8.6360000 0.7089438   hist
#> 173  204  -5.917   23.79  9.1185431    26.9240000 0.9467371   hist
#> 174  205  -5.180   23.25  6.8842299    10.9220000 0.7350240   hist
#> 175  206  -7.030   21.86  7.3883583    17.0180000 0.6167529   hist
#> 176  207  -2.215   23.85 11.2886097     3.0480000 0.7535546   hist
#> 177  208  -1.891   27.59 14.9931451    12.1920000 0.9202397   hist
#> 178  209   1.902   32.10 19.1400375     3.5560000 0.9671976   hist
#> 179  210   8.480   37.78 24.5587639     3.0480000 0.9104722   hist
#> 180  211  17.910   43.87 30.0727874     3.5560000 0.8864295   hist
#> 181  212  17.120   39.55 28.4607361    18.7960000 0.8407906   hist
#> 182  213  17.150   38.04 26.1839799    47.2440000 0.8560231   hist
#> 183  214  13.140   34.95 23.9333764    11.6840000 1.0756582   hist
#> 184  215   2.634   30.24 17.3371333    13.2080000 0.9094507   hist
#> 185  216  -6.061   26.64  9.7336207    63.7540000 0.9621205   hist
#> 186  217  -2.096   20.19  7.5230986    77.4700000 0.7279804   hist
#> 187  218  -5.131   21.81  6.9285963    21.0820000 0.3820418   hist
#> 188  219  -0.725   22.90 11.3269278    21.3360000 1.9241013   hist
#> 189  220  -3.887   25.66 12.5276753    16.0020000 1.0493595   hist
#> 190  221  -1.755   29.78 15.3380125     1.5240000 1.6990468   hist
#> 191  222   2.985   33.37 20.4020097     2.5400000 1.4630554   hist
#> 192  223   9.280   37.92 25.8065374     0.0000000 1.1234442   hist
#> 193  224  14.400   42.32 27.4418750    39.1160000 1.3687018   hist
#> 194  225  17.610   41.11 27.9051437    27.9400000 1.0106297   hist
#> 195  226  10.760   38.18 24.3610278    60.9600000 1.3270818   hist
#> 196  227   5.144   31.47 20.4440517    12.7000000 1.5259721   hist
#> 197  228   4.798   29.68 15.0525778     7.6200000 1.6395770   hist
#> 198  229  -4.261   25.02 10.6706063     4.0640000 0.9047893   hist
#> 199  230  -6.390   40.99  7.8518722     0.5080000 0.5945451   hist
#> 200  231  -5.048   25.58 11.2732471     6.3500000 0.8281120   hist
#> 201  232  -3.105   25.80 12.1724278     1.7780000 0.7712942   hist
#> 202  233   0.228   31.15 15.7045503     4.3180000 1.0865517   hist
#> 203  234   5.274   38.88 23.8743042     0.0000000 1.1268188   hist
#> 204  235   8.750   38.70 26.3704454    26.1620000 1.0054495   hist
#> 205  236  15.190   40.10 26.8715000    67.3100000 0.9030040   hist
#> 206  237  17.270   38.31 26.3999444    52.3240000 0.9779368   hist
#> 207  238  14.570   38.83 24.1762931    21.8440000 1.2263030   hist
#> 208  239  11.120   34.06 21.7910694    34.7980000 1.4897214   hist
#> 209  240  -1.422   29.10 13.5512658    43.9420000 0.7341925   hist
#> 210  241  -3.633   26.11 11.1145181     6.8580000 0.8498784   hist
#> 211  242  -9.830   21.41  6.9165517    11.9380000 0.8974578   hist
#> 212  243  -4.381   20.91  7.5957625    13.9700000 0.8086074   hist
#> 213  244  -1.273   25.14  9.3463994    40.6400000 0.9721571   hist
#> 214  245  -2.423   28.97 14.8487153     3.0480000 1.1717235   hist
#> 215  246   2.139   35.05 18.9308391     2.7940000 1.0279353   hist
#> 216  247  11.780   38.15 23.5794306    32.0040000 0.9918165   hist
#> 217  248  12.030   41.09 27.4351006     2.2860000 1.1106296   hist
#> 218  249  16.220   38.77 26.4266667    40.6400000 0.7171574   hist
#> 219  250  16.710   36.90 25.5058190    34.2900000 1.0324127   hist
#> 220  251  13.600   36.64 24.6428611    27.4320000 1.3505526   hist
#> 221  252   0.197   32.99 16.5841653     8.8900000 0.6508895   hist
#> 222  253  -1.007   26.94 11.9689109    16.0020000 0.6534885   hist
#> 223  254  -6.621   18.60  4.6297569    51.8160000 0.8243953   hist
#> 224  255  -3.555   21.20  7.9044799     1.2700000 0.7457902   hist
#> 225  256  -2.394   21.80  7.2969681    30.4800000 0.7837895   hist
#> 226  257  -4.167   29.09 11.4488506    17.2720000 0.8494762   hist
#> 227  258  -1.842   30.90 13.0964875    12.7000000 1.1317241   hist
#> 228  259  -2.200   30.43 22.3670795     0.0000000 1.0760310   hist
#> 229  260   2.800   33.30 27.5793103     4.3000000 1.1307745   hist
#> 230  261   8.300   36.10 26.9896552   122.2000000 0.9971322   hist
#> 231  262   9.400   31.10 26.2137931    49.5000000 1.1080989   hist
#> 232  263  12.200   36.51 26.1762233    28.9560000 1.0628434   hist
#> 233  264   3.873   35.08 22.0880347     0.0000000 0.9106251   hist
#> 234  265  -0.825   26.78 14.0441394    32.7660000 1.0103421   hist
#> 235  266  -5.078   24.89  8.9215583    20.5740000 1.0598884   hist
#> 236  267  -4.506   22.25  9.0706208     0.0000000 1.2194690   hist
#> 237  268  -6.447   26.77  9.7372313     0.0000000 0.8550300   hist
#> 238  269  -0.471   27.08 13.3529194     1.7780000 0.6867843   hist
#> 239  270  -5.057   28.61 13.9274468     3.5560000 0.5861795   hist
#> 240  271   2.273   32.91 20.1734500     0.0000000 0.7211451   hist
#> 241  272   6.324   34.75 24.2027644     0.0000000 0.5516956   hist
#> 242  273  15.940   40.00 26.6196695    47.2440000 0.9927058   hist
#> 243  274  16.180   34.43 23.0574028   118.1100000 1.5354452   hist
#> 244  275  16.350   35.88 25.0256322    49.7840000 2.0580431   hist
#> 245  276   8.550   34.02 22.4519167    22.0980000 2.6868937   hist
#> 246  277  -0.111   31.26 16.9727414     0.2540000 1.9684873   hist
#> 247  278  -5.276   28.11 12.6239667     0.0000000 1.4151075   hist
#> 248  279  -6.688   19.00  5.8826722     4.8260000 0.8998612   hist
#> 249  280  -4.997   25.79 10.3619375     0.0000000 1.2344397   hist
#> 250  281  -1.998   25.92 12.6094353     1.0160000 1.3952603   hist
#> 251  282  -2.110   28.42 13.3195236     4.8260000 1.0774605   hist
#> 252  283   3.298   35.33 21.0929612     0.0000000 1.2327855   hist
#> 253  284   9.580   39.88 25.9895139     0.0000000 1.2541998   hist
#> 254  285  15.350   38.92 25.7440805    41.4020000 1.3082427   hist
#> 255  286  17.440   39.23 27.5517529    21.3360000 1.5908658   hist
#> 256  287  16.750   38.93 26.2400000    34.5440000 1.8519476   hist
#> 257  288  12.330   38.24 25.0756034     8.6360000 1.4115558   hist
#> 258  289   3.802   35.55 17.8582347   126.7460000 1.3739654   hist
#> 259  290  -2.392   20.40  8.0682170    30.2260000 1.7334135   hist
#> 260  291  -4.821   21.41  8.6783889     6.8580000 1.3143874   hist
#> 261  292  -5.807   19.86  5.2978458    21.8440000 1.6104376   hist
#> 262  293  -5.015   23.59  8.0268681    10.9220000 3.2077380   hist
#> 263  294  -1.712   28.18 10.7736063     3.0480000 2.3762505   hist
#> 264  295  -1.641   31.38 16.4949167    13.2080000 2.2623165   hist
#> 265  296   7.350   34.44 22.0801724     9.6520000 2.2930869   hist
#> 266  297  12.930   38.91 27.7643333     0.0000000 2.0160450   hist
#> 267  298  16.320   39.28 25.9574856    60.4520000 1.8318931   hist
#> 268  299  16.550   36.55 25.1411782    56.3880000 1.8349707   hist
#> 269  300  14.840   37.84 25.7126944    16.0020000 2.6296862   hist
#> 270  301   4.879   35.15 20.9509080     6.3500000 1.7784186   hist
#> 271  302   4.713   32.05 18.9632601     4.5720000 1.3612857   hist
#> 272  303  -6.545   23.93  8.1328431     2.7940000 1.6945172   hist
#> 273  304  -4.361   22.37  7.2340514     3.3020000 1.4354970   hist
#> 274  305  -8.900   21.98  7.5574444    23.8760000 1.6683087   hist
#> 275  306 -13.300   23.90 15.1931034     4.9000000 1.5440824   hist
#> 276  307  -6.700   26.70 19.0700000     0.0000000 1.1995178   hist
#> 277  308  -4.400   28.90 23.7379310     0.0000000 1.2539433   hist
#> 278  309   1.700   35.60 27.6689655     0.0000000 1.4768373   hist
#> 279  310   9.400   36.70 31.8833333     2.3000000 1.2055539   hist
#> 280  311  10.600   34.40 24.8068966    86.4000000 1.2055558   hist
#> 281  312   8.300   32.80 27.1827586    81.1000000 3.5665632   hist
#> 282  313  -2.200   28.90 22.4344828    24.0000000 2.9802597   hist
#> 283  314  -3.300   26.10 15.8793103    11.0000000 1.6537128   hist
#> 284  315  -6.700   22.80 10.2533333    23.8000000 1.7565814   hist
#> 285  316 -11.100   18.90  4.2182221    23.3000000 1.0365508   hist
#> 286  317  -1.920   26.14 10.6113194     0.9000000 1.1334405   hist
#> 287  318  -4.710   22.88  9.0420402    17.1000000 1.3291683   hist
#> 288  319  -1.920   28.68 13.2760694     3.6000000 1.2432005   hist
#> 289  320  -0.200   29.76 17.0315694     0.0000000 1.1313898   hist
#> 290  321   6.650   37.36 23.5898851     0.0000000 0.9247814   hist
#> 291  322  15.020   37.92 26.9256667     7.5000000 1.1549359   hist
#> 292  323  16.980   41.44 29.4389224    20.8000000 1.2517108   hist
#> 293  324  17.380   39.65 27.0685139    23.8000000 1.1293278   hist
#> 294  325  13.840   34.57 25.3903448     6.1000000 1.0993284   hist
#> 295  326  11.830   35.50 22.5589722     9.5000000 1.0212618   hist
#> 296  327  -2.120   28.66 13.5612644    19.2000000 1.0125218   hist
#> 297  328  -5.060   23.47  7.9980139     4.1000000 1.2053718   hist
#> 298  329  -9.710   18.52  6.9919253    37.8000000 1.4658889   hist
#> 299  330  -5.924   25.30  7.2006125     8.7000000 1.5585069   hist
#> 300  331  -0.220   29.13 11.5910129    43.5000000 1.3378172   hist
#> 301  332   1.931   28.41 15.7327944    23.9000000 1.6589129   hist
#> 302  333   6.568   34.47 21.3195503     0.9000000 1.8536561   hist
#> 303  334  11.590   38.42 25.9842500     1.4000000 1.7613298   hist
#> 304  335  14.370   37.12 26.7649444    20.3000000 1.8406462   hist
#> 305  336  16.520   38.16 25.3841667    91.0000000 1.4423643   hist
#> 306  337  12.490   35.46 23.4804444    60.5000000 2.2987828   hist
#> 307  338   9.210   35.36 20.1534195    25.2000000 1.9730919   hist
#> 308  339   0.845   29.72 14.1841500    16.1000000 1.7147049   hist
#> 309  340  -7.080   19.95  7.3102701    37.8000000 1.5927022   hist
#> 310  341  -7.550   21.81  7.0875111    34.0000000 1.4002997   hist
#> 311  342  -3.633   20.83  7.5798721    39.5000000 1.8526678   hist
#> 312  343   1.203   24.52 10.3916944    29.1000000 2.8374067   hist
#> 313  344  -3.201   28.05 13.2913060     0.5000000 2.4414969   hist
#> 314  345   3.199   29.53 18.0230181     4.8000000 2.1672405   hist
#> 315  346   7.230   39.50 24.4200718     3.1000000 1.9932998   hist
#> 316  347  13.210   39.91 28.7982083     0.0000000 1.6909883   hist
#> 317  348  17.380   40.61 27.9472557    31.6000000 1.9365892   hist
#> 318  349  15.360   37.52 24.6816389    45.9000000 1.9530155   hist
#> 319  350  11.740   35.74 24.3541250    31.3000000 1.9437895   hist
#> 320  351   4.987   33.82 17.3408362    22.1000000 1.7502957   hist
#> 321  352  -8.450   28.91 13.0243361     0.0000000 1.5955765   hist
#> 322  353  -5.618   24.89  8.7721667     3.9000000 0.6300134   hist
#> 323  354  -5.786   22.63  8.6606694     2.9000000 1.0582922   hist
#> 324  355  -3.835   24.57 11.0308204     2.8000000 1.0825948   hist
#> 325  356  -4.632   26.09 12.8914250     6.3000000 1.1444709   hist
#> 326  357   1.662   31.81 19.1419626     1.1000000 0.8329579   hist
#> 327  358   5.212   35.63 23.6112716     0.0000000 1.3241023   hist
#> 328  359  13.350   39.11 27.9417500    12.2000000 1.2051095   hist
#> 329  360  16.620   39.25 27.6588649    59.7000000 1.5361902   hist
#> 330  361  16.090   36.98 22.8279722   152.0000000 2.3209209   hist
#> 331  362   9.350   32.91 20.6254444    86.9000000 3.9126304   hist
#> 332  363   2.242   32.42 18.4836925    58.4000000 1.9128724   hist
#> 333  364   1.376   28.92 13.9423181     4.5000000 1.9415065   hist
#> 334  365  -9.840   24.69  7.9807931     2.2000000 1.8788084   hist
#> 335  366 -10.030   20.52  4.9323966     9.8000000 1.1700893   hist
#> 336  367  -8.380   23.53  6.8046861    30.0000000 1.3438488   hist
#> 337  368  -7.880   30.07 12.4973405     0.7000000 1.4776131   hist
#> 338  369  -0.216   30.10 15.6118194    21.4000000 1.4924329   hist
#> 339  370   2.299   33.80 20.0746523    25.1000000 1.8977147   hist
#> 340  371  10.740   35.90 24.2580316     9.5000000 1.6136647   hist
#> 341  372  17.690   41.27 28.8890972    12.2000000 1.9626687   hist
#> 342  373  15.800   39.10 24.5175000    34.0000000 1.9269279   hist
#> 343  374  14.490   35.82 24.8217222     5.6000000 3.1885733   hist
#> 344  375   6.078   34.64 22.3961925     1.8000000 2.5451683   hist
#> 345  376   1.177   32.58 17.5941667     0.0000000 1.6278552   hist
#> 346  377  -2.574   27.34 12.0387139    26.0000000 1.4808712   hist
#> 347  378  -7.440   20.59  5.0905708    13.8000000 0.9881464   hist
#> 348  379  -6.717   19.30  6.3367313     8.5000000 1.0482803   hist
#> 349  380  -3.933   26.94 10.9124542     3.4000000 1.6662994   hist
#> 350  381  -3.400   28.15 14.8502543     0.0000000 1.6682746   hist
#> 351  382   2.075   31.35 18.2427597     0.0000000 1.6788960   hist
#> 352  383   4.455   37.55 22.1220014     4.7000000 1.6964188   hist
#> 353  384  12.130   39.23 28.3487069     0.3000000 1.5471369   hist
#> 354  385  17.580   36.49 24.7530139   104.4000000 2.0509262   hist
#> 355  386  15.680   35.48 23.9715086    67.5000000 2.4524343   hist
#> 356  387  10.450   34.24 21.9150833    14.4000000 3.1764466   hist
#> 357  388   0.645   32.00 18.4878578     2.0000000 1.8503730   hist
#> 358  389  -0.845   29.84 13.1177139    11.0000000 1.7473721   hist
#> 359  390  -8.870   23.05  8.7082792    11.3000000 1.4169052   hist
#> 360  391  -6.905   22.22  8.7900056    11.6000000 1.2529679   hist
#> 361  392  -5.319   28.85  9.6923606     3.8000000 0.9304622   hist
#> 362  393   0.512   28.25 14.8713056    14.0000000 1.1895369   hist
#> 363  394  -0.845   31.35 15.8422974     0.4000000 1.0214003   hist
#> 364  395   6.867   36.46 23.1141486    16.8000000 1.0504478   hist
#> 365  396  12.630   37.42 24.5223563     1.2000000 1.0315265   hist
#> 366  397  17.700   40.04 27.2678592    36.1000000 1.5052973   hist
#> 367  398  17.950   39.51 27.9586250    17.4000000 1.5391328   hist
#> 368  399 -24.000   40.00 25.5927853   132.6300000 1.5266904   hist
#> 369  400   4.000   40.00 22.3783404    92.5600000 1.3500921   hist
#> 370  401  -1.814   30.21 15.0882639    12.8000000 1.1838994   hist
#> 371  402  -4.214   22.22  8.0833569    24.0000000 1.2925411   hist
#> 372  403  -5.341   18.59  5.7415611    14.6000000 1.0254293   hist
#> 373  404  -2.956   18.07  6.8230388    80.9000000 1.0366946   hist
#> 374  405  -1.458   22.34  9.1700569    11.9000000 1.3475194   hist
#> 375  406   0.406   28.32 14.3631083     2.2000000 1.8219926   hist
#> 376  407   1.497   33.22 17.7017529    10.6000000 2.3649539   hist
#> 377  408   6.273   40.56 25.0151361     0.0000000 2.0791778   hist
#> 378  409  13.280   38.32 27.3915805    10.2000000 1.6707861   hist
#> 379  410  17.750   40.33 25.8390661    27.3000000 1.4618431   hist
#> 380  411  15.900   39.02 25.5218472    39.1000000 1.8702144   hist
#> 381  412  12.530   36.61 23.9773994    11.4000000 2.1488145   hist
#> 382  413   5.675   30.30 17.3611394    14.8000000 1.6914544   hist
#> 383  414  -7.690   26.83 10.1351014     1.1000000 1.7557419   hist
#> 384  415  -9.920   24.02  8.3952194     7.5000000 1.0401264   hist
#> 385  416 -13.380   23.07  6.9889569     2.6000000 1.3553371   hist
#> 386  417 -16.590   25.31  8.9440903     1.5000000 1.2525393   hist
#> 387  418  -0.459   32.82 16.8791125     0.4000000 1.3651641   hist
#> 388  419   2.061   32.26 18.9170388     0.3000000 1.6264248   hist
#> 389  420   4.842   36.21 22.0674319     0.0000000 1.4935002   hist
#> 390  421  13.740   41.33 28.8372778     0.6000000 1.3027435   hist
#> 391  422  18.140   38.34 26.7557759    86.3000000 1.2557284   hist
#> 392  423  16.700   40.08 26.4203305    29.8000000 2.9245141   hist
#> 393  424  13.120   38.72 24.6886250    41.3000000 2.5245915   hist
#> 394  425   1.804   34.51 19.7785517     4.9000000 2.3296287   hist
#> 395  426  -2.745   29.26 11.6880690     7.7000000 1.4480468   hist
#> 396  427  -6.834   21.10  5.3765722    69.5000000 0.9077362   hist
#> 397  428  -5.462   22.77  8.0116764     2.9000000 0.8170257   hist
#> 398  429  -4.042   21.72  8.8318621    14.3000000 1.0662336   hist
#> 399  430  -6.200   26.61 11.2393806     5.7000000 1.3607495   hist
#> 400  431   1.497   32.71 17.6262833     0.7000000 0.9659065   hist
#> 401  432   8.250   34.90 22.6111063     4.3000000 1.0964495   hist
#> 402  433   7.950   38.35 27.2261528     0.3000000 1.1703588   hist
#> 403  434  17.290   40.23 27.9275431    26.6000000 0.9791053   hist
#> 404  435  16.930   38.95 26.7946667    15.3000000 1.0686758   hist
#> 405  436   8.790   37.45 23.7013075    14.7000000 1.2732084   hist
#> 406  437   5.442   34.70 22.2060819     4.3000000 0.8201655   hist
#> 407  438  -2.146   31.09 17.0818362     0.2000000 0.5120926   hist
#> 408  439  -3.177   25.51 12.6074819     2.9000000 0.9483855   hist
#> 409  440  -7.430   16.71  4.7199368    12.5000000 0.6376255   hist
#> 410  441 -11.140   22.11  7.3042903     9.0000000 1.0614008   hist
#> 411  442  -8.260   27.70  8.2927888    11.4000000 0.8757837   hist
#> 412  443   2.659   30.36 17.0721583     1.4000000 1.2426597   hist
#> 413  444   2.829   31.81 19.3743836     0.0000000 1.3167757   hist
#> 414  445   9.150   38.12 25.0641667     0.8000000 1.5524968   hist
#> 415  446  16.830   40.00 29.0011111    11.8000000 1.7154050   hist
#> 416  447  16.260   35.96 24.5918678    65.7000000 1.8185877   hist
#> 417  448  15.800   37.12 25.4782639    46.8000000 2.2171052   hist
#> 418  449  10.080   32.66 21.7059339    65.3000000 3.1666663   hist
#> 419  450   2.130   30.45 16.7892125     0.8000000 2.2136023   hist
#> 420  451  -1.250   26.12 11.2001897    21.2000000 1.6691923   hist
#> 421  452  -4.763   22.54  6.5759042    15.0000000 1.4681438   hist
#> 422  453  -4.436   24.46  8.8906810     0.0000000 0.8686261   hist
#> 423  454  -3.343   29.11 12.8990639     1.4000000 0.8829976   hist
#> 424  455   1.631   27.56 14.3885374     9.3000000 1.0795753   hist
#> 425  456   2.981   31.54 18.0964368     5.8000000 1.1514726   hist
#> 426  457   6.041   36.44 21.2732431     0.1000000 1.0050695   hist
#> 427  458  16.930   40.03 29.1495556     6.6000000 1.2678138   hist
#> 428  459  17.420   39.57 26.7969397    56.3000000 1.2708371   hist
#> 429  460  15.390   36.18 23.4991944    95.5218133 2.0352203   hist
#> 430  461  14.220   35.71 22.9463793   111.1131467 3.2064012   hist
#> 431  462   7.940   31.14 19.3932500    36.1154133 2.1366627   hist
#> 432  463  -3.404   31.24 13.6591417     7.8638400 2.1470154   hist
#> 433  464  -2.802   24.62  9.7918190    13.3976533 1.1679867   hist
#> 434  465  -7.720   24.17  5.5780472    12.5238933 0.8808318   hist
#> 435  466  -7.120   25.58 10.7776638    33.7853867 0.4385335   hist
#> 436  467  -3.403   25.65 12.3228097    23.5915200 1.0464857   hist
#> 437  468   0.446   29.58 17.4044612     0.0000000 1.4931015   hist
#> 438  469   3.827   31.96 17.9796638     5.8250667 1.1505500   hist
#> 439  470   6.632   36.64 24.4465306     4.2231733 1.1049765   hist
#> 440  471  17.060   41.13 27.1261638    55.3381333 1.1216413   hist
#> 441  472  16.290   38.25 25.2416944    75.5802400 2.3991911   hist
#> 442  473  15.720   39.43 24.5236351    46.8917867 1.4262694   hist
#> 443  474   9.160   35.83 22.0989306    30.1447200 1.6802046   hist
#> 444  475   1.473   33.20 15.8614164    18.3489600 1.3417106   hist
#> 445  476  -7.000   25.00  9.4851509    28.5382133 0.9396909   hist
#> 446  477  -5.669   18.06  5.0087430   190.5746667 0.6763127   hist
#> 447  478  -7.420   22.70  6.7150089     6.8444533 1.3432473   hist
#> 448  479   0.047   28.81 14.8750747     0.0000000 1.2206617   hist
#> 449  480   0.205   28.95 15.5014778     0.0000000 1.2690105   hist
#> 450  481   3.885   32.49 18.2091968     3.6406667 1.2215228   hist
#> 451  482   7.990   38.41 22.7370833     2.1844000 1.1234165   hist
#> 452  483  15.970   41.66 27.3998611    35.0960267 1.4531980   hist
#> 453  484  17.530   38.75 27.1393103   104.9968267 1.8169478   hist
#> 454  485  13.860   36.02 24.1674583    39.6104533 2.1776035   hist
#> 455  486  10.250   36.26 22.0348994    68.4445333 1.8857273   hist
#> 456  487   6.873   31.90 20.5846986     4.9513067 2.1876310   hist
#> 457  488  -0.424   27.98 13.2572778     5.6726667 1.8460336   hist
#> 458  489  -3.630   24.43  8.3240359    26.9240000 0.9555673   hist
#> 459  490  -4.598   23.75  7.5574889    44.1960000 0.9911794   hist
#> 460  491  -5.151   28.17 10.5668403    19.0500000 1.2821842   hist
#> 461  492  -0.283   30.23 16.7191451     0.7620000 1.1111326   hist
#> 462  493   2.699   32.20 18.9714528     0.5080000 1.4708110   hist
#> 463  494   4.446   36.12 21.2836379     0.1456267 1.3447614   hist
#> 464  495  13.000   41.88 28.0530972     0.0000000 1.1599795   hist
#> 465  496  17.210   39.83 27.2870259    53.9360533 1.0241655   hist
#> 466  497  15.580   35.64 24.6419828    75.1820000 2.3728685   hist
#> 467  498  14.980   35.90 24.8877500    12.4460000 2.2244980   hist
#> 468  499   8.170   33.86 21.6153592     4.5720000 1.2024657   hist
#> 469  500   3.432   31.56 17.2683222     3.3020000 1.0805843   hist
#> 470  501  -7.877   28.91 10.6596722    15.4940000 0.7751793   hist
#> 471  502  -5.219   23.47  9.3776681     1.0160000 0.7804968   hist
#> 472  503  -5.909   24.54 10.5719813     8.1280000 0.6860383   hist
#> 473  504  -3.488   24.97 11.1408994    11.9380000 0.5236605   hist
#> 474  505  -0.625   33.09 17.5854833     1.0160000 0.7230805   hist
#> 475  506   4.912   36.26 22.2140345     0.5080000 0.6981137   hist
#> 476  507  10.550   39.74 26.6023194     1.7780000 0.9496771   hist
#> 477  508  13.380   41.00 27.7197845    38.1000000 0.5695964   hist
#> 478  509  15.870   41.51 27.9241111    14.9860000 0.9606167   hist
#> 479  510  14.950   36.42 24.9176437    41.9080000 1.1332452   hist
#> 480  511   5.921   36.51 23.7177347    29.7140000 1.2202223   hist
#> 481  512   3.865   27.56 14.6400675    53.5940000 0.7401333   hist
#> 482  513  -6.136   24.84  9.3208764    41.1480000 1.2556021   hist
#> 483  514 -10.780   19.60  5.9930819    22.3520000 0.2816760   hist
#> 484  515  -5.178   22.74  8.4884736     5.8420000 1.2514946   hist
#> 485  516  -7.962   27.64  8.9374792     9.1440000 1.3931755   hist
#> 486  517  -2.544   28.51 13.1076839    21.0820000 1.8742687   hist
#> 487  518   1.151   32.48 18.8530694     6.0960000 1.2763324   hist
#> 488  519   4.380   33.71 20.4781307     6.6040000 0.9540730   hist
#> 489  520  11.580   38.23 27.3467778     8.6360000 1.2581853   hist
#> 490  521  16.440   39.83 28.2153448    38.6080000 1.0408416   hist
#> 491  522  17.480   39.64 27.6673611    34.0320000 1.3363447   hist
#> 492  523  12.850   37.06 23.9887931    59.1820000 1.0996533   hist
#> 493  524   1.083   32.00 18.6318014     3.3020000 1.5893304   hist
#> 494  525  -3.542   26.86 12.2994009    18.0340000 0.4788715   hist
#> 495  526  -4.544   22.58  7.4431681    90.6780000 0.4013013   hist
#> 496  527  -6.362   20.98  6.5901580    14.4780000 1.2002468   hist
#> 497  528  -7.191   22.24  8.9023083    13.4620000 0.8862174   hist
#> 498  529  -3.629   23.64 11.5896451    50.2920000 0.7790793   hist
#> 499  530  -1.058   28.53 15.5750708     0.2540000 1.2199426   hist
#> 500  531   8.880   37.01 23.9787500     0.0000000 0.6955318   hist
#> 501  532  11.000   40.53 27.0118889     0.2540000 0.6853989   hist
#> 502  533  17.020   41.95 29.2115694    11.9380000 0.4603117   hist
#> 503  534  17.170   41.29 28.3565661    43.1760000 0.7596091   hist
#> 504  535  10.390   39.63 25.9981715    12.1920000 0.9063107   hist
#> 505  536   9.000   36.43 23.4306210     0.0000000 0.8589713   hist
#> 506  537  -3.000   33.64 15.7989107    88.7200000 0.5978336   hist
#> 507  538  -6.000   30.00 10.0778580   193.6800000 0.5199673   hist
#> 508  539  -8.000   21.50  5.8770017     3.2500000 0.3352475   hist
#> 509  540  -5.230   23.50  7.4701167   142.9890000 0.1518755   hist

 # }