Prepare a rodents data tables for forecasting

This function wraps around summarize_rodent_data to produce a data.frames associated with a set of data specifications.

prep_rodents_table(
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = TRUE,
  clean = FALSE,
  type = "Rodents",
  level = "Site",
  plots = "all",
  treatment = NULL,
  min_plots = 24,
  min_traps = 1,
  output = "abundance",
  shape = "crosstab",
  fillweight = (output != "abundance"),
  unknowns = FALSE,
  time = "newmoon",
  na_drop = FALSE,
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site = TRUE),
  ref_species = all_species(),
  effort = FALSE,
  quiet = TRUE,
  verbose = FALSE,
  control_files = files_control(),
  save = TRUE,
  overwrite = TRUE,
  filename = "rodents_all.csv",
  arg_checks = TRUE
)

process_rodent_data(
  rodents_tab,
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = FALSE,
  clean = FALSE,
  level = "Site",
  treatment = NULL,
  na_drop = FALSE,
  time = "newmoon",
  ref_species = all_species(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

moons
Moons data.frame. See prep_moons.

end_moon
integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

species
character-valued vector of species names to include.

total
logical value indicating if a total (sum across species) should be added or not. Only available if more than one species is included.

interpolate
logical value indicating if the data should be interpolated to fill in NA values (using na.interp).

clean
logical indicator of if only the rodent data that passed QA/QC (clean = TRUE) or if all data (clean = FALSE) should be loaded.

type
character value of the rodent data set type, according to pre-existing definitions. An alternative toggle to species.
Either all species (type = "Rodents") or only granivoes (type = "Granivores").

level
character indicating the type of summary: "Plot", "Treatment", or "Site". Pipes directly to summarize_rodent_data.

plots
Specification of subset of plots. Can be a vector of numeric plots indicators or specific sets indicated by character values: "all" plots or "Longterm" plots (plots that have had the same treatment for the entire time series).

treatment
character indicating the specific treatment(s) to trim to if level = "Treatment": "control", "exclosure", "removal", or "spectabs"

min_plots
integer (or integer numeric) of the minimum number of plots surveyed for a survey to be used. Pipes directly to summarize_rodent_data.

min_traps
integer (or integer numeric) of the minimum number of traps collected for a plot to be used. Pipes directly to summarize_rodent_data.

output
character indicating the type of data: "abundance", "biomass", or "energy". Pipes directly to summarize_rodent_data.

shape
character value indicating a "crosstab" or "flat" output.

fillweight
logical specifier of whether to fill in unknown weights with other records from that individual or species, where possible.

unknowns
logical indicator to either remove all individuals not identified to species (unknowns = FALSE) or sum them in an additional column (unknowns = TRUE.

time
character value specifying the format of the time index in the output. Options are "period" (sequential Portal surveys), "newmoon" (lunar cycle numbering), and "date" (calendar date).
The default time = "newmoon" produces an equispaced observation timestep, a common format format for discrete-time modeling.

na_drop
logical indicator of if NA values (representing insufficient sampling) should be dropped.

zero_drop
logical indicator of if 0 values (representing sufficient sampling but no detection) should be dropped.

ref_species
character vector of the reference species. Sets a boundary outside of the species included in the model.

effort
logical indicator of if the effort columns should be included in the output.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control. Only used here to point to the moons data.

save
logical indicator controlling if the output should be saved out.

overwrite
logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

filename
character name of the file for saving the output. The extension of the filename may be used for identifying object requirements (for example, if the extension of filename is ".csv", it will trigger use of write.csv).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

rodents_tab
Rodents data.frame data table.

Value

data.frame for the specified data set.

Examples

 # \donttest{
   create_dir()

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

   fill_raw()

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

   prep_rodents_table()

#>     moon BA DM DO DS NA OL OT  PB PE PF PH PL PM  PP RF RM RO SF SH SO total
#> 1     28  0 16  0 11  1  6 12   0  2  2  0  0  0   3  0  0  0  0  0  0    53
#> 2     29  0 15  5 15  5  4 10   0  3  1  0  0  0   2  0  1  0  0  0  0    61
#> 3     30  0 25  3 11  4  9 10   0  5  5  0  0  1   1  0  5  0  0  0  0    79
#> 434   31  0 30  5 11  4 12 10   0  3  6  0  0  0   0  0  6  0  0  0  0    87
#> 4     32  0 35  7 11  5 15  9   0  1  8  0  0  0   0  0  6  0  0  0  0    97
#> 5     33  0 19  6 11  1 10  4   0  3  7  0  0  0   0  0 11  0  0  0  0    72
#> 6     34  0 23  1 23  1  6  5   0  2 10  0  0  0   0  0 10  0  0  0  0    81
#> 7     35  0 22  2 26  1  6  5   0  1 11  0  0  0   1  0  6  0  0  0  0    81
#> 8     36  0 23  3 36  9  5  1   0  1  6  0  0  0   2  0  0  0  0  0  0    86
#> 9     37  0 21  6 30  4  2  4   0  2  3  0  0  0   3  0  0  0  0  0  0    75
#> 10    38  0 15  6 16  4  1  2   0  2  4  0  0  0   1  0  0  0  0  0  0    51
#> 11    39  0 14  4 17  4  2  2   0  0  6  0  0  0   2  0  0  0  0  0  0    51
#> 12    40  0 17  1 21  4  1  6   0  1  9  0  0  0   5  0  0  0  0  0  0    65
#> 13    41  0 19  1 18  3  7 14   0  1  5  0  0  0   2  0  1  0  0  0  0    71
#> 14    42  0 24  4 22  5  9  7   0  1  7  0  0  0   0  0  0  0  0  0  0    79
#> 15    43  0 32  7 21  6  7  9   0  1  6  0  0  0   0  0  0  0  0  0  0    89
#> 435   44  0 32 10 22  5  7  5   0  2  6  0  0  0   0  0  2  0  0  0  0    91
#> 16    45  0 31 14 24  4  7  1   0  3  5  0  0  0   1  0  3  0  0  0  0    93
#> 17    46  0 24  3 26  2  3 10   0  1  6  0  0  1   0  0  2  0  0  0  0    78
#> 436   47  0 28  6 29  2  2  6   0  2  7  0  0  1   0  0  2  0  0  0  0    85
#> 18    48  0 33  8 32  3  1  2   0  3  8  0  0  1   1  0  1  0  0  0  0    93
#> 19    49  0 33  9 34  6  2  3   0  2  4  0  0  2   4  0  2  0  0  0  0   101
#> 20    50  0 34  8 24  6  0  1   0  0  7  0  0  1   2  0  0  0  0  0  0    83
#> 21    51  0 23  4 13  2  3  4   0  0  2  0  0  0   5  0  0  0  0  0  0    56
#> 22    52  0 21  5 11  5  0  0   0  0  1  0  0  0   1  0  0  0  0  0  0    44
#> 23    53  0 21  3 12  5  2  6   0  4  0  0  0  0   0  0  1  0  0  0  0    54
#> 24    54  0 33  3 12  7  8 10   0  6  3  0  0  1   2  0  0  0  0  0  0    85
#> 25    55  0 49  6 19  9  6  8   0  7  5  0  0  2   0  0  5  0  0  0  0   116
#> 26    56  0 36  7 20  9  2  2   0 14  7  0  0  4   0  0 17  0  0  0  0   118
#> 27    57  0 43  7  9  6  6  5   0 14  8  0  0  3   0  0 15  0  0  0  0   116
#> 28    58  0 54  9 26 10  7  3   0 17 10  0  0  3   0  0 36  0  0  0  0   175
#> 29    59  0 48 10 25  7  5  5   0 11 10  0  0 11   2  0 34  0  0  0  0   168
#> 30    60  0 34 10 38  6  2  3   0 11  6  0  0 11   3  1 14  0  0  0  0   139
#> 31    61  0 41  7 39 12  4  5   0 10  7  0  0  3   3  0 11  0  0  1  0   143
#> 32    62  0 43  8 38 15  3  6   0 10 15  0  0  6   4  0 10  0  0  1  0   159
#> 33    63  0 44  8 35 19  3 11   0  9 19  0  0  9   6  0  4  0  0  0  0   167
#> 34    64  0 57  9 30 12  4  8   0  7 18  0  0 10   8  0  8  0  0  0  0   171
#> 35    65  0 48  7 20  5  6 13   0  3 10  0  0  4   4  0  0  0  0  0  0   120
#> 437   66  0 54 12 22  7 10 10   0  4 17  0  0  4   2  0  6  0  0  0  0   148
#> 36    67  0 61 17 25  9 13  7   0  4 24  0  0  4   0  1 12  0  0  0  0   177
#> 438   68  0 66 18 22  6 12  6   0  2 17  0  0  5   0  0 16  0  0  0  0   170
#> 37    69  0 71 19 18  3 11  4   0  1 10  0  0  6   0  0 19  0  0  0  0   162
#> 38    70  0 54 18 10  1  9  7   0  3  5  0  0  2   0  0 22  0  0  0  0   131
#> 39    71  0 57 18 19  0 11  5   0  3 22  0  0  1   0  0 25  0  0  1  0   162
#> 40    72  0 41 13 37  3  4  0   0  9 27  0  0  4   2  0 17  0  0  1  0   158
#> 41    73  0 38  6 30  5  0  9   0 12 14  0  0  3   5  0 14  0  0  0  0   136
#> 42    74  0 28 16 27  9  4  7   0  2 10  0  0  3   3  0  2  0  0  0  0   111
#> 43    75  0 33 10 13 10  4  8   0  0  6  0  0  4   5  0  4  0  0  0  0    97
#> 44    76  0 40  8 18 14  2  9   0  1  7  0  0  3   9  0  1  0  0  0  0   112
#> 45    77  0 40 15 27 22  4  4   0  2  6  0  0  3   6  0  4  0  0  0  0   133
#> 46    78  0 44 15 32 14  3 10   0  2  3  0  0  3   1  0  3  0  0  0  0   130
#> 47    79  0 42 16 26  9  6 14   0  0  3  0  0  6   1  0  8  0  0  0  0   131
#> 48    80  0 40 13 23  8  7  9   0  3  1  0  0  4   0  0  8  0  0  0  0   116
#> 439   81  0 40 12 20  6  6  6   0  2  0  0  0  3   0  0 12  0  0  0  0   107
#> 49    82  0 39 11 18  3  6  3   0  0  0  0  0  2   0  0 17  0  0  0  0    99
#> 50    83  0 27  8 17  1  5  1   0  0  0  0  0  2   2  0  3  0  0  0  0    66
#> 51    84  0 20  5 15  7  3  3   0  1  2  0  0  3   4  0  3  0  0  0  0    66
#> 440   85  0 25  6  9  8  6  4   0  2  2  0  0  2   4  0  4  0  0  0  0    72
#> 52    86  0 30  8  3 10  8  5   0  2  1  0  0  1   4  0  5  0  0  0  0    77
#> 53    87  0 28  7  1  7  8  5   0  3  0  0  0  0   4  0  8  0  0  0  0    71
#> 54    88  0 22  5  3 13  6  7   0  4  0  0  0  1   5  0  5  0  0  0  0    71
#> 55    89  0 22  4  0  9  4  6   0  2  1  2  0  0   1  0  3  0  0  0  0    54
#> 56    90  0 49  9  3  7  6  5   0  1  1  1  0  0   3  0  4  0  0  0  0    89
#> 57    91  0 50 12  3  3 14  4   0  0  1  3  0  0   1  0  4  0  0  0  0    95
#> 441   92  0 44 13  2  2  9  5   0  1  0  2  0  0   0  0  4  0  0  0  0    82
#> 58    93  0 37 14  1  0  4  6   0  2  0  1  0  1   0  0  4  0  0  0  0    70
#> 59    94  0 65 29  2  0  8  7   0  3  0  0  0  3   0  0 11  0  0  0  0   128
#> 60    95  0 53 18  3  1 13  2   0  5  1  0  0  1   0  0  8  0  0  0  0   105
#> 61    96  0 47 18  4  1 14  4   0  4  0  0  0  4   0  0  5  0  0  0  0   101
#> 442   97  0 40 14  6  2  9  4   0  3  0  0  0  2   2  0  6  0  0  0  0    88
#> 62    98  0 32 11  7  3  4  3   0  2  0  0  0  1   4  0  6  0  0  0  0    73
#> 63    99  0 37 13 12  2  2  0   0  1  0  0  0  0   3  0  8  0  0  0  0    78
#> 64   100  0 43 13 10  9  3  4   0  2  0  0  0  1   3  0  3  0  0  0  0    91
#> 65   101  0 42 17  6 10  3  5   0  0  0  0  0  0   7  0  1  0  0  0  0    91
#> 66   102  0 50 13  8  4  5  6   0  5  0  1  0  0   8  0  2  0  0  0  0   102
#> 67   103  0 50 18 11  5 13  9   0  8  0  0  0  1   2  0  3  0  0  0  0   120
#> 68   104  0 61 17 11  4  7  8   0  9  0  1  0  0   1  0  7  0  0  0  0   126
#> 69   105  0 58 16  9  4  7  8   0  8  0  1  0  0   1  0 14  0  0  0  0   126
#> 443  106  0 53 16 11  4  7  6   0  8  0  1  0  0   1  0 14  0  0  0  0   121
#> 444  107  0 48 15 12  5  6  5   0  7  0  1  0  0   2  0 13  0  0  0  0   114
#> 70   108  0 43 15 14  5  6  3   0  7  0  1  0  0   2  0 13  0  0  0  0   109
#> 71   109  0 43 11 13  5  5  1   0  2  0  0  0  2   3  0  4  0  0  0  0    89
#> 72   110  0 39  9 12 10  2  4   0  8  0  0  0  5   2  0 10  0  0  0  0   101
#> 73   111  0 39 11  5 11  2  4   0  2  0  0  0  4   4  0  3  0  0  0  0    85
#> 74   112  0 44 12  7  7  3  3   0  2  1  0  0  4   5  0  8  0  0  0  0    96
#> 75   113  0 37  6  7  4  2  0   0  0  0  0  0  1   4  0  0  0  0  0  0    61
#> 76   114  0 37  7  7  6  4  3   0  0  1  0  0  0   5  0  0  0  0  0  0    70
#> 77   115  0 36  6  6  6  5  3   0  2  0  0  0  2   1  0  0  0  0  1  0    68
#> 78   116  0 50 18  9  2 10  3   0  0  0  0  0  8   0  0  5  0  0  0  0   105
#> 79   117  0 38 20  8  4  4  1   0  3  0  0  0  8   0  0 10  0  0  0  0    96
#> 445  118  0 40 20  8  6  6  2   0  6  0  0  0 10   0  0 16  0  0  0  0   114
#> 80   119  0 43 21  9  7  9  3   0  8  0  0  0 11   0  0 23  0  0  0  0   134
#> 81   120  0 45 29 10  5 10  0   0  4  0  0  0  8   0  0 32  0  0  0  0   143
#> 82   121  0 38 27 12  7  5  3   0 10  0  0  0 11   2  0 14  0  0  0  0   129
#> 83   122  0 33 22 14 11  4  0   0 15  0  0  0  8   6  0  4  0  0  0  0   117
#> 84   123  0 40 16 10 13  8  1   0 13  0  0  0 16   3  0 12  0  0  0  0   132
#> 85   124  0 49 17 14  9  5  5   0  9  0  0  0 12   5  0 18  0  0  0  0   143
#> 86   125  0 50 15  8 11  6  2   0 13  0  0  0 15   9  0 20  0  0  1  0   150
#> 87   126  0 44 16  7 13  7  6   0 13  0  0  0 12   8  0 12  0  0  1  0   139
#> 88   127  0 37 15  8 20  6 13   0  9  0  0  0 10  11  0  2  0  0  0  0   131
#> 89   128  0 39 16  7 18  8  7   0 20  2  0  0 16   1  0 15  0  0  2  0   151
#> 90   129  0 51 19  5 14 11  3   0 28  2  0  0 14   0  0 37  0  0  1  0   185
#> 446  130  0 46 22  5 10  6  3   0 22  1  0  0 12   0  0 40  0  0  2  0   169
#> 91   131  0 42 24  5  7  2  3   0 15  0  0  0 11   0  0 42  0  0  2  0   153
#> 92   132  0 38 24  5  9  5  5   0 18  0  0  0 11   0  1 38  0  0  3  0   157
#> 93   133  0 42 17  4  6  4  3   0 12  1  0  0 12   2  1 32  0  0  2  0   138
#> 94   134  0 23 11  6  7  5  8   0 16  1  0  0  3   3  1 14  0  0  3  0   101
#> 95   135  0 26 12  7  8  1  4   0 10  0  0  0  2   5  0  2  0  0  3  0    80
#> 96   136  0 28 10  6 11  4  7   0 19  0  0  0  1   8  1  9  0  0  2  0   106
#> 97   137  0 25 10  4 12  3  6   0 11  0  0  0  3   8  0  3  0  0  2  0    87
#> 98   138  0 26 11  4 11  2  6   0 17  0  0  0  2  18  0  7  0  0  3  0   107
#> 99   139  0 22  4  2  9  2 11   0 11  0  0  0  1   7  1  1  0  0  1  0    72
#> 100  140  0 31  5  3  9  1 14   0 16  0  0  0  3   2  0  1  0  0  2  0    87
#> 101  141  0 28  6  4  8 15 13   0 28  0  1  0  3   0  1 19  0  0  6  0   132
#> 102  142  0 34 10  4  5  7  7   0 17  0  0  0  2   0  5 43  0  0  2  0   136
#> 103  143  0 35  9  3  6 10 14   0 18  0  0  0  4   0  6 49  0  0  6  0   160
#> 104  144  0 35 13  3  5 14  9   0 26  0  0  0  6   0  8 60  0  0  4  0   183
#> 105  145  0 22  6  5  6  7  7   0 24  0  0  0  5   1  5 57  0  0  8  0   153
#> 106  146  0 22  7  4  6  7  3   0 15  0  0  0  1   8  7 44  0  0  6  0   130
#> 107  147  0 26  8  3  9  4  6   0 16  0  0  0  4   3  3 25  0  0  7  0   114
#> 108  148  0 26  6  3  7  3  3   0 19  0  0  0  4   1  4 35  0  0  7  0   118
#> 109  149  0 22  5  3  9  2  4   0  7  0  0  0  2   3  5 24  0  0  7  0    93
#> 110  150  0 31  9  2  5  1  8   0  8  0  0  0  1   8  1 12  0  0  2  0    88
#> 111  151  0 20 12  1  2  2 12   0  4  2  0  0  0   8  0  0  0  0  4  0    67
#> 112  152  0 27 11  2  4  1 13   0 12  3  0  0  0   6  2  9  0  0  3  0    93
#> 113  153  2 25 15  2  5  5 17   0 15  5  1  0  0   6  4 30  0  2  1  0   135
#> 114  154  1 30 18  2  3  1  6   0 13  3  1  0  0   1  4 53  0  3  0  0   139
#> 115  155  1 27 18  2  1  2  6   0  4  3  0  0  0   1  2 24  0  5  2  0    98
#> 116  156  0 33 13  1  1  2  4   0  9  5  1  0  0   0  4 30  0  3  0  0   106
#> 117  157  2 35 17  1  5  2  6   0 10  5  1  0  0   0  1 31  0  3  2  0   121
#> 118  158  0 37 17  2  3  1  6   0 14  6  1  0  0   0  3 30  0  2  3  0   125
#> 119  159  2 37 15  1  2  3  5   0  8  4  1  0  0   4  2 23  0  2  1  0   110
#> 120  160  1 35 12  1  3  0  5   0  4  3  1  0  0   2  0 11  0  2  1  0    81
#> 121  161  1 31 14  2  1  0  4   0  2  5  0  0  0   3  0 11  0  1  0  0    75
#> 122  162  0 37 11  1  1  2  7   0  2  2  2  0  0   6  0  5  0  0  0  0    76
#> 123  163  1 33 14  2  1  1  7   0  2  5  0  0  0   7  0  3  0  0  0  0    76
#> 124  164  0 27  9  1  4  2  2   0  5  1  0  0  0   8  0  1  0  0  0  0    60
#> 125  165  1 41  9  1  3  2  8   0  3  4  0  0  0   3  0 10  0  0  0  0    85
#> 126  166  0 42 10  1  2  1  9   0  4  3  0  0  0   0  0 22  0  0  0  0    94
#> 127  167  2 47 12  1  2  7  8   0  8  5  0  0  0   0  0 30  0  0  0  0   122
#> 128  168  1 40 13  1  2  6  9   0  6  4  0  0  0   1  0 40  0  0  0  0   123
#> 129  169  1 43 15  1  3  5  8   0  2  3  0  0  0   0  0 43  0  1  0  0   125
#> 130  170  1 44 14  1  2  2  4   0  3  2  0  0  0   0  0 41  0  1  0  0   115
#> 131  171  2 32 11  1  6  4  2   0  2  1  0  0  0   4  0 18  0  0  0  0    83
#> 132  172  2 27 10  1  3  2  7   0  3  2  0  0  0   9  0 11  0  0  0  0    77
#> 133  173  2 31  9  1  2  7  2   0  5  1  0  0  0   3  0 10  0  0  0  0    73
#> 134  174  2 29  7  1  5  0  6   0  4  2  0  0  0  19  0 13  1  2  0  0    91
#> 135  175  1 30  7  1  3  0 14   0  5  5  0  0  0  19  0  7  0  2  0  0    94
#> 136  176  1 27  8  1  1  0 12   0  7  7  0  0  0  12  0  2  0  0  1  0    79
#> 137  177  3 30  8  1  0  0 12   0  7 13  0  0  0  14  0  7  0  0  0  1    96
#> 138  178  5 34  7  1  2  2 18   0 10 15  0  0  0   4  0 57  0  1  0  7   163
#> 139  179  5 36 14  0  1  8 13   0 11  7  0  0  0   3  0 58  0  0  0  3   159
#> 140  180  3 28 10  1  1  1 12   0  5  1  1  0  0   2  0 26  0  1  0  1    93
#> 141  181  1 28 17  1  1  6  7   0 14  2  1  0  0   2  0 20  0  0  0  5   105
#> 142  182  2 25 12  1  0  3  3   0 11  4  1  0  0   4  0 34  0  0  0  2   102
#> 143  183  0 33 10  2  0  3  0   0  3  6  3  0  0   6  0  8  0  0  0  1    75
#> 144  184  0 28  5  1  0  3  1   0  1  3  1  0  0  12  0  3  0  0  0  3    61
#> 145  185  0 31  4  2  4  1  5   0  2  2  0  0  0   7  0  4  0  1  0  1    64
#> 146  186  0 24  9  2  1  3  3   0  2  4  0  0  0  22  0  8  0  0  0  1    79
#> 147  187  0 18  7  1  2  4  4   0  0  3  0  0  0  22  0  8  0  0  0  1    70
#> 148  188  0 19  8  2  1  5  2   0  1  7  0  0  0  30  0  3  0  0  0  2    80
#> 149  189  0 23  6  2  2  7  1   0  3  8  0  0  0  21  0 10  0  0  0  1    84
#> 150  190  0 29  5  1  2  7  4   0  0 14  0  0  0   3  0 19  0  0  0  2    86
#> 447  191  0 25  3  2  2  4  2   0  0  8  0  0  0   2  0 17  0  0  0  1    66
#> 151  192  0 21  1  2  1  2  0   0  0  1  0  0  0   1  0 15  0  1  0  0    45
#> 152  193  0 17  4  2  2  3  1   0  2  3  0  0  0   0  0 16  0  0  0  0    50
#> 153  194  0 15  2  1  0  4  2   0  0  4  0  0  0   2  0 23  0  1  0  0    54
#> 154  195  0 43  2  2  2  4  3   0  1  8  0  0  0   6  0 12  0  0  0  0    83
#> 155  196  0 15  2  2  3  2  0   0  1  4  0  0  0  21  0  5  0  0  0  1    56
#> 156  197  0 18  5  3  1  3  4   0  1  2  0  0  0  11  0  2  0  0  0  1    51
#> 157  198  0 19  3  3  1  3  1   0  4  2  0  0  0  24  0  4  0  1  0  1    66
#> 158  199  0 19  1  1  2  4  0   0  1  8  0  0  0  16  0  6  0  1  0  1    60
#> 159  200  0 21  1  0  1  2  3   0  1  8  0  0  0  11  0  7  0  0  0  1    56
#> 160  201  0 18  2  1  3  2  0   0  0 12  0  0  0   6  0  6  0  0  0  0    50
#> 161  202  0 24  1  1  3  4  1   0  0  7  0  0  0   3  0  5  0  0  0  0    49
#> 162  203  0 24  3  1  4  1  4   0  2  7  0  0  0   0  0  8  0  0  0  1    55
#> 163  204  0 20  3  1  2  3  3   0  0  5  0  0  0   0  0  9  0  0  0  0    46
#> 164  205  0 26  1  0  1  2  7   0  0  2  0  0  0   0  0  3  0  0  0  2    44
#> 165  206  0 29  0  0  2  0  3   0  0  5  0  0  0   0  0  5  0  0  0  0    44
#> 166  207  0 26  1  1  1  0  4   0  0  3  0  0  0   0  0  8  0  0  0  0    44
#> 167  208  0 37  2  2  2  0  4   0  0  7  0  0  0   7  0  6  0  0  0  0    67
#> 168  209  0 24  2  1  1  1  2   0  0  6  0  0  0   5  0  3  0  0  0  1    46
#> 169  210  0 15  2  1  0  3  0   0  0  3  0  0  0  11  0  1  0  0  0  0    36
#> 170  211  0 17  3  1  0  1  1   0  0  2  0  0  0  10  0  4  0  0  0  0    39
#> 171  212  0 17  0  1  1  3  0   0  0  6  0  0  0  16  0  2  0  0  0  0    46
#> 172  213  0 17  0  1  0  4  0   0  0  6  0  0  0  13  0  2  0  0  0  0    43
#> 173  214  0 28  4  1  1  3  0   0  0 10  0  0  0  10  0  4  0  0  0  0    61
#> 174  215  0 28  4  0  1  2  0   0  1 14  0  0  0   2  0  1  0  0  0  0    53
#> 175  216  0 29  6  0  0  2  0   0  2  9  0  0  0   0  0  1  0  0  0  0    49
#> 176  217  0 25  6  0  0  2  0   0  0  5  0  0  0   0  0  2  0  0  0  0    40
#> 177  218  0 32  6  0  0  2  0   0  1  4  0  0  0   0  0  6  0  0  0  0    51
#> 178  219  0 30  6  0  0  2  0   0  0  6  0  0  0   1  0  7  0  0  0  0    52
#> 179  220  0 36  5  0  0  0  0   0  0  8  0  0  0   3  0  6  0  0  0  0    58
#> 180  221  0 28  6  0  0  0  1   0  2  4  0  0  0   6  0  7  0  0  0  0    54
#> 181  222  0 32  2  0  1  4  0   0  3  4  0  0  1   9  0  1  0  0  0  0    57
#> 182  223  0 33  1  0  0  2  1   0  0  4  0  0  1  17  0  6  0  0  0  0    65
#> 183  224  0 33  5  0  1  3  1   0  0  9  2  0  1  40  0  5  0  0  0  0   100
#> 184  225  0 31  1  0  0  4  6   0  1  9  0  0  0  60  0  5  0  0  0  0   117
#> 185  226  0 33  4  0  1  5  7   1  1 28  1  1  0  67  0 12  0  0  0  0   161
#> 186  227  0 41  2  0  2  3  9   3  4 37  0  0  0  51  0  4  0  0  0  0   156
#> 187  228  0 42  6  0  2  2  5   1  1 27  0  0  2  17  0 13  0  0  0  0   118
#> 188  229  0 40  9  0  1  0  8   2  1 13  0  2  2   6  0  7  0  0  0  0    91
#> 189  230  0 57 20  0  0  1  1   3  4 24  0  3  1   5  0  4  0  0  0  0   123
#> 190  231  0 58 17  0  0  3  4   2  7 34  0  0  6  11  0  8  0  0  0  0   150
#> 191  232  0 35 11  0  0  2  5   1  8 25  0  1  8  28  0 20  0  0  0  0   144
#> 192  233  0 48 30  0  0  3  3   0  7 58  0  1 10  40  0 25  0  0  0  0   225
#> 193  234  0 38 13  0  1  1  3   0  2 35  0  0  1  27  0  3  0  0  0  0   124
#> 194  235  0 35 18  0  0  2  5   2  2 36  0  0  4  41  0  3  0  0  0  0   148
#> 195  236  0 22 13  0  0  0  5   1  0 21  0  0  0  38  0  1  0  0  0  0   101
#> 196  237  0 23  6  0  0  0  4   2  0 13  0  0  0  57  0  0  0  0  0  0   105
#> 197  238  0 39  8  0  1  0 26   9  1 25  0  0  2  34  0  2  0  0  0  0   147
#> 198  239  0 41  9  0  2  1 22   8  2 27  0  1  1  16  0  0  0  0  0  0   130
#> 199  240  0 45 11  0  3  0 18   5  3 19  0  0  5   1  0  8  0  0  0  0   118
#> 200  241  0 51 18  0  0  0 12   6  0 13  0  0 12   0  0 16  0  0  0  0   128
#> 448  242  0 69 26  0  0  0 16   8  1 17  0  2 13   0  0 28  0  0  0  0   180
#> 201  243  0 87 35  0  0  1 19  10  2 21  0  5 14   0  0 39  0  0  0  0   233
#> 202  244  0 91 28  0  1  1 20   8  7 34  0  0 16  15  0 33  0  0  0  0   254
#> 203  245  0 55 25  0  0  3 16   6  4 30  1  0 28  65  1 33  0  0  0  0   267
#> 204  246  0 43 29  0  6  1 29  19  7 21  1  0 31  52  0  9  0  0  0  3   251
#> 205  247  0 51 23  0  8  1 33  41  7 16  1  0 38  54  0  7  0  0  0  0   280
#> 206  248  0 49 27  1  4  1 27  45  6 25  0  3 29  77  0  9  0  0  2  0   305
#> 207  249  0 41 20  1  9  0 16  31  5 23  0  2 15  50  0  5  0  0  2  0   220
#> 449  250  0 37 19  1  8  0 24  32  4 16  0  4 20  30  0  6  0  0  1  0   202
#> 208  251  0 33 18  1  6  0 31  34  2 10  0  5 26   9  0  6  0  0  0  0   181
#> 209  252  0 42 15  1  8  0 21  25  8  5  0  1 35   1  0  5  0  0  1  0   168
#> 210  253  0 47 18  1  3  0 27  25  6  0  0  1 22   0  0  6  0  0  2  0   158
#> 211  254  0 37 15  1  3  0 19  15  3  0  0  2 17   1  0  6  0  0  0  0   119
#> 212  255  0 54 13  1  1  0 21  17  0  0  0  2 15   0  0  2  0  0  0  0   126
#> 213  256  0 50 17  2  2  0 16  21  1  0  0  0  8   0  0  3  0  0  0  0   120
#> 450  257  0 46 11  1  3  0 13  16  2  0  0  1  8  10  0  2  0  0  0  0   113
#> 214  258  0 43  5  0  4  0 10  11  3  1  0  2  7  21  0  1  0  0  0  0   108
#> 215  259  0 60 12  0  7  0 14  16  4  6  0  3  7  23  0  0  0  0  0  0   152
#> 216  260  0 35 10  1  6  0 12  30  6  7  0  0  8  26  0  0  0  0  0  0   141
#> 217  261  0 34  7  1  1  0 10  39  2  6  0  0 11  40  0  1  0  0  0  0   152
#> 218  262  0 29  5  0  2  0 10  53  3  3  0  0  8  45  0  0  0  0  0  0   158
#> 219  263  0 32 11  1  4  0 10  45  2  1  0  0  9  40  0  0  0  0  0  0   155
#> 220  264  0 36 10  1  2  0 24  37  0  0  0  0 16   8  0  0  0  0  0  0   134
#> 221  265  0 48 13  1  1  0 19  30  1  0  0  0  7   0  0  3  0  0  0  0   123
#> 222  266  0 49  5  1  1  0 15  20  0  0  0  0  6   0  0  2  0  0  0  0    99
#> 223  267  0 55 12  1  2  0 21  23  2  0  0  0  7   1  0  6  0  0  0  0   130
#> 451  268  0 54 10  1  2  0 17  24  1  0  0  0  8   2  0  6  0  0  0  0   125
#> 224  269  0 54  9  1  2  0 13  25  0  0  0  0 10   2  0  7  0  0  0  0   123
#> 225  270  0 54 16  2  2  0 11  33  0  0  0  0  7   9  0  1  0  0  0  0   135
#> 226  271  0 41 12  1  2  0 12  28  1  0  0  0  5  23  0  2  0  0  0  0   127
#> 227  272  0 32 10  1  1  0  7  39  1  0  0  0  5  20  0  1  0  0  0  0   117
#> 452  273  0 22  7  1  2  0  5  29  1  0  0  0  3  17  0  1  0  0  0  0    88
#> 453  274  0 13  3  0  4  0  3  18  0  0  0  0  2  15  0  0  0  0  0  0    58
#> 228  275  0  3  0  0  5  0  1   8  0  0  0  0  0  12  0  0  0  0  0  0    29
#> 229  276  0 16  5  0  1  0  6  20  2  0  0  0  1  23  0  0  0  0  0  0    74
#> 230  277  0 18  3  0  2  0 13  31  0  0  0  0  0  44  0  0  0  0  1  0   112
#> 231  278  0 11  5  0  2  0  6  37  0  0  0  0  0  33  0  3  0  0  1  0    98
#> 232  279  0 19 11  0  4  0  7  40  0  0  0  0  1  32  0  3  0  0  1  0   118
#> 233  280  0 19 13  0  3  0  9  57  0  0  0  0  0  31  0  0  0  0  0  0   132
#> 454  281  0 19 13  0  3  0 10  55  1  0  0  0  0  33  0  1  0  0  0  0   135
#> 455  282  0 20 13  0  2  0 10  52  1  0  0  0  0  35  0  1  0  0  1  0   135
#> 234  283  0 20 13  0  2  0 11  50  2  0  0  0  0  37  0  2  0  0  1  0   138
#> 235  284  0 18 13  0  3  0  8  63  1  0  0  0  0  59  0  1  0  0  3  0   169
#> 236  285  0 19  7  0  2  0 11  74  1  0  0  0  0  72  0  0  0  0  1  0   187
#> 456  286  0 22  6  0  2  0 14  68  2  0  0  0  0  65  0  0  0  0  0  0   179
#> 237  287  0 25  6  0  1  0 18  62  2  0  0  0  0  58  0  0  0  0  0  0   172
#> 238  288  0 30  5  0  3  0 25  49  2  0  0  0  1  14  0  1  0  0  0  0   130
#> 457  289  0 28  6  0  4  0 24  41  4  0  0  0  0   8  0  3  0  0  0  0   118
#> 239  290  0 27  7  0  4  0 23  33  6  0  0  0  0   3  0  5  0  0  0  0   108
#> 240  291  0 23  4  0  2  0 28  31  1  0  0  0  0   1  0  0  0  0  0  0    90
#> 458  292  0 24  3  0  2  0 21  26  1  0  0  0  0   7  0  0  0  0  0  0    84
#> 459  293  0 25  1  0  2  0 15  22  0  1  0  0  1  12  0  1  0  0  0  0    80
#> 241  294  0 26  0  0  2  0  8  17  0  1  0  0  1  18  0  1  0  0  0  0    74
#> 242  295  0 26  8  0  2  0  6  21  1  3  0  0  1  23  0  0  0  0  1  0    92
#> 243  296  0 32  7  0 10  0  6  28  3  2  0  0  0  16  0  0  0  0  3  0   107
#> 460  297  0 32  7  0  8  0 12  46  3  3  0  0  0  33  0  0  0  0  2  0   146
#> 461  298  0 31  7  0  5  0 17  63  3  4  0  0  0  49  0  0  0  0  1  0   180
#> 244  299  0 31  7  0  3  0 23  81  3  5  0  0  0  66  0  0  0  0  0  0   219
#> 245  300  0 35  7  0  4  0 22  87  2  2  0  0  0  52  0  0  0  0  0  0   211
#> 246  301  0 37 12  0  5  0 38  78  4  8  0  0  0  24  0  0  0  0  1  0   207
#> 247  302  0 34 19  0  8  0 28  80 12  1  0  0  0   7  0  2  0  0  1  0   192
#> 462  303  0 35 22  0  7  0 22  74 10  1  0  0  0   6  0  4  1  0  1  0   183
#> 463  304  0 36 25  0  5  0 16  69  9  2  0  0  1   5  0  7  2  0  1  0   178
#> 248  305  0 37 28  0  4  0 10  63  7  2  0  0  1   4  0  9  3  0  1  0   169
#> 249  306  0 37 26  0  5  0 10  53  2  4  0  0  0  13  0  2  0  0  1  0   153
#> 250  307  0 34 31  0  5  0 12  54  1  2  0  0  0  49  0  0  0  0  3  0   191
#> 251  308  0 37 33  0  5  0  6  88 11  2  0  0  0  51  0  2  0  0  2  0   237
#> 252  309  0 27 23  0  4  0  6  95  7  1  0  0  0  59  0  2  0  0  0  0   224
#> 253  310  0 23 19  0  4  0 11  95  8  1  0  0  0  86  0  2  1  0  3  0   253
#> 464  311  0 18 13  0  4  0  8  61  4  0  0  0  0  56  0  1  0  0  2  0   167
#> 254  312  0 14  7  0  3  0  6  27  0  0  0  0  0  26  0  0  0  0  0  0    83
#> 255  313  0 25 12  0  5  2  7 100  5  1  0  0  0  53  0  0  0  0  0  0   210
#> 256  314  0 21 20  0  2  3 21 110  6  2  0  0  0  25  0  2  0  1  1  0   214
#> 257  315  0 23 22  0  5  1 18 105  5  2  0  0  0  11  0  0  0  1  0  0   193
#> 258  316  0 21 20  0  5  2  9  77  3  1  0  0  0   7  0  1  3  5  0  0   154
#> 259  317  0 31 33  0  4  1 15 109  9  6  0  0  0  11  0  9  1  4  1  0   234
#> 260  318  0 34 48  0  4  1 14 106 18 12  0  0  0  15  0  5  1  5  2  0   265
#> 261  319  0 35 34  0  8  4 18  97 16  9  0  0  0  52  0  8  1  5  1  0   288
#> 262  320  0 32 26  0  5 12 36 116 19  7  0  0  0  65  0  6  2  4  1  0   331
#> 263  321  0 32 19  0  7  6 29 136  7  9  0  0  1  60  0  3  0  4  1  0   314
#> 264  322  0 29 18  0  4  4 28 107 10  6  0  0  0  95  0  2  0  2  1  0   306
#> 265  323  0 26 13  0  6  2 25 108 11  4  0  0  0  97  0  2  0  3  2  0   299
#> 266  324  0 31 10  1  2  4 26  91  7  2  0  0  0  68  0  3  0  0  0  0   245
#> 267  325  0 20 12  0  1  3 29  83  8  2  0  0  0  26  0  5  0  0  0  0   189
#> 268  326  0 20  8  0  2  1 24  63 10  0  0  0  0   7  0  3  0  0  0  0   138
#> 269  327  0 17 13  0  3  0 23  58  9  0  0  0  0   3  0  1  0  0  0  0   127
#> 270  328  0 30  8  0  1  0 13  52  6  0  0  0  0   0  0  1  0  0  0  0   111
#> 271  329  0 24  9  0  2  3 15  45  5  0  0  0  0   0  0  1  0  0  0  0   104
#> 272  330  0 24 13  0  2  0  9  39  4  0  0  0  0   1  0  0  0  0  0  0    92
#> 273  331  0 21  6  0  0  0 11  39  2  0  0  0  0   5  0  0  0  0  0  0    84
#> 465  332  0 24 11  0  1  0 10  33  2  1  0  0  0  16  0  0  0  0  0  0    98
#> 274  333  0 26 16  0  2  0  9  27  1  2  0  0  0  28  0  0  0  0  0  0   111
#> 466  334  0 23 14  0  1  0  9  33  1  2  0  0  0  34  0  0  0  0  0  0   117
#> 467  335  0 19 11  0  1  0 10  38  1  2  0  0  0  41  0  0  0  0  0  0   123
#> 275  336  0 16  9  0  0  0 10  44  1  2  0  0  0  47  0  0  0  0  0  0   129
#> 276  337  0  8  5  0  0  0  4  43  0  0  0  0  0  61  0  0  0  0  0  0   121
#> 277  338  0 14 10  0  2  0 15  48  4  0  0  0  0  54  0  0  0  0  0  0   147
#> 468  339  0 18 11  0  2  0 14  46  3  2  0  0  0  32  0  0  0  0  0  0   128
#> 278  340  0 21 12  0  1  0 14  45  2  4  0  0  0  11  0  0  0  0  0  0   110
#> 279  341  0 11 26  0  2  0  9  39  5  2  0  0  1   3  0  1  0  0  0  0    99
#> 469  342  0 12 22  0  1  0  9  34  4  1  0  0  1  16  0  1  0  0  0  0   101
#> 470  343  0 13 19  0  1  0 10  30  3  1  0  0  1  28  0  0  0  0  0  0   106
#> 280  344  0 14 15  0  0  0 10  25  2  0  0  0  1  41  0  0  0  0  0  0   108
#> 471  345  0 14 17  0  1  0  8  34  3  0  0  0  0  46  0  0  0  0  0  0   123
#> 281  346  0 15 19  0  2  0  7  43  4  0  0  0  0  52  0  1  0  0  0  0   143
#> 282  347  0 12 11  0  1  0 11  62  5  0  0  0  0  69  0  1  0  0  0  0   172
#> 283  348  0 16 12  0  1  0  8  80  2  0  0  0  0 110  0  1  0  0  0  0   230
#> 284  349  0 17 12  0  1  0 14  99  0  0  0  0  0 104  0  1  0  0  0  0   248
#> 285  350  0 12 10  0  0  2 19  99  4  1  0  0  0  97  0  0  0  0  0  0   244
#> 286  351  0 20 12  0  1  2 30  96  5  5  0  0  0  41  0  2  0  0  0  0   214
#> 287  352  0 18 14  0  1  1 17  60  7  0  0  0  1   6  0  3  2  0  0  0   130
#> 288  353  0 24 17  0  0  0 17  64  5  0  0  0  1   8  0  1  1  0  0  0   138
#> 289  354  0 23 18  0  2  1 13  58  7  0  0  0  1   8  0  3  2  0  0  0   136
#> 290  355  0 36  9  0  1  0  6  55  6  1  0  0  0  15  0  2  1  0  0  0   132
#> 291  356  0 24  9  0  1  0  8  58  5  3  0  0  0  28  0  1  0  0  0  0   137
#> 292  357  0 30 12  0  0  0  4  47  3  2  0  0  0  68  0  0  0  0  0  0   166
#> 293  358  0 24 12  0  3  1  3  56  1  1  0  0  1  67  0  0  0  0  0  0   169
#> 472  359  0 22 10  0  2  0  2  66  1  1  0  0  0 100  0  0  0  0  0  0   204
#> 294  360  0 20  8  0  1  0  2  77  1  1  0  0  0 132  0  0  0  0  0  0   242
#> 473  361  0 14  5  0  0  0  2  53  0  0  0  0  0  96  0  0  0  0  0  0   170
#> 295  362  0  8  2  0  0  0  1  29  0  0  0  0  0  59  0  0  0  0  0  0    99
#> 296  363  0 19  2  0  1  0  3  59  1  0  0  0  0 103  0  0  0  0  0  0   188
#> 297  364  0 16  5  0  1  0  5  78  2  1  0  0  1 169  0  0  0  0  1  0   279
#> 298  365  0 24  6  0  1  0  3 102  7  0  0  0  1 132  1  2  1  0  8  0   288
#> 474  366  0 22 10  0  2  0  3 100  7  1  0  0  2 103  0  2  1  0 18  0   271
#> 299  367  0 20 15  0  3  0  3  99  7  2  0  0  2  74  0  3  1  0 28  0   257
#> 300  368  0 18 23  0  1  2 11  90 13  6  0  0  7 134  0 14  0  0 32  0   351
#> 301  369  0 25 13  0  2  0  6 113 18  0  0  1  3  99  0 13  0  0 29  0   322
#> 302  370  0 20 15  0  4  1 14 197 10  4  0  1  9  86  0  3  0  2 37  0   403
#> 303  371  0 29 29  0  4  0 16 231 11 10  0  0 10 131  0 12  0  2 36  0   521
#> 304  372  0 23 20  0  2  2 26 248 17 16  0  3  5 160  0  6  0  2 24  0   554
#> 475  373  1 25 25  0  3  2 38 217 11 18  0  5  6 149  0  4  0  7 18  0   529
#> 305  374  2 27 30  0  4  1 49 186  5 19  0  7  6 138  0  1  0 12 13  0   500
#> 306  375 24 34 32  0  7  5 52 141 17 12  0  9 12  43  0  9  0 21  4  0   422
#> 307  376 24 40 32  0  9 10 63 134 11 13  0 16 16  26  0 41  0 13 10  0   458
#> 308  377 12 43 41  0  2  4 31  81  6  1  0  9 19   4  0 52  0 17  2  0   324
#> 309  378  4 40 38  1  3  3 18  67  4  1  0  9 11   1  0 37  0 16  3  0   256
#> 310  379  2 45 40  1  5  4 12  61  4  0  0  7 11   7  0 31  0 21  1  0   252
#> 311  380  2 44 42  1  5  4 15  60  5  1  0  6 14  13  0 31  0  9  3  0   255
#> 312  381  9 41 46  2  4  1 10  62  3  8  0  3 12  45  0 22  0 14  5  0   287
#> 313  382  4 43 56  2  8  4 14  61  4 10  0  4 10  75  0 20  0 17  4  0   336
#> 314  383  8 46 51  3  7  3 14  63  6 19  0  1  5  69  0  7  0 14  1  0   317
#> 315  384  4 40 42  4  5  3 12  70  3 20  0  1  3  83  0  1  0  5  1  0   297
#> 316  385  0 24 25  1  1  2  6  20  2  4  0  0  1  42  0  0  0  5  1  0   134
#> 317  386  0 12 12  3  1  0  9  41  3  2  0  0  0  49  0  0  0  2  0  0   134
#> 318  387  0 13 16  2  6  0 11  44  1  2  0  0  3  48  0  0  0  1  0  0   147
#> 319  388  0 15 18  4  4  1 12  39  3  2  0  0  1  17  0  0  0  0  0  0   116
#> 320  389  0 21 21  3  6  1  7  31  5  2  0  0  1   5  0  1  0  1  0  0   105
#> 321  390  0 22 24  2  2  0  5  24  1  2  0  0  0   0  0  0  0  0  0  0    82
#> 322  391  1 20 23  1  5  2  6  22  0  0  0  0  0   0  0  2  0  3  0  0    85
#> 323  392  1 26 22  1  5  0  3  21  2  2  0  0  0   3  0  6  0  1  0  0    93
#> 324  393  0 23 29  0  5  3  3  18  3  4  0  0  0  19  0  7  0  2  0  0   116
#> 325  394  3 18 25  0  5  5  7  37  1  4  0  0  0  50  0  1  0  3  0  0   159
#> 326  395  3 21 28  0  3  2  2  50  2  8  0  0  0  51  0  1  0  1  0  0   172
#> 327  396  1 15 24  1  2  3  1  56  1  5  0  0  0  60  0  0  0  3  0  0   172
#> 328  397  2 16 18  2  2  3  2  36  1  3  0  1  0  49  0  0  0  0  0  0   135
#> 329  398  1 15 13  2  3  2  5  24  1  0  0  0  0  28  0  0  0  0  0  0    94
#> 330  399  0 14  5  3  3  3  5  14  0  0  0  0  1   9  0  0  0  0  0  0    57
#> 331  400  0 12  3  2  1  2  4  11  0  0  0  0  0   4  0  0  0  0  0  0    39
#> 332  401  0 11  1  2  1  2  3  10  0  0  0  0  0   1  0  0  0  0  0  0    31
#> 333  402  0 11  1  2  1  1  4   6  1  0  0  0  0   0  0  0  0  0  0  0    27
#> 334  403  0 12  0  1  1  0  6   3  0  0  0  0  0   0  0  0  0  0  0  0    23
#> 335  404  0 15  1  1  1  3  4   4  0  0  0  0  0   2  0  0  0  0  0  0    31
#> 336  405  0 16  1  0  0  1  2   6  0  0  0  0  0   7  0  0  0  0  0  0    33
#> 337  406  0 17  1  0  0  1  1   1  0  0  0  0  0  22  0  0  0  0  0  0    43
#> 338  407  0 21  1  0  0  1  0   2  0  1  0  0  0  22  0  0  0  0  0  0    48
#> 476  408  0 18  1  0  0  1  1   2  0  1  0  0  0  30  0  0  0  0  0  0    54
#> 339  409  0 15  1  0  0  1  2   2  0  1  0  0  0  39  0  0  0  0  0  0    61
#> 477  410  0 15  2  0  1  1  1   2  0  2  0  0  0  72  0  0  0  0  0  0    96
#> 340  411  0 15  3  0  2  1  0   2  0  2  0  0  0 104  0  0  0  0  0  0   129
#> 478  412  0 16  3  0  2  0  0   1  0  2  0  0  0  70  0  0  0  0  0  0    94
#> 341  413  0 18  3  0  1  0  1   0  0  2  0  0  0  36  0  0  0  0  0  0    61
#> 479  414  0 20  6  0  0  0  2   0  0  1  0  0  0  19  0  0  0  0  0  0    48
#> 342  415  0 23  8  0  0  0  4   1  0  0  0  0  0   2  0  0  0  0  0  0    38
#> 343  416  0 20  3  0  0  1  3   0  0  0  0  0  0   0  0  0  0  0  0  0    27
#> 344  417  0 19  4  0  0  1  2   0  0  0  0  0  0  12  0  0  0  0  0  0    38
#> 345  418  0 21  6  0  0  0  3   1  0  1  0  0  0  56  0  0  0  0  0  0    88
#> 346  419  0 23  2  0  0  0  4   1  0  1  0  0  0  73  0  0  0  0  0  0   104
#> 480  420  0 17  2  0  0  0  2   1  0  0  0  0  0  96  0  0  0  0  0  0   118
#> 347  421  0 11  2  0  0  0  1   1  0  0  0  0  0 118  0  0  0  0  0  0   133
#> 348  422  0 18  2  0  0  0  0   1  0  0  0  0  0 147  0  0  0  0  0  0   168
#> 349  423  0 14  0  0  0  0  0   0  0  0  0  0  0 135  0  0  0  0  0  0   149
#> 350  424  0 15  0  0  0  0  3   1  1  0  0  0  0 165  0  0  0  0  0  0   185
#> 351  425  0 17  0  0  0  1  5   2  4  2  0  0  0 161  0  0  0  0  0  0   192
#> 481  426  0 23  1  0  0  3  4   3  7  3  0  0  0 110  0  0  0  0  0  0   154
#> 482  427  0 30  1  0  0  5  2   3 10  5  0  0  0  59  0  1  1  0  0  0   117
#> 352  428  0 36  2  0  0  7  1   4 13  6  0  0  0   8  0  1  1  0  0  0    79
#> 353  429  0 33  4  0  0  7  1   4  8 11  0  0  0  26  0  3  1  0  0  0    98
#> 354  430  0 35  4  0  0  4  2   5  7  9  0  0  0  61  0  0  0  0  0  0   127
#> 355  431  0 36  2  0  1  8  2   5  7  4  0  0  0 116  0  1  0  0  0  0   182
#> 483  432  0 44  4  0  1  4  4   8  8  4  0  0  0 160  0  0  0  0  0  0   237
#> 356  433  0 52  5  0  1  0  6  10  8  5  0  0  0 203  0  0  0  0  0  0   290
#> 357  434  0 51  4  0  0  0  4  29  5  1  0  0  0 285  0  0  0  0  0  0   379
#> 358  435  0 50  3  0  0  0  8  26  4  3  0  0  0 238  0  1  0  0  0  0   333
#> 359  436  0 45  2  0  0  4  9  25 10  1  0  0  0 166  0  0  0  0  0  0   262
#> 360  437  0 49  3  0  0  6 12  18  7  2  0  0  0 122  0  0  0  0  0  0   219
#> 361  438  0 50  2  0  1  7 10  12 10  1  0  0  0  39  0  0  0  0  0  0   132
#> 362  439  0 71  7  0  1  7  3  13 10  0  0  0  0   8  0  0  0  0  0  0   120
#> 363  440  0 51  3  0  0  5  5   9  7  1  0  0  0   0  0  1  0  0  0  0    82
#> 364  441  0 55  4  0  1  7  2  10  7  1  0  0  0   4  0  0  0  0  0  0    91
#> 365  442  0 63 12  0  1  7  3  11  9  9  0  0  0  17  0  2  1  0  0  0   135
#> 366  443  2 58  4  0  1  3  4   7  6  3  0  0  0  95  0  0  0  0  0  0   183
#> 367  444  0 52  5  0  0  7  2  15  5  4  0  0  0 100  0  0  0  0  0  0   190
#> 484  445  0 46  5  0  0  4  1  14  4  5  0  0  0 126  0  0  0  0  0  0   205
#> 368  446  0 40  5  0  0  0  0  14  4  6  0  0  0 152  0  0  0  0  0  0   221
#> 485  447  0 40  6  0  0  0  1  14  5  3  0  0  0 140  0  0  0  0  0  0   209
#> 369  448  0 39  6  0  0  0  2  13  6  0  0  0  0 128  0  0  0  0  0  0   194
#> 486  449  0 42  5  0  0  2  2  14 10  0  0  0  0 103  0  0  0  0  0  0   178
#> 370  450  0 46  4  0  1  4  2  16 13  0  0  0  0  78  0  0  0  0  0  0   164
#> 487  451  0 50  4  0  0  2  4  16 12  0  0  0  0  42  0  0  0  0  0  0   130
#> 371  452  0 54  3  0  0  0  5  16 11  0  0  0  0   5  0  1  0  0  0  0    95
#> 372  453  0 67 10  0  1  5  1  10 30  1  0  0  0   5  0  4  0  0  0  0   134
#> 373  454  0 61  7  0  1  1  2  12 31  1  0  0  0  36  0  6  0  1  0  0   159
#> 374  455  3 57 14  0  0  2  0  10 23  3  0  0  0  97  0  9  0  0  0  0   218
#> 375  456  0 57 19  0  2  3  4  17 19  0  0  0  0 103  0  3  0  0  0  0   227
#> 376  457  1 54 12  0  1  2  0  15 14  1  0  0  0 141  0  5  1  0  0  0   247
#> 377  458  1 48 13  0  0  3  2  26  5  0  0  0  0 202  0  2  0  1  0  0   303
#> 378  459  0 45 15  0  1  1  5  14 11  0  0  0  0 111  0  3  1  0  0  0   207
#> 379  460  1 38 10  0  0  1  6  28  8  0  0  0  0 210  0  0  0  0  0  0   302
#> 380  461  1 36  6  0  1  3 10  28 13  1  0  0  0 135  0  0  0  0  0  0   234
#> 381  462  0 39  9  0  1  0 14  24 14  1  1  0  0  78  0  0  0  3  0  0   184
#> 382  463  1 35  9  0  1  4 15  25 16  0  0  0  0  42  0  7  3  2  2  0   162
#> 488  464  2 45 11  0  1  6 14  19 16  2  0  0  0  30  0 12  8  2  2  0   170
#> 383  465  2 55 13  0  1  7 12  13 17  3  0  0  0  17  0 17 12  2  2  0   173
#> 384  466 11 61 23  0  1 11 16  18 23  6  0  0  0  41  0 32 14 14  4  0   275
#> 385  467 27 49 28  0  1  4 13  15 24  4  0  0  0  94  0 64  4 10  8  0   345
#> 386  468 37 30 23  0  1  7 11  11 31  7  0  0  0 134  0 81  3 12 10  0   398
#> 387  469 18 48 21  0  0 14 18  11 26  4  0  0  2 102  0 32  2  9  7  0   314
#> 388  470 31 44 13  0  1  6 11  18 32 11  0  0  6 164  0 20  1 21 17  0   396
#> 389  471 19 45 12  0  1  7 10  12  9  9  0  0  5 229  0  9  0 18  5  0   390
#> 390  472 15 41 18  0  4 10 19  14 14  8  0  0  4 247  0  3  0  1  1  0   399
#> 391  473 12 47 14  0  2 11 19  14 18  1  0  2  3 144  0  1  0  1  1  0   290
#> 392  474 15 54 16  0  2 17 27  16 23  9  0  0  9  87  0 12  0  4  1  0   292
#> 393  475  5 56 17  0  4 17 32  15 25  6  0  1 11  22  0 17  0  1  4  0   233
#> 489  476  4 56 19  0  3 13 22  11 21  4  0  1 10  15  0 22  1  1  3  0   206
#> 490  477  4 55 21  0  3 10 11   7 18  2  0  0  9   9  0 28  2  0  2  0   181
#> 394  478  3 55 23  0  2  6  1   3 14  0  0  0  8   2  0 33  3  0  1  0   154
#> 395  479  8 50 22  0  2  2  3   2  9  1  0  0  4  29  1 23  1  2  0  0   159
#> 491  480  6 54 21  0  3  2  4   3  6  0  0  0  4  53  0 14  1  1  0  0   172
#> 396  481  5 59 20  1  4  3  5   4  4  0  0  0  5  77  0  4  1  0  0  0   192
#> 397  482  1 44 15  1  3  1  2   6  4  5  0  0  1  61  0  2  0  1  0  0   147
#> 398  483  0 35 12  1  1  2  2   4  0  5  0  0  0  84  0  0  0  1  0  0   147
#> 399  484  0 30 12  1  1  1  4   6  3  2  0  0  0  93  0  0  0  0  0  0   153
#> 400  485  0 34  9  0  3  2  4   6  5  1  0  0  0  76  0  0  0  0  0  0   140
#> 401  486  0 34  5  1  3  1  7   5  5  0  0  0  0  38  0  0  0  0  0  0    99
#> 402  487  0 42  6  1  4  2 10   5  5  0  0  1  0   9  0  0  0  0  1  0    86
#> 492  488  0 48  9  1  4  2 10   4  5  0  0  2  1   6  0  4  1  0  2  0    99
#> 493  489  0 55 13  1  5  1  9   2  4  0  0  3  2   3  0  7  1  0  2  0   108
#> 403  490  0 61 16  1  5  1  9   1  4  0  0  4  3   0  0 11  2  0  3  0   121
#> 404  491  0 62 20  1  7  0  6   2  6  0  0  5  2   2  1 10  2  0  1  0   127
#> 405  492  0 55 17  0  0  3 11   1 11  0  0  3  4  44  0 10  4  0  0  0   163
#> 406  493  0 63 19  0  6  2  8   2 11  1  0  2  1  44  1  1  0  0  8  0   169
#> 407  494  0 44 24  0  5  1 11   5 10  1  1  1  0  92  0  1  1  0  0  0   197
#> 408  495  0 33  7  1  7  0  8   2  2  1  2  0  0 108  0  0  0  0  2  0   173
#> 494  496  0 33  8  0  4  0 12   5  2  0  2  0  0 133  0  0  0  0  1  0   200
#> 409  497  0 33  9  0  1  0 15   8  2  0  1  0  0 158  1  0  0  0  0  0   228
#> 410  498  0 42  7  0  6  0 15   6  1  0  1  0  0 213  0  0  0  0  0  0   291
#> 411  499  0 41  5  1  6  0 26   5  9  0  2  1  1  94  0  1  0  0  1  0   193
#> 412  500  0 41  3  1  5  0 36   6 14  0  0  1  1  33  0  1  0  0  0  0   142
#> 495  501  0 42  2  1  4  0 30   6 12  0  0  1  1  30  0  1  0  0  0  0   130
#> 496  502  0 44  2  1  2  0 25   6  9  0  0  0  1  28  0  0  0  0  0  0   118
#> 497  503  0 46  1  1  1  0 20   6  6  0  0  0  1  26  0  0  0  0  0  0   108
#> 413  504  0 47  0  1  0  0 14   6  4  0  0  0  1  23  0  0  0  0  0  0    96
#> 414  505  0 48  2  1  1  1 11   3  1  0  0  0  0  51  0  0  0  0  0  0   119
#> 415  506  0 21  3  1  2  0 11   2  4  0  0  0  0  89  0  0  0  0  0  0   133
#> 416  507  0 33  1  1  1  0  8   6  2  0  0  0  0  77  0  0  0  0  1  0   130
#> 417  508  0 25  0  1  0  0  7   6  5  0  0  0  0  83  0  0  0  0  0  0   127
#> 418  509  0 23  1  0  2  0  7   3  1  0  0  0  0  53  0  0  0  0  0  0    90
#> 419  510  0 22  2  1  1  0  9   5  3  0  0  0  0  48  0  0  0  0  0  0    91
#> 420  511  0 22  2  0  1  0  9   5  1  0  0  0  0  29  0  0  0  0  0  0    69
#> 421  512  0 27  0  0  1  1  1   4  4  0  0  0  0   2  0  0  0  0  0  0    40
#> 498  513  0 30  0  0  1  1  1   4  3  0  0  0  0   4  0  1  0  0  0  0    45
#> 499  514  0 34  0  0  0  1  2   4  2  0  0  0  0   6  0  2  0  0  0  0    51
#> 500  515  0 37  1  0  0  1  2   3  2  0  0  0  0   7  0  2  0  0  0  0    55
#> 422  516  0 40  1  0  0  1  2   3  1  0  0  0  0   9  0  3  0  0  0  0    60
#> 423  517  0 32  0  0  0  0  3   4  6  0  0  0  0  26  0  4  0  0  0  0    75
#> 424  518  0 22  1  0  1  0  2   3  3  0  0  0  0  26  0  0  0  0  0  0    58
#> 425  519  0 18  5  0  1  0  1   5  1  0  0  0  0  39  0  0  0  0  0  0    70
#> 426  520  0 14  1  0  0  0  5   3  1  0  0  0  0  48  0  0  0  0  1  0    73
#> 427  521  0  8  4  0  0  1  4   3  1  0  0  0  0  69  0  0  0  0  0  0    90
#> 428  522  0 14  1  0  1  0  4   3  0  1  0  0  0  69  0  0  0  0  0  0    93
#> 429  523  0 19  4  0  0  0 10   8  2  0  0  0  0  78  0  0  0  0  0  0   121
#> 430  524  0 19  4  0  0  0  6   7  6  1  0  0  0  31  0  0  0  0  0  0    74
#> 431  525  0 25  9  0  0  0  9   8  5  0  0  0  0   7  0  2  0  0  0  0    65
#> 432  526  0 28  6  0  0  0  5   3  2  0  0  0  0   1  0  3  0  0  0  0    48
#> 433  527  0 37  5  0  0  0  8   7  5  0  0  0  0   0  0  2  0  0  0  0    64

 # }