Prepare a rodents data tables for forecasting

This function wraps around summarize_rodent_data to produce a data.frames associated with a set of data specifications.

prep_rodents_table(
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = TRUE,
  clean = FALSE,
  type = "Rodents",
  level = "Site",
  plots = "all",
  treatment = NULL,
  min_plots = 24,
  min_traps = 1,
  output = "abundance",
  shape = "crosstab",
  fillweight = (output != "abundance"),
  unknowns = FALSE,
  time = "newmoon",
  na_drop = FALSE,
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site = TRUE),
  ref_species = all_species(),
  effort = FALSE,
  quiet = TRUE,
  verbose = FALSE,
  control_files = files_control(),
  save = TRUE,
  overwrite = TRUE,
  filename = "rodents_all.csv",
  arg_checks = TRUE
)

process_rodent_data(
  rodents_tab,
  main = ".",
  moons = NULL,
  end_moon = NULL,
  species = base_species(),
  total = TRUE,
  interpolate = FALSE,
  clean = FALSE,
  level = "Site",
  treatment = NULL,
  na_drop = FALSE,
  time = "newmoon",
  ref_species = all_species(),
  quiet = TRUE,
  verbose = FALSE,
  arg_checks = TRUE
)

Arguments

main
character value of the name of the main component of the directory tree.

moons
Moons data.frame. See prep_moons.

end_moon
integer (or integer numeric) newmoon number of the last sample to be included. Default value is NULL, which equates to the most recently included sample.

species
character-valued vector of species names to include.

total
logical value indicating if a total (sum across species) should be added or not. Only available if more than one species is included.

interpolate
logical value indicating if the data should be interpolated to fill in NA values (using na.interp).

clean
logical indicator of if only the rodent data that passed QA/QC (clean = TRUE) or if all data (clean = FALSE) should be loaded.

type
character value of the rodent data set type, according to pre-existing definitions. An alternative toggle to species.
Either all species (type = "Rodents") or only granivoes (type = "Granivores").

level
character indicating the type of summary: "Plot", "Treatment", or "Site". Pipes directly to summarize_rodent_data.

plots
Specification of subset of plots. Can be a vector of numeric plots indicators or specific sets indicated by character values: "all" plots or "Longterm" plots (plots that have had the same treatment for the entire time series).

treatment
character indicating the specific treatment(s) to trim to if level = "Treatment": "control", "exclosure", "removal", or "spectabs"

min_plots
integer (or integer numeric) of the minimum number of plots surveyed for a survey to be used. Pipes directly to summarize_rodent_data.

min_traps
integer (or integer numeric) of the minimum number of traps collected for a plot to be used. Pipes directly to summarize_rodent_data.

output
character indicating the type of data: "abundance", "biomass", or "energy". Pipes directly to summarize_rodent_data.

shape
character value indicating a "crosstab" or "flat" output.

fillweight
logical specifier of whether to fill in unknown weights with other records from that individual or species, where possible.

unknowns
logical indicator to either remove all individuals not identified to species (unknowns = FALSE) or sum them in an additional column (unknowns = TRUE.

time
character value specifying the format of the time index in the output. Options are "period" (sequential Portal surveys), "newmoon" (lunar cycle numbering), and "date" (calendar date).
The default time = "newmoon" produces an equispaced observation timestep, a common format format for discrete-time modeling.

na_drop
logical indicator of if NA values (representing insufficient sampling) should be dropped.

zero_drop
logical indicator of if 0 values (representing sufficient sampling but no detection) should be dropped.

ref_species
character vector of the reference species. Sets a boundary outside of the species included in the model.

effort
logical indicator of if the effort columns should be included in the output.

quiet
logical indicator if progress messages should be quieted.

verbose
logical indicator if detailed messages should be shown.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control. Only used here to point to the moons data.

save
logical indicator controlling if the output should be saved out.

overwrite
logical indicator of whether or not the existing files should be updated (most users should leave as TRUE).

filename
character name of the file for saving the output. The extension of the filename may be used for identifying object requirements (for example, if the extension of filename is ".csv", it will trigger use of write.csv).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

rodents_tab
Rodents data.frame data table.

Value

data.frame for the specified data set.

Examples

 # \donttest{
   create_dir()
   fill_raw()
   prep_rodents_table()
 # }