Read in cast output from a given cast

Read in the various output files of a cast or casts in the casts sub directory.

read_cast_tab(main = ".", cast_id = NULL, arg_checks = TRUE)

read_cast_tabs(main = ".", cast_ids = NULL, arg_checks = TRUE)

read_cast_metadata(main = ".", cast_id = NULL, arg_checks = TRUE)

read_model_fits(main = ".", cast_id = NULL, arg_checks = TRUE)

read_model_casts(main = ".", cast_id = NULL, arg_checks = TRUE)

Arguments

main
character value of the name of the main component of the directory tree.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

cast_ids, cast_id
integer (or integer numeric) value(s) representing the cast(s) of interest, as indexed within the directory in the casts sub folder. See the casts metadata file (casts_metadata.csv) for summary information. If NULL (the default), the most recently generated cast's output is read in.
cast_ids can be NULL, one value, or more than one values, cast_id can only be NULL or one value.

Value

read_cast_tab: data.frame of the cast_tab.

read_cast_tabs: data.frame of the cast_tabs with a cast_id column added to distinguish among casts.

read_cast_metadata: list of cast_metadata.

read_model_fits: list of model_fits.

read_model_casts: list of model_casts.

Examples

 # \donttest{
  setup_dir()
  portalcast(models = c("AutoArima", "NaiveArima"), end_moons = 515:520)
  read_cast_tab(cast_id = 1)
  read_cast_tabs(cast_ids = 1:2)
  read_cast_metadata(cast_id = 1)
  read_model_fits(cast_id = 1)
  read_model_casts(cast_id = 1)
# }