Create a control lists for generating a rodents data table

Given the number of arguments into summarize_rodent_data via prep_rodents_table, it helps to have a control list to organize them. rodents_control provides a template for the controls list used to define the specific arguments for a user's novel data set, setting the formal arguments to the basic default values

rodents_control(name = "name", species = base_species(),
  total = TRUE, interpolate = FALSE, clean = FALSE,
  type = "Rodents", level = "Site", plots = "all",
  treatment = NULL, min_plots = 24, min_traps = 1,
  output = "abundance", fillweight = (output != "abundance"),
  unknowns = FALSE, time = "newmoon", na_drop = FALSE,
  zero_drop = switch(tolower(level), plot = FALSE, treatment = TRUE, site
  = TRUE), effort = FALSE, filename = paste0("rodents_", name, ".csv"),
  arg_checks = TRUE)

Arguments

name
character value of the rodent data set name used to enforce certain arguments. Currently used prefab data sets are "all", "all_interp", "controls", and "controls_interp".

species
character-valued vector of species names to include.

total
logical value indicating if a total (sum across species) should be added or not. Only available if more than one species is included.

interpolate
logical value indicating if the data should be interpolated to fill in NA values (using na.interp).

clean
logical indicator of if only the rodent data that passed QA/QC (clean = TRUE) or if all data (clean = FALSE) should be loaded.

type
character value of the rodent data set type, according to pre-existing definitions. An alternative toggle to species.
Either all species (type = "Rodents") or only granivores (type = "Granivores").

level
character indicating the type of summary: "Plot", "Treatment", or "Site". Pipes directly to summarize_rodent_data.

plots
Specification of subset of plots. Can be a vector of numeric plots indicators or specific sets indicated by character values: "all" plots or "Longterm" plots (plots that have had the same treatment for the entire time series).

treatment
character indicating the specific treatment(s) to trim to if level = "Treatment": "control", "exclosure", "removal", or "spectabs"

min_plots
integer (or integer numeric) of the minimum number of plots surveyed for a survey to be used. Pipes directly to summarize_rodent_data.

min_traps
integer (or integer numeric) of the minimum number of traps collected for a plot to be used. Pipes directly to summarize_rodent_data.

output
character indicating the type of data: "abundance", "biomass", or "energy". Pipes directly to summarize_rodent_data.

fillweight
logical specifier of whether to fill in unknown weights with other records from that individual or species, where possible.

unknowns
logical indicator to either remove all individuals not identified to species (unknowns = FALSE) or sum them in an additional column (unknowns = TRUE.

time
character value specifying the format of the time index in the output. Options are "period" (sequential Portal surveys), "newmoon" (lunar cycle numbering), and "date" (calendar date).
The default time = "newmoon" produces an equispaced observation timestep, a common format format for discrete-time modeling.

na_drop
logical indicator of if NA values (representing insufficient sampling) should be dropped.

zero_drop
logical indicator of if 0 values (representing sufficient sampling but no detection) should be dropped.

effort
logical indicator of if the effort columns should be included in the output.

filename
character name of the file for saving the output. The extension of the filename may be used for identifying object requirements (for example, if the extension of filename is ".csv", it will trigger use of write.csv).

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

Named list of data_set's data-generating controls, for input as part of controls_rodents in prep_rodents.

Examples

 rodents_controls(c("all", "controls"))
 all_PPonly <- rodents_control("all_PPonly", species = "PP")
 rodents_controls("all_PPonly", all_PPonly, arg_checks = FALSE)
 rodents_controls(c("all", "all_PPonly"), all_PPonly, arg_checks = FALSE)