Create a control list for a runjags JAGS model run

We leverage the run.jags function in the runjags (Denwood 2016) package to run JAGS (Plummer 2003) models in portalcasting. That function has a number of control parameters that users may be interested in changing, and this function wraps those parameters with a few portalcasting-specific parameters into a control list for input into specific model functions.

runjags_control(
  nchains = 4,
  adapt = 10000,
  burnin = 10000,
  sample = 10000,
  thin = 10,
  modules = "glm",
  method = "interruptible",
  factories = "",
  mutate = NA,
  cast_obs = TRUE,
  silent_jags = FALSE
)

Arguments

nchains

Non-negative integer-conformable value of the number of parallel chains to use. See run.jags.

adapt

Non-negative integer-conformable value of the number of adaptation steps to use. See run.jags.

burnin

Non-negative integer-conformable value of the number of burnin steps to use. See run.jags.

sample

Non-negative integer-conformable value of the number of sampling steps to use. See run.jags.

thin

Non-negative integer-conformable value of the thinning interval to use. See run.jags.

modules

character vector of external modules to add to JAGS. See run.jags.

method

character value of the run.jags method to use. Options include "rjags", "simple", "interruptible", "parallel", "rjparallel", "background", "bgparallel", and "snow". See run.jags.

factories

character vector of factory modules to add to JAGS. See run.jags.

mutate

A function or list (with the first element being a function) used to add variables to the posterior chain (rather than throughout sampling). See run.jags.

cast_obs

logical value indicating if the observations should be forecast as well (should be kept as TRUE for the vast majority of uses).

silent_jags

logical value for quieting the output from the runjags function, including the underlying JAGS output.

Value

list of controls.

References

Denwood, M. J. 2016. runjags: an R package providing interface utilities, model templates, parallel computing methods and additional distributions for MCMC models in JAGS. Journal of Statistical Software, 71:9. URL.

Plummer, M. 2003. JAGS: A program for analysis of Bayesian graphical models using Gibbs Sampling. Proceedings of the 3rd International Workshop on Distributed Statistical Computing (DSC 2003). ISSN 1609-395X. URL.

Examples

 runjags_control()