Determine if columns in a table are species columns

Given a table, returns a logical vector indicating if each column is a species' column or not.

is_sp_col(rodents_tab, species = NULL, total = FALSE, nadot = FALSE,
  arg_checks = TRUE)

species_from_table(rodents_tab = NULL, total = FALSE, nadot = FALSE,
  arg_checks = TRUE)

Arguments

rodents_tab
data.frame of columns to be checked.

species
character vector indicating species names to use in determining if columns are species columns. Defaults to all_species.

total
logical indicator if the total should be included.

nadot
logical indicator if the dot should be added to the "NA" species name. Defaults to FALSE.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

is_sp_col: logical vector indicating if each column is a species' column or not.

species_from_table: character vector of species names from the column names in the rodents_tab data table.

Examples

 # \donttest{
  create_dir()
  fill_raw()
  rodents_tab <- prep_rodents_table()
  is_sp_col(rodents_tab)
  species_from_table(rodents_tab)
 # }