Trim a moons table based on target moons

Some functions require that a data table of time (the moons) table only includes specific time stamps (newmoon numbers). This function is a simple utility to reduce the moons table to that as well as the columns.

trim_moons(
  moons = NULL,
  target_moons = NULL,
  retain_target_moons = TRUE,
  target_cols = c("moon", "moondate", "period", "censusdate"),
  arg_checks = TRUE
)

Arguments

moons
Moons data.frame. See prep_moons.

target_moons
integer-conformable newmoon numbers, usually made through target_moons of the moons to either drop or retain based on retain_target_moons.

retain_target_moons
logical value that dictates if target_moons are to be dropped (retain_target_moons = FALSE or retained (retain_target_moons = TRUE)

target_cols
character vector of columns to retain.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.
However, in sandboxing, it is often desirable to be able to deviate from strict argument expectations. Setting arg_checks = FALSE triggers many/most/all enclosed functions to not check any arguments using check_args, and as such, caveat emptor.

Value

moons and a data.frame but trimmed.

Examples

 # \donttest{
  setup_dir()

#> ------------------------------------------------------------

#> This is portalcasting v0.20.0

#> ------------------------------------------------------------

#> Establishing portalcasting directory at
#>  /tmp/RtmppX5q9A/file4ca251e6eebb/reference

#> ------------------------------------------------------------

#> Filling directory with standard content

#>  -Downloading raw files

#>   -PortalData

#>   -portalPredictions

#>  -Filling casts folder with files from archive

#>  -Filling fits folder with files from archive

#>  -Writing model scripts

#>   -AutoArima

#>   -ESSS

#>   -NaiveArima

#>   -nbGARCH

#>   -nbsGARCH

#>   -pevGARCH

#>   -simplexEDM

#>   -GPEDM

#>   -jags_RW

#>  -Adding data files to data subdirectory

#>   -lunar data file

#>   -rodents data files

#>    -all

#>    -controls

#>    -exclosures

#>    -dm_controls

#>    -all_interp

#>    -controls_interp

#>    -exclosures_interp

#>    -dm_controls_interp

#>   -covariate data files

#>   -metadata file

#> ------------------------------------------------------------

#> Directory successfully instantiated

#> ------------------------------------------------------------

  moons <- prep_moons()
  trim_moons(moons, 300:310)

#>     moon   moondate period censusdate
#> 300  300 2001-09-17    280 2001-09-22
#> 301  301 2001-10-16    281 2001-10-13
#> 302  302 2001-11-14    282 2001-11-17
#> 303  303 2001-12-14    283 2001-12-15
#> 304  304 2002-01-13    284 2002-01-12
#> 305  305 2002-02-12    285 2002-02-09
#> 306  306 2002-03-13    286 2002-03-13
#> 307  307 2002-04-12    287 2002-04-16
#> 308  308 2002-05-12    288 2002-05-15
#> 309  309 2002-06-10    289 2002-06-15
#> 310  310 2002-07-10    290 2002-07-13

 # }