Write the template for a model into model subdirectory

write_model creates a template script (as written by model_template) for a given model.

model_template creates the character-valued text for a model script to be housed in the model directory, as written out by write_model.

control_list_arg creates the character-valued text for a specific list argument into model function within a model script to be housed in the model directory.

write_model(
  name = NULL,
  data_sets = NULL,
  covariatesTF = NULL,
  lag = NULL,
  main = ".",
  control_model = NULL,
  control_files = files_control(),
  control_runjags = NULL,
  max_E = NULL,
  quiet = FALSE,
  verbose = TRUE,
  arg_checks = TRUE
)

model_template(
  name = NULL,
  data_sets = NULL,
  covariatesTF = FALSE,
  lag = NULL,
  main = ".",
  control_files = files_control(),
  max_E = NULL,
  control_runjags = NULL,
  quiet = FALSE,
  verbose = FALSE,
  arg_checks = TRUE
)

control_list_arg(control_list = NULL, list_function = NULL, arg_checks = TRUE)

Arguments

name
character value of the name of the model.

data_sets
character vector of the rodent data set names that the model is applied to.

covariatesTF
logical indicator for if the model requires covariates.

lag
integer (or integer numeric) lag time used for the covariates or NULL if covariatesTF is FALSE.

main
character value of the name of the main component of the directory tree.

control_model
list of model-level controls, including name, a character value of the model's name; covariatesTF, a logical indicator for if the model requires covariates; and lag, a integer (or integer numeric) lag time used for the covariates or NA if covariatesTF = FALSE. Only used if the specific valued argument is NULL.

control_files
list of names of the folders and files within the sub directories and saving strategies (save, overwrite, append, etc.). Generally shouldn't need to be edited. See files_control.

control_runjags
list of arguments passed to run.jags via runjags_control.

max_E
integer (or integer numeric) for the maximum embedding dimension to search amongst for EDM models. See simplex for more information.

quiet
logical indicator controlling if messages are printed.

verbose
logical indicator if detailed messages should be shown.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

control_list
list of arguments passed to list_function.

list_function
character value name of the function to send control_list arguments to within the model script.

Value

write_mode writes the model script out and returns NULL.

model_template: character-valued text for a model script to be housed in the model directory.

control_list_arg: character-valued text for part of a model script.

Examples

 # \donttest{
  create_dir()
  write_model("AutoArima")
  model_template()
  control_list_arg(runjags_control(nchains = 3), "runjags_control")
 # }