Obtain the URL for a Zenodo record to be downloaded

zenodo_url obtains the URL for a given Zenodo record, identified either by the concept record identifier (concept_rec_id) and version (rec_version) or record identifier (rec_id). (Note: if rec_id is used, it overrides concept_rec_id).

zenodo_versions: determines the available version numbers and the corresponding record identifier for each version available for a given Zenodo concept (group of records).

zenodo_url(
  concept_rec_id = NULL,
  rec_version = "latest",
  rec_id = NULL,
  arg_checks = TRUE
)

zenodo_versions(concept_rec_id, arg_checks = TRUE)

Arguments

concept_rec_id
Concept record identifier, a character value corresponding to the Zenodo concept.

rec_version
character value of the version number or "latest" (default) for the data to be download.

rec_id
Optional input record identifier, a character value corresponding to the Zenodo record.

arg_checks
logical value of if the arguments should be checked using standard protocols via check_args. The default (arg_checks = TRUE) ensures that all inputs are formatted correctly and provides directed error messages if not.

Value

zenodo_url: character value of the URL for the zip to be downloaded.

zenodo_versions: a data.frame of version number and record identifier for each version available.

Examples

 # \donttest{
   zenodo_versions("1215988")

#>     version  rec_id
#> 1     1.2.0 1216139
#> 2    1.18.0 1229505
#> 3    1.34.0 1254536
#> 4    1.52.0 1318647
#> 5    1.71.0 2530876
#> 6    1.92.0 2574985
#> 7   1.105.0 2726986
#> 8   1.119.0 3338641
#> 9   1.133.0 3470238
#> 10    2.4.0 3600576
#> 11   2.20.0 3752088
#> 12   2.33.0 3934844
#> 13   2.53.0 4266863
#> 14   2.65.0 4448173
#> 15    1.7.0 1217173
#> 16   1.22.0 1231360
#> 17   1.40.0 1284495
#> 18   1.59.0 1466004
#> 19   1.82.0 2538173
#> 20   1.95.0 2591424
#> 21  1.113.0 3244287
#> 22  1.127.0 3379060
#> 23  1.142.0 3561133
#> 24   2.14.0 3695938
#> 25   2.30.0 3897522
#> 26   2.45.0 4044629
#> 27   2.61.0 4382814
#> 28   2.76.0 4477064
#> 29   1.11.0 1217742
#> 30   1.26.0 1238027
#> 31   1.44.0 1295667
#> 32   1.62.0 1469121
#> 33   1.83.0 2538177
#> 34   1.98.0 2610349
#> 35  1.114.0 3244359
#> 36  1.135.0 3471763
#> 37    2.6.0 3608225
#> 38   2.22.0 3774512
#> 39   2.37.0 3972384
#> 40   2.54.0 4272069
#> 41   2.71.0 4464887
#> 42    1.4.0 1217163
#> 43   1.20.0 1229647
#> 44   1.38.0 1257700
#> 45   1.60.0 1469070
#> 46   1.80.0 2537230
#> 47   1.94.0 2587251
#> 48  1.112.0 3243525
#> 49  1.124.0 3367759
#> 50    1.1.0 1215989
#> 51   1.17.0 1219889
#> 52   1.33.0 1254517
#> 53   1.48.0 1308970
#> 54   1.73.0 2530985
#> 55   1.89.0 2540867
#> 56  1.106.0 2825465
#> 57  1.122.0 3361798
#> 58   1.16.0 1219880
#> 59   1.32.0 1253350
#> 60   1.51.0 1318387
#> 61   1.77.0 2535684
#> 62   1.91.0 2573270
#> 63  1.103.0 2649460
#> 64  1.120.0 3348140
#> 65  1.134.0 3470249
#> 66    2.9.0 3629849
#> 67   2.25.0 3834857
#> 68   2.40.0 4001033
#> 69   2.55.0 4281385
#> 70   2.72.0 4465613
#> 71    1.8.0 1217574
#> 72   1.21.0 1229770
#> 73   1.42.0 1292980
#> 74   1.58.0 1465980
#> 75   1.79.0 2535760
#> 76   1.93.0 2582183
#> 77  1.110.0 3238678
#> 78  1.126.0 3372776
#> 79  1.140.0 3539625
#> 80   2.12.0 3674077
#> 81   2.31.0 3905330
#> 82   2.47.0 4069564
#> 83   2.59.0 4319116
#> 84   2.74.0 4470013
#> 85   1.10.0 1217661
#> 86   1.27.0 1239019
#> 87   1.41.0 1287966
#> 88   1.57.0 1465031
#> 89   1.75.0 2533401
#> 90   1.90.0 2541170
#> 91  1.104.0 2672933
#> 92  1.123.0 3367535
#> 93  1.136.0 3477572
#> 94    2.7.0 3609213
#> 95   2.23.0 3788097
#> 96   2.39.0 3990624
#> 97   2.63.0 4432817
#> 98   2.77.0 4479933
#> 99    2.0.0 3569995
#> 100  2.21.0 3759840
#> 101  2.38.0 3979869
#> 102  2.52.0 4243063
#> 103  2.69.0 4459499
#> 104 1.135.1 3471845
#> 105  2.15.0 3704629
#> 106  2.28.0 3880414
#> 107  2.46.0 4058883
#> 108  2.57.0 4292551
#> 109  2.73.0 4466847
#> 110   1.9.0 1217638
#> 111  1.25.0 1237877
#> 112  1.39.0 1283455
#> 113  1.55.0 1431753
#> 114  1.72.0 2530889
#> 115  1.87.0 2538187
#> 116 1.101.0 2634554
#> 117 1.117.0 3266417
#> 118 1.131.0 3461008
#> 119   2.3.0 3596168
#> 120  2.18.0 3735364
#> 121  2.36.0 3964947
#> 122  2.51.0 4143847
#> 123  2.67.0 4456009
#> 124  1.13.0 1219194
#> 125  1.29.0 1247550
#> 126  1.45.0 1295688
#> 127  1.61.0 1469105
#> 128  1.81.0 2538161
#> 129  1.96.0 2591628
#> 130 1.107.0 3100763
#> 131 1.121.0 3355702
#> 132 1.139.0 3522386
#> 133  2.13.0 3687252
#> 134  2.29.0 3887207
#> 135  2.42.0 4011276
#> 136  2.58.0 4306800
#> 137  2.75.0 4474309
#> 138   1.6.0 1217168
#> 139  1.24.0 1231506
#> 140  1.37.0 1257697
#> 141  1.56.0 1465022
#> 142  1.78.0 2535695
#> 143  1.92.1 2575806
#> 144 1.108.0 3234125
#> 145 1.128.0 3385239
#> 146 1.141.0 3547831
#> 147  2.11.0 3662910
#> 148  2.26.0 3858107
#> 149  2.43.0 4019810
#> 150  2.60.0 4323400
#> 151  2.78.0 4486720
#> 152  1.14.0 1219625
#> 153  1.31.0 1251303
#> 154  1.46.0 1295690
#> 155  1.74.0 2531062
#> 156  1.88.0 2538862
#> 157 1.102.0 2642797
#> 158 1.118.0 3279727
#> 159 1.132.0 3468535
#> 160   2.5.0 3603394
#> 161  2.19.0 3743831
#> 162  2.34.0 3945398
#> 163  2.50.0 4127222
#> 164  2.68.0 4458420
#> 165  1.12.0 1219185
#> 166  1.28.0 1243910
#> 167  1.43.0 1295592
#> 168  1.63.0 1469805
#> 169  1.85.0 2538183
#> 170  1.99.0 2621951
#> 171 1.115.0 3249316
#> 172 1.130.0 3438218
#> 173   2.2.0 3592516
#> 174  2.16.0 3713959
#> 175  2.35.0 3955151
#> 176  2.49.0 4123569
#> 177  2.64.0 4443742
#> 178   1.5.0 1217167
#> 179  1.23.0 1231499
#> 180  1.36.0 1256939
#> 181  1.53.0 1319008
#> 182  1.66.0 2211764
#> 183  1.86.0 2538186
#> 184 1.100.0 2633310
#> 185 1.116.0 3256662
#> 186 1.137.0 3491070
#> 187   2.8.0 3620856
#> 188  2.24.0 3823198
#> 189  2.44.0 4031783
#> 190  2.62.0 4396188
#> 191  2.80.0 4495683
#> 192   1.3.0 1216208
#> 193  1.19.0 1229506
#> 194  1.35.0 1255304
#> 195  1.54.0 1320480
#> 196  1.76.0 2533403
#> 197  1.94.1 2588043
#> 198 1.111.0 3243498
#> 199 1.129.0 3404213
#> 200   2.1.0 3581427
#> 201  2.17.0 3726284
#> 202  2.32.0 3924657
#> 203  2.48.0 4110852
#> 204  2.66.0 4453410
#> 205  1.15.0 1219752
#> 206  1.30.0 1251246
#> 207  1.47.0 1305401
#> 208  1.64.0 1475410
#> 209  1.84.0 2538180
#> 210  1.97.0 2598774
#> 211 1.109.0 3234999
#> 212 1.125.0 3371666
#> 213 1.138.0 3516417
#> 214  2.10.0 3637583
#> 215  2.27.0 3873675
#> 216  2.41.0 4006509
#> 217  2.56.0 4284000
#> 218  2.70.0 4460510

   zenodo_url("1215988", "latest")

#> [1] "https://zenodo.org/api/files/4e7f4eb8-9c3f-476f-80eb-529efad98027/weecology/PortalData-2.80.0.zip"

   zenodo_url("1215988", "1.71.0")

#> [1] "https://zenodo.org/api/files/5df87c60-f720-4336-91ce-d406d26fadfc/weecology/PortalData-1.71.0.zip"

   zenodo_url(rec_id = "1217163")

#> [1] "https://zenodo.org/api/files/f36b8cf7-de42-4111-b9c1-908bed08897f/weecology/PortalData-1.4.0.zip"

 # }